### 2024-02-06 #625 @b-rodrigues
rix: Reproducible Environments with Nix
Submitting Author Name: Bruno Rodrigues
Submitting Author Github Handle: @b-rodrigues
Other Package Authors Github handles:  @philipp-baumann
Repository:  https://github.com/b-rodrigues/rix
Version submitted: 0.60
Submission type: Standard

Version accepted: TBD

---

```
Package: rix
Title: Rix: Reproducible Environments with Nix
Version: 0.6.0
    person("Bruno", "Rodrigues", , "bruno@brodrigues.co", role = c("aut", "cre"),
    person("Philipp", "Baumann", , "baumann-philipp@protonmail.com", role = "aut",
  )
Description: Provides helper functions to create reproducible development
    environments using the Nix package manager.
    codetools,
    httr,
    sys,
    utils
    knitr
Config/fusen/version: 0.5.2
```

The package makes it easy to define `default.nix` files which can then be use by the Nix package manager to build a reproducible development environment. This is especially useful to make workflows completely reproducible.

This package will be useful to anyone with very strict reproducibility requirements.

Not quite: to replace what `{rix}` provides one would need to combine Docker and {renv} or {groundhog}.

https://github.com/ropensci/software-review/issues/624

@jhollist @mpadge

We are unable to have the coverage of 75%. This is due to several issues:

1. Some tests requires Nix to run. For these tests, we have a "skip if nix is not available" test.
2. Some tests make {covr} completely fail, even if they have the `skip_on_covr()` function, but on only on GA. On our computers, `covr::package_coverage()` runs and these tests are successfully skipped.
3. We have several snapshot tests, and these simply seem to be ignored by `covr`.

We have 26 tests, which all pass on different configurations: on Ubuntu, macOS and Windows, with or without Nix available (some tests get skipped if Nix is not available) and whether R itself is installed with Nix as well, or not (in other words, through the usual means for the operating system in use).

Because of these issues, our coverage is quite low:

```
> rix Coverage: 10.87%
R/get_latest.R: 0.00%
R/nix_build.R: 0.00%
R/rix.R: 0.00%
R/tar_nix_ga.R: 0.00%
R/with_nix.R: 0.68%
R/rix_init.R: 2.84%
R/detect_os.R: 28.57%
R/fetchgit.R: 90.00%
R/find_rev.R: 92.31%
R/available_r.R: 100.00%
R/detect_versions.R: 100.00%
R/get_sri_hash_deps.R: 100.00%
```

but if our snapshot tests (for the functions `rix()`, and `tar_nix_ga()`) would not be ignored, and if `covr` would successfully run the tests for `with_nix()`, our coverage would be more than 50%, we think. Some functions are very difficult to test, for example `get_latest()`, which gets the latest commit hash of the `nixpkgs` github repository.
