+++ dev/e-interactive_use.Rmd
@@ -32,7 +32,7 @@ these project-specific environments. This is because RStudio re-defines many
 environment variables and as such, a version of RStudio installed using the
 usual installer for your operating system will not be able "to see" an R
 interpreter installed with Nix. As stated in the macOS-specific vignette
-`vignette("b1-setting-up-and-using-rix-on-macos")`, RStudio is not available on
+`vignette("b2-setting-up-and-using-rix-on-macos")`, RStudio is not available on
 macOS through `nixpkgs`. The following instructions are thus only applicable to
 Linux and Windows. If you’re on macOS, read the macOS-specific vignette if
 that’s not done already, and then come back here and skip to the "scenario 3" of
