+++ dev/z-subshells.Rmd
@@ -330,8 +330,9 @@ Let's build a subshell with the latest version of R, but an older version
 of `{stringr}`:
 
 
-```{r, eval=FALSE}
+```{r}
 library("rix")
+
 path_env_stringr <- file.path(".", "_env_stringr_1.4.1")
 
 rix_init(
