### 2024-02-11 4d96207
generated correct inst/extdata/default.nix
