+++ dev/0-dev_history.Rmd
@@ -265,7 +265,7 @@ rix(r_ver = "latest",
       package_name = "rix",
       repo_url = "https://github.com/b-rodrigues/rix/",
       branch_name = "master",
-      commit = "11f6898fde38a4793a7f53900c88d2fd930e882f"
+      commit = "de009f777266a2836da26a5f3e118c7dcf911087"
     ),
     ide = "other",
     project_path = "inst/extdata",
