### 2024-02-11 9fb8d39
added vignette to generate inst/extdata/default.nix at each push
