+++ vignettes/z-advanced-topic-running-r-or-shell-code-in-nix-from-r.Rmd
@@ -366,8 +366,9 @@ of `{stringr}`:
 
 
 
-```{r eval = FALSE}
+```{r}
 library("rix")
+
 path_env_stringr <- file.path(".", "_env_stringr_1.4.1")
 
 rix_init(
