+++ dev/z-inst_rix.Rmd
@@ -1,0 +1,37 @@
+---
+title: "Developer Vignette: Generate a minimal environment containing only rix"
+output: html_document
+editor_options:
+  chunk_output_type: console
+---
+
+The code below generates the `default.nix` file included in the package
+at `inst/extdata`. You can use the call below in case you want to generate
+an environment with `{rix}` only.
+
+The latest commit gets used automatically:
+
+```{r, eval = T}
+library(rix)
+
+api_url <- "https://api.github.com/repos/b-rodrigues/rix/commits?sha=master"
+
+response <- httr::GET(url = api_url)
+
+commit_data <- jsonlite::fromJSON(httr::content(response, "text"))
+latest_commit <- commit_data$sha[1]
+
+rix(r_ver = "latest",
+    r_pkgs = NULL,
+    system_pkgs = NULL,
+    git_pkgs = list(
+      package_name = "rix",
+      repo_url = "https://github.com/b-rodrigues/rix/",
+      branch_name = "master",
+      commit = latest_commit
+    ),
+    ide = "other",
+    project_path = "inst/extdata",
+    overwrite = TRUE,
+    shell_hook = NULL)
+```
