### 2024-02-12 5291fd7
compile vignette using local()
