+++ dev/0-dev_history.Rmd
@@ -192,6 +192,14 @@ fusen::inflate(flat_file = "dev/z-inst_rix.Rmd",
                overwrite = TRUE)
 ```
 
+```{r}
+local({
+  knitr::knit("vignettes/z-developers-vignette-generate-an-environment-with-rix-only.Rmd",
+              output = tempfile())
+})
+
+```
+
 # Vignettes for users
 
 ```{r}
@@ -264,12 +272,15 @@ fusen::inflate(flat_file = "dev/z-subshells.Rmd",
 # Generating the inst/extdata/default.nix
 
 ```{r, eval = FALSE}
-api_url <- "https://api.github.com/repos/b-rodrigues/rix/commits?sha=master"
-
-response <- httr::GET(url = api_url)
-
-commit_data <- jsonlite::fromJSON(httr::content(response, "text"))
-latest_commit <- commit_data$sha[1]
+latest_commit <- tryCatch(
+  expr = {
+    sys::as_text(
+      sys::exec_internal("git", c("rev-parse", "master"))$stdout
+    )
+  },
+  # if it fails, use this commit instead
+  error = function(e) "de009f777266a2836da26a5f3e118c7dcf911087"
+)
 
 rix(r_ver = "latest",
     r_pkgs = NULL,
