### 2024-02-12 7e85209
added ropensci badge as per https://github.com/ropensci/software-review/issues/625
