+++ README.Rmd
@@ -16,7 +16,8 @@ output: github_document
 [![R-CMD-check](https://github.com/b-rodrigues/rix/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/b-rodrigues/rix/actions/workflows/R-CMD-check.yaml)
 [![runiverse-package rix](https://b-rodrigues.r-universe.dev/badges/rix?scale=1&color=pink&style=round)](https://b-rodrigues.r-universe.dev/rix)
 [![Docs](https://img.shields.io/badge/docs-release-blue.svg)](https://b-rodrigues.github.io/rix)
-[![pkgcheck](https://github.com/b-rodrigues/rix/workflows/pkgcheck/badge.svg)](https://github.com/b-rodrigues/rix/actions?query=workflow%3Apkgcheck)
+[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/625_status.svg)](https://github.com/ropensci/software-review/issues/625)
+
 <!-- badges: end -->
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
