+++ README.Rmd
@@ -17,7 +17,6 @@ output: github_document
 [![runiverse-package rix](https://b-rodrigues.r-universe.dev/badges/rix?scale=1&color=pink&style=round)](https://b-rodrigues.r-universe.dev/rix)
 [![Docs](https://img.shields.io/badge/docs-release-blue.svg)](https://b-rodrigues.github.io/rix)
 [![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/625_status.svg)](https://github.com/ropensci/software-review/issues/625)
-
 <!-- badges: end -->
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
