### 2024-02-12 c3b3677
fixed typo in macos vignette
