+++ dev/b2-macos.Rmd
@@ -119,7 +119,7 @@ As an example, call `rix::rix()` like this.
 # see ?rix_init for more details
 rix_init(
   project_path = "./rix_darwin",
-  profile_action = "create_missing"
+  rprofile_action = "create_missing"
 )
 ```
 
