### 2024-02-14 a26c591
raise warning if ide=rstudio on macos
