### 2024-02-14 49316c9
handling of r_pkgs = NULL when ide = 'rstudio'
