+++ dev/flat_rix.Rmd
@@ -524,8 +524,8 @@ testthat::test_that("If on darwin and ide = rstudio, raise warning", {
   save_default_nix_test <- function(path_default_nix) {
 
     rix(r_ver = "4.3.1",
-        r_pkgs = NULL,
         ide = "rstudio",
+        r_pkgs = NULL,
         project_path = path_default_nix,
         overwrite = TRUE,
         shell_hook = NULL
@@ -536,7 +536,7 @@ testthat::test_that("If on darwin and ide = rstudio, raise warning", {
   }
 
   testthat::expect_warning(
-              path = save_default_nix_test(path_default_nix),
+              save_default_nix_test(path_default_nix),
               regexp = "refer to the macOS"
               )
 
