+++ tests/testthat/test-rix.R
@@ -1,6 +1,8 @@
 # WARNING - Generated by {fusen} from dev/flat_rix.Rmd: do not edit by hand
 
-testthat::test_that("Snapshot test of rix()", {
+testthat::test_that("rix(), ide is 'rstudio', Linux", {
+
+  skip_if(Sys.info()["sysname"] == "Darwin")
 
   path_default_nix <- tempdir()
 
@@ -28,15 +30,47 @@ testthat::test_that("Snapshot test of rix()", {
 
   }
 
-  testthat::announce_snapshot_file("find_rev/rstudio_default.nix")
+  testthat::announce_snapshot_file("rix/rstudio_default.nix")
 
   testthat::expect_snapshot_file(
-              path = save_default_nix_test(ide = "rstudio",
-                                           path_default_nix),
-              name = "rstudio_default.nix",
-              )
+    path = save_default_nix_test(ide = "rstudio", path_default_nix),
+    name = "rstudio_default.nix",
+  )
 
-  testthat::announce_snapshot_file("find_rev/other_default.nix")
+
+})
+
+
+testthat::test_that("rix(), ide is 'other' or 'code'", {
+
+  path_default_nix <- tempdir()
+
+  save_default_nix_test <- function(ide, path_default_nix) {
+
+    rix(r_ver = "4.3.1",
+        r_pkgs = c("dplyr", "janitor", "AER@1.2-8", "quarto"),
+        tex_pkgs = c("amsmath"),
+        git_pkgs = list(
+          list(package_name = "housing",
+               repo_url = "https://github.com/rap4all/housing/",
+               branch_name = "fusen",
+               commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"),
+          list(package_name = "fusen",
+               repo_url = "https://github.com/ThinkR-open/fusen",
+               branch_name = "main",
+               commit = "d617172447d2947efb20ad6a4463742b8a5d79dc")
+        ),
+        ide = ide,
+        project_path = path_default_nix,
+        overwrite = TRUE,
+        shell_hook = NULL)
+
+    paste0(path_default_nix, "/default.nix")
+
+  }
+
+
+  testthat::announce_snapshot_file("rix/other_default.nix")
 
   testthat::expect_snapshot_file(
               path = save_default_nix_test(ide = "other",
@@ -44,7 +78,7 @@ testthat::test_that("Snapshot test of rix()", {
               name = "other_default.nix"
             )
 
-  testthat::announce_snapshot_file("find_rev/code_default.nix")
+  testthat::announce_snapshot_file("rix/code_default.nix")
 
   testthat::expect_snapshot_file(
               path = save_default_nix_test(ide = "code",
@@ -71,34 +105,53 @@ testthat::test_that("Quarto gets added to sys packages", {
 
       paste0(path_default_nix, "/default.nix")
 
-    }
+  }
 
-    testthat::announce_snapshot_file("rix/no_quarto_default.nix")
-
-    testthat::expect_snapshot_file(
-                path = save_default_nix_test(pkgs = "dplyr",
-                                             interface = "other",
-                                             path_default_nix),
-                name = "no_quarto_default.nix",
-                )
-
-    testthat::announce_snapshot_file("rix/yes_quarto_default.nix")
-
-    testthat::expect_snapshot_file(
-                path = save_default_nix_test(pkgs = c("dplyr", "quarto"),
-                                             interface = "other",
-                                             path_default_nix),
-                name = "yes_quarto_default.nix"
-              )
-
-    testthat::announce_snapshot_file("rix/null_pkgs_rstudio.nix")
+  testthat::announce_snapshot_file("rix/no_quarto_default.nix")
 
   testthat::expect_snapshot_file(
[…]
