### 2024-02-24 a2ff1e3
Update c-using-rix-to-build-project-specific-environments.Rmd
Minor fix to shebang symbol: !#  -> #!
