### 2024-02-24 ce6e724
fixed shebang in thanks to nono gueye
