### 2024-02-25 4579e68
added fix for rstudio on linux
