### 2024-03-06 #625 @wdwatkins
Estimated hours spent reviewing:6

---

Due to my organization's IT policies, I can only run `nix` in a Docker container,so all my experimenting with the actual `nix` libraries is done there.  I can use the CLI to interact with nix,though its a bit awkward without other things installed (e.g. R).  I was able to generate the default.nix files with `rix`, then copy them into the container to build them with shell commands.

You may need to update the `nix` install command at some point, just be aware if that changes

Good selection of vignettes, help files are complete.

I also found your ["What is nix" blog post](https://www.brodrigues.co/blog/2023-07-13-nix_for_r_part1/) particularly helpful for understanding what was actually happening under the hood.  It seems likely to me that a good portion of your users will also be using `nix` for the first time, so it is worth making sure you explain or link out to explanations of how `nix` itself works at an intuitive level, in the interest of usability.

#### From goodpractice results

Is there a reason for using `=` here `R/with_nix.R:564:13`?

#### code comments

I see a couple 'list versions' functions rely on a stored data file.  Is there a way to get this information directly from the nix archive so you don't need to keep that up to date?

There are a fair number of internal functions that don't have inputs/outputs documented, I would suggest documenting those for your own/future maintainer's benefit down the road.

Also, several todos still in comments.

#### test comments

Seems like get_latest.R could be tested at least that it doesn't error and returns a hash,
outside of environments where web service requests would fail?

Re: coverage—just to make sure I understand, the 25% coverage value is ignoring the tests
that actually run `nix` commands since you would need an environment with `nix`
installed to run them, which you don't get in CRAN or other non-containerized environments?
It seems like it should be possible to make a Docker container with both R and `nix`
installed to get a more realistic test environment, but maybe you've done this already.
