### 2024-03-11 #625 @assignUser
---

{rix} is a very interesting package that will allow R developers to make use of nix powerful features without having to learn (much) of its arcane secrets :mage: . For me personally it was a great opportunity to finally get to know 'nix' (special thanks for pointing to the DS installer as I have previously destroyed a shell env via half-removed nix :grin:).

While it is afaik not required by rOpenSci, I would recommend to change the default branch from 'master' to 'main'. It's a small, easy change that is well supported by GitHub but is quite meaningful.

Overall the documentation is very good and covers basic as well as advanced features of the package and points to a plethora of external resources about nix. The README especially guide both beginner and advanced users and points to the specific useful further materials. I have made a few notes while reading/following along with the vignettes that I will list as bullet points for brevity. These are all what I would consider 'nits'/nice to haves and not show stoppers.

- All vignettes start with a code block with a `library(rix)` call, which seems uneccessary (-> `echo=False`).
- The 'numbering' of the vignettes being 'hardcoded' in the title is a bit unusual and I personally would remove it as the vignettes all link to their respective follow up vignettes (well except C 'Using rix'). If this is done to make sure they appear in a certain order in the pkgdown menu: this can be controlled via the `_pkgdown.yml`.
- Getting Started: more paragraphs would make for a smoother read of the first block of text.
- b2:  'generates R expressions' -> 'nix expressions'?
- d2: Should the link to the nix package search by default be to the stable rather than unstable channel?
- e: I tried using nvim with lsp in an nix shell and that crashed but I think that might be due to my use of mason which provides a separate R env for the languageserver clashing with nix. haven't done any further investigations for now.
- d1: It's imo superflous to require both `branch` and `commit` as a commit is unique to a repo and not a branch so the additional info of branch is not necessary. Branch could be resolved to the lates commit on that branch (not entireliy inline with literate programming but useful when working against a quickly changing internal pkg?). It would also be nice to add a short hand similar to {pak} ala `owner/repo@ref` (ref = tag, branch or commit)
- Remote dependencies: I think {rix} **should** provide functionality to install packages with remote dependencies not found in nixpkgs. While it might be relativly rare case for packages publicly hosted on GitHub I think it is much less unusual for companies to have a number of internal packages that also depend on each other. The issue with no default way to know which commit could be solved by printing a warning and maybe the required shorthand to explicitly add the pkg to `rix()`.
- The developers vignettes are both quite short and could be combined into one.

- I agree with @wdwatkins regarding the test coverage.
- There is a probably unintentional partial match in `flag_git_archive`: `In cran_pkgs$archive : partial match of 'archive' to 'archive_pkgs'`
- I like the use of {fledge} but personally don't think {fusen} is the right choice for a technically advanced package like {rix}. It is of course preference and up to you but I think {fusen} changes the established workflow of an R package dev quite a bit. I see how it can lower the bar for people to get started with creating packages and that's great. But I think in this case the people that are likely going to contribute (experienced devs) might not do it if that means they have to spend time getting used to {fusen} vs. just adding a quick patch. (I certainly looked through R/ and not dev/)
- The two main functions `rix` and `with_rix` and some of the helper functions are quite long and would benefit from some refactoring (a comment hinting which if() your closing is a good pointer that some refactoring is in order :wink:  ). This would also remove duplicate code (e.g. in `serialize_globals`)
- The large error message in `with_rix` will never be reached due to the preceding `stopifnot`, maybe combine the two?
- I was surprised that you are not using {rlang} and quosures, it has been a while since I have used them but I would think they would be a good fit here. (but maybe not, it's been a while since I have used them 'under the hood')
- GHAs are my specialty so some extended notes on extdata/run-pipeline.yml (most of this applies in general as well):
  - Please explicitly add the `contents: write` [permission](https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs). This is technically not required as the token has the correct permisson but it is best practice to minimise token permissions.
    - You should also do this on the workflows in `.github/workflows/` with the minimal permissions each (e.g. `contents: read`)
  - Always use _at least_ tags for actions for stability, in an action that has write permission I would even highly recommend to explicitly use a commit hash. That ensures that even a malicous take over of one of the action repos does not allow a third-party to run code in your repo. Github's [dependabot](https://docs.github.com/en/code-security/dependabot/working-with-dependabot/automating-dependabot-with-github-actions) can handle updating shas and tags.
  - Make sure tags/hashes are updated regularly (actions/checkout is at v4 but v2 is used in the workflow).
  - secrets should always be assigned to envvars as close to where they are used as possible (usually step `env:`) and not for the entire workflow.
  - I would recommend to use an explicit [output](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-output-parameter) instead of `continue-on-error` as that still creates a big red error message in the job summary.

- It seems like the `codetools::check` is run quite a lot, I assume it's to make debugging easier?
- `with_nix` is very verbose with ~ a terminal length of 'preamble' output of checks etc., I think a 'quite' option would be nice when running it repeatedly.
- 'hidden' bindings like {arrow}'s dplyr bindings are not forwarded to the nix shell `with_nix` (the workaround of explicitly calling some function in the expression you want to run is probably way easier than an actual fix, but I don't really know as I am mostly concerned with arrow's build system^^):
    ```r
    library(rix)
    library(arrow)
    arrow_cars <- arrow_table(cars)
    arrow_test <- function() {
      arrow_cars %>%
        dplyr::filter(speed > 10)
    }
    with_nix(arrow_test())
    ```
  > => Serializing package(s) required to run `expr`:
 dplyr
