### 2024-03-12 6483e1d
added case study 3 to subshell vignette
