+++ vignettes/z-advanced-topic-running-r-or-shell-code-in-nix-from-r.Rmd
@@ -98,7 +98,7 @@ that takes care of runtime-pure R package libraries from declaratively defined
 Nix builds. Additionally, it modifies `.libPaths()` in the running R session.
 
 
-```{r parsermd-chunk-1}
+```{r}
 library("rix")
 
 path_env_1 <- file.path(".", "_env_1_R-4-1-3")
@@ -114,7 +114,7 @@ list.files(path = path_env_1, all.files = TRUE)
 This will generate the following `.Rprofile` file.
 
 
-```{r parsermd-chunk-2, echo = FALSE}
+```{r, echo = FALSE}
 cat(readLines(file.path(path_env_1, ".Rprofile")), sep = "\n")
 ```
 
@@ -122,7 +122,7 @@ Next, we write a `default.nix` file containing Nix expressions that pin R
 version 4.2.0 from Nixpkgs.
 
 
-```{r parsermd-chunk-3}
+```{r}
 rix(
   r_ver = "4.1.3",
   overwrite = TRUE,
@@ -134,7 +134,7 @@ The following expression is written to default.nix in the subfolder
 `./_env_1_R-4-1-3/`.
 
 
-```{r parsermd-chunk-4, echo = FALSE}
+```{r, echo = FALSE}
 cat(readLines(file.path(path_env_1, "default.nix")), sep = "\n")
 ```
 
@@ -146,7 +146,7 @@ higher 4.2.0 available on your system, you can check what that
 `as.vector.data.frame()` S3 method returns a list.
 
 
-```{r parsermd-chunk-5}
+```{r}
 df <- data.frame(a = 1:3, b = 4:6)
 as.vector(x = df, mode ="list")
 ```
@@ -162,7 +162,7 @@ To formally validate in a 'System-to-Nix' approach that the object returned from
 the computation above.
 
 
-```{r parsermd-chunk-6}
+```{r}
 df_as_vector <- function(x) {
   out <- as.vector(x = x, mode = "list")
   return(out)
@@ -200,7 +200,7 @@ current R console the specific Nix paths that will be downloaded and copied into
 your Nix store automatically.
 
 
-```{r parsermd-chunk-7, eval = FALSE}
+```{r, eval = FALSE}
 # now run it in `nix-shell`; `with_nix()` takes care
 # of exporting global objects of `df_as_vector` recursively
 out_nix_1 <- with_nix(
@@ -226,7 +226,7 @@ In the previous code snippet we wrapped the top-level `expr` function with
 arguments when assigning the function used as `expr` input like this:
 
 
-```{r parsermd-chunk-8}
+```{r}
 df_as_vector <- function(x = df) {
   out <- as.vector(x = x, mode = "list")
   return(out)
@@ -237,7 +237,7 @@ Then, you just supply the name of the function to evaluate with default
 arguments.
 
 
-```{r parsermd-chunk-9, eval = FALSE}
+```{r, eval = FALSE}
 out_nix_1_b <- with_nix(
   expr = df_as_vector, # provide name of function
   program = "R", 
@@ -250,7 +250,7 @@ out_nix_1_b <- with_nix(
 It yields the same results.
 
 
-```{r parsermd-chunk-10, eval = FALSE}
+```{r, eval = FALSE}
 Reduce(f = identical, list(out_nix_1, out_nix_1_b))
 ```
 
@@ -266,7 +266,7 @@ from. Accordingly, we define the R 4.2.0 environment in a
 the target environment where `df_as_vector()` will be evaluated in.
 
 
-```{r parsermd-chunk-11}
+```{r}
 library("rix")
[…]
