### 2024-03-12 #625 @ldecicco-USGS
Thanks so much @wdwatkins and @assignUser ! There's no specific due date for a response. The sooner you respond, the fresher the comments are in the reviewers' mind.
