+++ vignettes/b2-setting-up-and-using-rix-on-macos.Rmd
@@ -46,7 +46,7 @@ a company that provides services and tools built on Nix. Simply open a terminal
 and run the following line:
 
 
-```{sh eval = FALSE}
+```{sh, eval = FALSE}
 curl --proto '=https' --tlsv1.2 -sSf \
     -L https://install.determinate.systems/nix | \
      sh -s -- install
@@ -101,7 +101,7 @@ R session, and install `{rix}` if that's not already done. Because `{rix}` is
 not yet on CRAN, the easiest way is to install it from its r-universe:
 
 
-```{r eval = FALSE}
+```{r, eval = FALSE}
 install.packages("rix", repos = c("https://b-rodrigues.r-universe.dev",
   "https://cloud.r-project.org"))
 ```
@@ -112,6 +112,15 @@ to learn more.
 
 
 ## More macOS specificities
+
+### R Support for Apple Silicon in Nixpkgs
+
+In Nixpkgs, the M series processors of the ARM64 family, also known under
+AArch64, supports R since R version 3.5.3. Earlier versions of R will not 
+compile on modern Apple processor architectures for the correspoding Nixpkgs
+revisions at that time. Hence, the `darwin-aarch64` architecture has constrained
+backwards-compatibility.
+
 
 ### Shared libraries issue
 
