+++ dev/flat_with_nix.Rmd
@@ -570,7 +570,7 @@ recurse_find_check_globals <- function(expr, args_vec) {
     env_other <- lapply(result_list, "[", "env_other")
     env_other <- unlist_unname(env_other)
     
-    env_fun = lapply(result_list, "[", "env_fun")
+    env_fun <- lapply(result_list, "[", "env_fun")
     env_fun <- unlist_unname(env_fun)
     
     exports <- list(
