+++ dev/flat_get_latest.Rmd
@@ -35,3 +35,16 @@ get_latest <- function() {
   })
 }
 ```
+
+```{r, tests-get_latest}
+testthat::test_that("get_latest() fails as expected if error", {
+  is_internet_down <- function() {
+    !curl::has_internet()
+  }
+
+  testthat::local_mocked_bindings(is_internet_down = function(...) TRUE)
+  testthat::expect_error(get_latest())
+
+})
+
+```
