### 2024-03-28 14632fe
added test to flat_get_latest.Rmd
