### 2024-03-28 #625 @b-rodrigues
> I also found your ["What is nix" blog post](https://www.brodrigues.co/blog/2023-07-13-nix_for_r_part1/) particularly helpful for understanding what was actually happening under the hood. It seems likely to me that a good portion of your users will also be using `nix` for the first time, so it is worth making sure you explain or link out to explanations of how `nix` itself works at an intuitive level, in the interest of usability.

I’ve added a new section to the "Getting started vignette" that includes the introductory text of the blog post: https://github.com/b-rodrigues/rix/pull/153/commits/7542610d41e5e2f29475710b43b7d2458d41334b

> #### From goodpractice results
> Is there a reason for using `=` here `R/with_nix.R:564:13`?

No, fixed with: https://github.com/b-rodrigues/rix/pull/153/commits/8ed3a4a6e4727cea151d17dfde1a524745f45d89

> #### code comments
> I see a couple 'list versions' functions rely on a stored data file. Is there a way to get this information directly from the nix archive so you don't need to keep that up to date?

It would always be the same data, so we thought it might be better to provide that information with the package as a dataset. Also, it would work without an active internet connection.

> There are a fair number of internal functions that don't have inputs/outputs documented, I would suggest documenting those for your own/future maintainer's benefit down the road.

Indeed, this will be done in the coming days.

> Also, several todos still in comments.

These todos are more for future references of functionalities we want to add, but agreed, this should be moved to specific issues. Will address this in the coming days as well.

> #### test comments
> Seems like get_latest.R could be tested at least that it doesn't error and returns a hash, outside of environments where web service requests would fail?

Test added with: https://github.com/b-rodrigues/rix/pull/153/commits/14632fee59feda19365ae1381730c6a1636c1847

> Re: coverage—just to make sure I understand, the 25% coverage value is ignoring the tests that actually run `nix` commands since you would need an environment with `nix` installed to run them, which you don't get in CRAN or other non-containerized environments? It seems like it should be possible to make a Docker container with both R and `nix` installed to get a more realistic test environment, but maybe you've done this already.

Actually, the issue is more complex: we do have our tests running on Github Actions with nix installed (and also without, to mimick environments with and without nix), but for some complex reason, {covr} doesn’t run the tests that require Nix, these end up failing. So can’t compute the coverage.

We’ve collected the commits in this pull request: https://github.com/b-rodrigues/rix/pull/153/ and we’ll likely add more commits to address @assignUser points as well
