### 2024-03-28 #625 @b-rodrigues
@assignUser

regarding your suggestions for GHAs:

>GHAs are my specialty so some extended notes on extdata/run-pipeline.yml (most of this applies in general as well):

> Please explicitly add the contents: write [permission](https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs). This is technically not required as the token has the correct permisson but it is best practice to minimise token permissions.
> You should also do this on the workflows in .github/workflows/ with the minimal permissions each (e.g. contents: read)
> Always use at least tags for actions for stability, in an action that has write permission I would even highly recommend to explicitly use a commit hash. That ensures that even a malicous take over of one of the action repos does not allow a third-party to run code in your repo. Github's [dependabot](https://docs.github.com/en/code-security/dependabot/working-with-dependabot/automating-dependabot-with-github-actions) can handle updating shas and tags.
> Make sure tags/hashes are updated regularly (actions/checkout is at v4 but v2 is used in the workflow).
> secrets should always be assigned to envvars as close to where they are used as possible (usually step env:) and not for the entire workflow.
> I would recommend to use an explicit [output](https://docs.github.com/en/actions/using-workflows/workflow-commands-for-github-actions#setting-an-output-parameter) instead of continue-on-error as that still creates a big red error message in the job summary.

Do you happen to perhaps have a couple examples on hand I could take inspiration from? I’m struggling a bit
