### 2024-03-28 dd88d8a
combined two error messages into one flat_with_nix.Rmd
