### 2024-03-28 854a202
combined two error messages into one flat_with_nix.Rmd
