### 2024-03-28 #625 @b-rodrigues
> #### Code
> * I agree with @wdwatkins regarding the test coverage.

Same answer :)

> * There is a probably unintentional partial match in `flag_git_archive`: `In cran_pkgs$archive : partial match of 'archive' to 'archive_pkgs'`

thanks, solved with https://github.com/b-rodrigues/rix/pull/153/commits/922a2c013ba78ba39ac3d2747c1eeb45a0c1c95d

> * I like the use of {fledge} but personally don't think {fusen} is the right choice for a technically advanced package like {rix}. It is of course preference and up to you but I think {fusen} changes the established workflow of an R package dev quite a bit. I see how it can lower the bar for people to get started with creating packages and that's great. But I think in this case the people that are likely going to contribute (experienced devs) might not do it if that means they have to spend time getting used to {fusen} vs. just adding a quick patch. (I certainly looked through R/ and not dev/)

You may be right, but I think that the trade-off is worth it; fusen takes care of many annoying aspects of package development. But to meet devs not familiar with fusen halfway, I’ve changed `contributing.md` to state that it’s ok that PRs change files in R/ directly: https://github.com/b-rodrigues/rix/commit/1264b7c0c43d557c2f01b3a7aac82066519eabb2

> * The large error message in `with_rix` will never be reached due to the preceding `stopifnot`, maybe combine the two?
thanks, solved with https://github.com/b-rodrigues/rix/commit/854a20265a198c5f7bf4b82f8278dafecba095df

We’ll address the other comments later this week
