+++ .github/workflows/cachix-dev-env.yml
@@ -1,0 +1,30 @@
+name: "Update cachix cache for development environment"
+
+on:
+  push:
+    branches: [ master, main ]
+  schedule:
+    - cron: '30 0/6 * * *'
+
+jobs:
+  r-update-cachix:
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest]
+    steps:
+    - uses: actions/checkout@v4
+    - uses: cachix/install-nix-action@v25
+      with:
+        nix_path: nixpkgs=channel:nixos-unstable
+    - uses: cachix/cachix-action@v14
+      with:
+        name: b-rodrigues
+        # If you chose signing key for write access
+        # signingKey: '${{ secrets.CACHIX_SIGNING_KEY }}'
+        # If you chose API tokens for write access OR if you have a private cache
+        authToken: '${{ secrets.CACHIX_AUTH }}'
+    - run: | 
+        nix-build
+        nix-store -qR --include-outputs $(nix-instantiate default.nix) | cachix push b-rodrigues
+    - run: nix-shell --run "echo OK"
