+++ .github/workflows/cachix-dev-env.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches: [ master, main ]
   schedule:
-    - cron: '30 0/6 * * *'
+    - cron: '30 0/12 * * *'
 
 jobs:
   r-update-cachix:
