+++ .github/workflows/tests-r-via-nix.yaml
@@ -22,7 +22,7 @@ jobs:
 
       - uses: cachix/install-nix-action@v25
         with:
-          nix_path: nixpkgs=channel:nixos-unstable
+          nix_path: nixpkgs=https://github.com/rstats-on-nix/nixpkgs/archive/refs/heads/r-daily.tar.gz
 
       - uses: cachix/cachix-action@v14
         with:
