+++ .github/workflows/tests-r-via-nix.yaml
@@ -27,10 +27,6 @@ jobs:
       - uses: cachix/cachix-action@v14
         with:
           name: b-rodrigues
-          # If you chose signing key for write access
-          # signingKey: '${{ secrets.CACHIX_SIGNING_KEY }}'
-          # If you chose API tokens for write access OR if you have a private cache
-          authToken: '${{ secrets.CACHIX_AUTH }}'
       
       - name: devtools::test() via nix-shell
         run: nix-shell --pure -p cacert nix R rPackages.covr rPackages.codetools rPackages.xml2 rPackages.sys rPackages.devtools --run "Rscript -e 'devtools::test()'"
