### 2024-04-05 #625 @b-rodrigues
>with_nix is very verbose with ~ a terminal length of 'preamble' output of checks etc., I think a 'quite' option would be nice when running it repeatedly.

What would be the best way to do so? Is there some way that we could add this without needing to touch the overall structure of the function?
