+++ .github/workflows/cachix-dev-env.yml
@@ -17,6 +17,7 @@ jobs:
     - uses: cachix/install-nix-action@v25
       with:
         nix_path: nixpkgs=channel:nixos-unstable
+  
     - uses: cachix/cachix-action@v14
       with:
         name: b-rodrigues
@@ -24,7 +25,14 @@ jobs:
         # signingKey: '${{ secrets.CACHIX_SIGNING_KEY }}'
         # If you chose API tokens for write access OR if you have a private cache
         authToken: '${{ secrets.CACHIX_AUTH }}'
-    - run: | 
-        nix-build
-        nix-store -qR --include-outputs $(nix-instantiate default.nix) | cachix push b-rodrigues
+  
+    - run:   |
+         if [ "$RUNNER_OS" == "Linux" ]; then
+              nix-build --argstr system x86_64-linux
+              nix-build --argstr system aarch64-linux
+         else
+              nix-build
+         fi
+        
+    - run: nix-store -qR --include-outputs $(nix-instantiate default.nix) | cachix push b-rodrigues
     - run: nix-shell --run "echo OK"
