### 2024-04-06 b82ca33
test for get_latest()
