+++ R/get_latest.R
@@ -4,11 +4,17 @@
 #' @return A character. The commit hash of the latest nixpkgs-unstable revision
 #' @importFrom httr content GET stop_for_status
 #' @importFrom jsonlite fromJSON
+#' @importFrom curl has_internet
 #'
 #' @noRd
 get_latest <- function() {
   api_url <- "https://api.github.com/repos/NixOS/nixpkgs/commits?sha=nixpkgs-unstable"
 
+  is_online <- curl::has_internet()
+
+  if(!is_online){
+    stop("ERROR! You don't seem to be connected to the internet.")
+  } else {
   tryCatch({
     response <- httr::GET(url = api_url)
     httr::stop_for_status(response)
@@ -19,4 +25,7 @@ get_latest <- function() {
     cat("Error:", e$message, "\n")
     return(NULL)
   })
+
+  }
+
 }
