+++ tests/testthat/test-get_latest.R
@@ -1,12 +1,11 @@
 # WARNING - Generated by {fusen} from dev/flat_get_latest.Rmd: do not edit by hand
 
-testthat::test_that("get_latest() fails as expected if error", {
-  is_internet_down <- function() {
-    !curl::has_internet()
-  }
+testthat::test_that("get_latest() fails gracefully if no internet", {
 
-  testthat::local_mocked_bindings(is_internet_down = function(...) TRUE)
-  testthat::expect_error(get_latest())
+with_mock(`curl::has_internet` = function(...) FALSE,
+          expect_error(get_latest(), "You don't seem to be connected")
+          )
+
 
 })
 
