### 2024-04-07 65a4a19
Update flat_rix.Rmd: refactor rix()
