+++ dev/flat_rix.Rmd
@@ -152,49 +152,10 @@ for more details.")
 
   rix_call <- match.call()
 
-  generate_rix_call <- function(rix_call, nix_revision){
-
-    rix_call$r_ver <- nix_revision
-
-    rix_call <- paste0("# >", deparse1(rix_call))
-
-    gsub(",", ",\n#  >", rix_call)
-  }
-
   # Get the rix version
   rix_version <- utils::packageVersion("rix")
 
-  generate_header <- function(rix_version,
-                              nix_revision,
-                              r_ver_text,
-                              rix_call){
 
-    if(identical(Sys.getenv("TESTTHAT"), "true")){
-      sprintf('
-let
- pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/%s.tar.gz") {};
-',
-nix_revision)
-      } else {
-        sprintf('# This file was generated by the {rix} R package v%s on %s
-# with following call:
-%s
-# It uses nixpkgs\' revision %s for reproducibility purposes
-# which will install R %s
-# Report any issues to https://github.com/b-rodrigues/rix
-let
- pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/%s.tar.gz") {};
-',
-rix_version,
-Sys.Date(),
-generate_rix_call(rix_call, nix_revision),
-nix_revision,
-r_ver_text,
-nix_revision
-)
-    }
-
- }
 
   # Now we need to generate all the different sets of packages
   # to install. Let's start by the CRAN packages, current
@@ -415,6 +376,7 @@ flag_rpkgs
 
 }
 ```
+
 ```{r, tests-rix_rstudio}
 testthat::test_that("rix(), ide is 'rstudio', Linux", {
 
@@ -546,6 +508,59 @@ testthat::test_that("Quarto gets added to sys packages", {
 })
 
 ```
+
+```{r, function-generate_rix_call}
+#' generate_rix_call Internal function used to generate the call to `rix()` as shown in `default.nix`
+#' @param rix_call Character, call to rix().
+#' @param nix_revision Character, nixpkgs revision to be used.
+generate_rix_call <- function(rix_call, nix_revision){
+
+  rix_call$r_ver <- nix_revision
+
+  rix_call <- paste0("# >", deparse1(rix_call))
+
+  gsub(",", ",\n#  >", rix_call)
+}
+```
+
+```{r, function-generate_header}
+#' generate_header Internal function used to generate the header of the `default.nix` file.
+#' @param rix_version Character, current, version of the package
+#' @param nix_revision Character, nixpkgs revision to be used.
+#' @param r_ver_text Character, human-readable text giving the version.
+#' @param rix_call Character, call to rix().
+generate_header <- function(rix_version,
+                            nix_revision,
+                            r_ver_text,
+                            rix_call){
+
+  if(identical(Sys.getenv("TESTTHAT"), "true")){
+     sprintf('
+let
+ pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/%s.tar.gz") {};
+',
+nix_revision)
+ } else {
+   sprintf('# This file was generated by the {rix} R package v%s on %s
+# with following call:
+%s
+# It uses nixpkgs\' revision %s for reproducibility purposes
+# which will install R %s
+# Report any issues to https://github.com/b-rodrigues/rix
+let
+ pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/%s.tar.gz") {};
+',
+rix_version,
+Sys.Date(),
+generate_rix_call(rix_call, nix_revision),
+nix_revision,
+r_ver_text,
+nix_revision
+)
+  }
+}
+```
[…]
