### 2024-04-07 #625 @philipp-baumann
>     * I was surprised that you are not using {rlang} and quosures, it has been a while since I have used them but I would think they would be a good fit here. (but maybe not, it's been a while since I have used them 'under the hood')

When I started `with_nix()`I thought about it, but then decided against it. Base R has very good computing on the language foundations, which suffices for the functionality we want to provide in `with_nix()`. For a package that facilitates reproducibility -- although nix provides all the ways to do so -- I prefer to reduce the amount of dependencies whenever possible (we for sure also have room to improve this in current {rix} ;-)).
