### 2024-04-07 #625 @philipp-baumann
@assignUser

>     * It seems like the `codetools::check` is run quite a lot, I assume it's to make debugging easier?

Exactly, it's another way to get feedback on possible scoping problems of variables etc. There is minor things we can deal with and runs in code exported to the nix session: for example, we support when an `expr` function or the ones downstream the call stack contain global variables . Here, The codetools `checkUsage()` also informs but the code runs because we also export the object in the global environment of the current R session invoking `with_nix()`. `with_nix()` is able to detect and export complex call stacks, because it recursively calls globals found by `findGlobals()`. Along this, we also repeat `checkUsage()`,  which "checks a single R closure".
