+++ DESCRIPTION
@@ -14,9 +14,10 @@ URL: https://b-rodrigues.github.io/rix/
 BugReports: https://github.com/b-rodrigues/rix
 Depends: 
     R (>= 2.10)
-Imports: 
+Imports:
     codetools,
     curl,
+    dplyr,
     httr,
     jsonlite,
     sys,
