### 2024-04-12 #625 @philipp-baumann
@b-rodrigues @assignUser

> > * I like the use of {fledge} but personally don't think {fusen} is the right choice for a technically advanced package like {rix}. It is of course preference and up to you but I think {fusen} changes the established workflow of an R package dev quite a bit. I see how it can lower the bar for people to get started with creating packages and that's great. But I think in this case the people that are likely going to contribute (experienced devs) might not do it if that means they have to spend time getting used to {fusen} vs. just adding a quick patch. (I certainly looked through R/ and not dev/)
>
> You may be right, but I think that the trade-off is worth it; fusen takes care of many annoying aspects of package development. But to meet devs not familiar with fusen halfway, I’ve changed `contributing.md` to state that it’s ok that PRs change files in R/ directly: [b-rodrigues/rix@1264b7c](https://github.com/b-rodrigues/rix/commit/1264b7c0c43d557c2f01b3a7aac82066519eabb2)

I see there are advantages for {fusen} for smaller packages, but as things grow and we mainly want to take away complexity, I agree with @assignUser and am in favor of completely removing it in the workflow. There is `fusen::inflate_all_no_check()` which I mostly use use to safe time, but when developing `with_nix()` it meant mostly more overhead to wait when doing `browser()` etc. Adding it would also mean that we force newcomers and possible contributers to to use a nix development environment. I have experienced different transformation behaviors, and I believe there are also formatting changes that will come since fusen hasn't completely stabilized.
