+++ dev/z-bleeding_edge.Rmd
@@ -53,7 +53,7 @@ builds packages and these get cached. So if you’re using Nix, pre-compiled bin
 built from source. This is not the case if you use our fork, **unless** you also use the cache that we provide, courtesy
 of [cachix.org](cachix.org). Cachix provides a cache for your own projects and works seamlessly with the official Nix cache.
 However, we are limited in space, and cannot possible cache all the CRAN packages. So only the most popular packages
-get built and cached, and hopefully the packages you need are part of it. To use the cache, run the following commands on
+get built and cached, and hopefully the packages you need are part of this limited set. To use the cache, run the following commands on
 your computer. First, install the Cachix client:
 
 ```
@@ -67,3 +67,18 @@ cachix use rstats-on-nix
 ```
 
 that’s it! Packages now not only will get pulled from the official cache, but also from the dedicated `rstats-on-nix` cache.
+
+## Using your own cache
+
+You could also set up your own cache for your own purposes. If you’re using Github, you could take advantage of Github Actions
+to periodically build a dedicated development environment and cache it. Simply define a `default.nix` file and put it on the root of
+your Github repository. Then, define a Github Actions workflow file to build the environment periodically and push the built
+packages on Cachix (create a Cachix account first). You can take inspiration from the workflow file of `{rix}`, found 
+[here](https://github.com/b-rodrigues/rix/blob/master/.github/workflows/cachix-dev-env.yml). As you can see on 
+[line 27](https://github.com/b-rodrigues/rix/blob/517dc636da752e4d92d0ed7251c4d4d01b14c088/.github/workflows/cachix-dev-env.yml#L27)
+you need to define a repository secret called `CACHIX_AUTH`. You need to create an authentication token on Cachix, and then
+paste it into the repository secrets of Github Actions. Then, on [line 37](https://github.com/b-rodrigues/rix/blob/517dc636da752e4d92d0ed7251c4d4d01b14c088/.github/workflows/cachix-dev-env.yml#L37)
+simply replace `rstats-on-nix` with your cache’s name. This environment will be built periodically on Github Actions
+according to the instructions in the `default.nix` file, which you can then use on your local machine to develop your package
+(or whatever project). Don’t forget to enable Cachix on your computer and to call `nix-build` locally to install the 
+environment on your machine!
