### 2024-04-28 758c7b7
Add more IDE-s: rserver and radian
Adds support to RStudio server and radian. It is done in a way that
avoids core repetition, and allows easy addition of future wrappers.
