### 2024-04-29 #625 @assignUser
> What would be the best way to do so? Is there some way that we could add this without needing to touch the overall structure of the function?

Use a wrapper function around `cat` with a `quiet` or something arg that skips printing and add that to with_nix too and pass it through as the default arg. You can give that to `codetools::check_usage` via `report` too :)
(or something like that, I just had a quick glimpse at the code)
