### 2024-04-29 #625 @assignUser
@b-rodrigues for better notification :)

> This is required because the underlying Nix function fetchgit requires both the repo and the commit. It might be more cumbersome from a UX perspective,

Ok in that case leave branch out as it's redundant or allow adding repo + branch only and fetch the HEAD of the branch for the user.

Looking at [master](https://github.com/b-rodrigues/rix/blob/master/R/with_nix.R#L206) I still see the stopifnot that you removed in the linked commit above?

With regards to GHA examples you can take a look in https://github.com/apache/arrow/tree/main/.github/workflows we have lot's of workflows that do a lot of things ^^ for more piece meal examples I'd recommend the [docs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions). You can also ping me on a PR with specific questions if you are still stuck :)
