### 2024-05-06 #625 @llrs
@ldecicco-USGS I was asked to [provide some feedback](https://fosstodon.org/@specphil/112395876328309796). I do have some comments (see reply in mastodon) but there might be more if I look into the code.

I am not sure if it is best to post it during the review or once it ends.
In any case, please do not wait for my feedback and let me know what would you (and all the involved people) prefer.
