### 2024-05-06 #625 @ldecicco-USGS
Hi @llrs ! Thanks for the offer. We do have 2 reviewers' already, I think creating an issue on the nix repository would probably be the best plan.

@philipp-baumann and @b-rodrigues  are you at a point where I should be verifying that the reviewers are satisfied with your response to the reviews? It looks like there's still some back and forth happening (?), but mostly ready?
