+++ .github/workflows/R-CMD-check.yaml
@@ -8,6 +8,9 @@ on:
 
 name: R-CMD-check
 
+permissions:
+  contents: read
+
 jobs:
   R-CMD-check:
     runs-on: ${{ matrix.config.os }}
