+++ .github/workflows/cachix-dev-env.yml
@@ -6,6 +6,9 @@ on:
   schedule:
     - cron: '30 0/12 * * *'
 
+permissions:
+  contents: read
+
 jobs:
   r-update-cachix:
     runs-on: ${{ matrix.os }}
