### 2024-05-10 e3966b0
nix permissions
