+++ .github/workflows/tests-r-via-system.yaml
@@ -8,6 +8,9 @@ on:
 
 name: tests-r-via-system
 
+permissions:
+  contents: read
+
 jobs:
   devtools:
     runs-on: ${{ matrix.os }}
