### 2024-05-10 #625 @b-rodrigues
> @b-rodrigues for better notification :)
>
> > This is required because the underlying Nix function fetchgit requires both the repo and the commit. It might be more cumbersome from a UX perspective,
>
> Ok in that case leave branch out as it's redundant or allow adding repo + branch only and fetch the HEAD of the branch for the user.
>
> Looking at [master](https://github.com/b-rodrigues/rix/blob/master/R/with_nix.R#L206) I still see the stopifnot that you removed in the linked commit above?
>
> With regards to GHA examples you can take a look in https://github.com/apache/arrow/tree/main/.github/workflows we have lot's of workflows that do a lot of things ^^ for more piece meal examples I'd recommend the [docs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions). You can also ping me on a PR with specific questions if you are still stuck :)

So I've tried with the following commits:

https://github.com/b-rodrigues/rix/commit/82890b30cb3315e264993da344cd46a8d1ee7aab

https://github.com/b-rodrigues/rix/commit/3b8335fc473e5d75185808749f530fd94f1322ff

https://github.com/b-rodrigues/rix/commit/fb4a8f27784d33f9ca9f98359f95382870b9dd2c

https://github.com/b-rodrigues/rix/commit/e3966b008f7a069789ba66b6136b12413ffd9214

https://github.com/b-rodrigues/rix/commit/a93771bea03d1f7d0be64275c4d0a681eb7db628

is that ok like this?
