+++ dev/flat_rix.Rmd
@@ -30,10 +30,13 @@ you use another IDE, you can leave the "ide" argument blank:
 #'   are not R packages such as command line applications for example. You can look for
 #'   available software on the NixOS website \url{https://search.nixos.org/packages?channel=unstable&from=0&size=50&sort=relevance&type=packages&query=}
 #' @param git_pkgs List. A list of packages to install from Git. See details for more information.
-#' @param tex_pkgs Vector of characters. A set of tex packages to install. Use this if you need to compile `.tex` documents, or build PDF documents using Quarto. If you don't know which package to add, start by adding "amsmath". See the Vignette "Authoring LaTeX documents" for more details.
+#' @param tex_pkgs Vector of characters. A set of tex packages to install.
+#'   Use this if you need to compile `.tex` documents,
+#'   or build PDF documents using Quarto. If you don't know which package to add,
+#'   start by adding "amsmath". See the Vignette "Authoring LaTeX documents" for more details.
 #' @param ide Character, defaults to "other". If you wish to use RStudio to work
 #'   interactively use "rstudio" or "rserver" for the server version. Use "code"
-#    for Visual Studio Code. You can also use "radian", an interactive REPL. For
+#'   for Visual Studio Code. You can also use "radian", an interactive REPL. For
 #'   other editors, use "other". This has been tested with RStudio, VS Code and
 #'   Emacs. If other editors don't work, please open an issue.
 #' @param project_path Character, defaults to the current working directory.
