+++ dev/z-binary_cache.Rmd
@@ -55,4 +55,49 @@ if you add a packages to it, or re-run it, you’ll end up with a new `default.n
 rebuild the environment. Again, depending on the packages you include, this could take quite some time to build.
 To generate a Github Actions workflow file that will build the environment on Github Actions, run 
 `rix::ga_cachix(cache_name = "", path = "")` where `cache_name` is the name of the cache you made on Cachix, and
-`path` is the path to the `default.nix` file generated by `generate_env.R`.
+`path` is the path to the `default.nix` file generated by `generate_env.R`. By default, the environment gets rebuilt
+every time you push changes to the `master/main` branch of your repository, but you can also re-build the environment
+periodically, by changing these lines:
+
+```
+on:
+  push:
+    branches: [ master, main ]
+```
+
+to this:
+
+```
+on:
+  push:
+    branches: [ master, main ]
+  schedule:
+    - cron: '30 0 * * *'
+```
+
+By using [cron syntax](https://en.wikipedia.org/wiki/Cron#Overview) you can specify how often you want the environment
+to be re-built. This is can be useful if you need to develop against the current state of CRAN every day (for instance,
+for package development). To then use the cache, you need to open an account on [Cachix](https://www.cachix.org/). The free
+tier includes 5GB of space, which is more than enough for several development environments. Once your account is done,
+create a personal auth token so that the workflow can authenticate to your Cachix account. Then, copy this token,
+and go to your Github repository’s settings, then *Secrets and variables > Actions* and add a new repository secret.
+Copy the token in the `Secret` field and name the secret `CACHIX_AUTH`. Now the action can authenticate to Cachix, and
+push the binaries it builds!
+
+To use your personalized cache on your computers, run the following commands on
+your computer. First, install the Cachix client:
+
+```
+nix-env -iA cachix -f https://cachix.org/api/v1/install
+```
+
+then use the cache:
+
+```
+cachix use your-cache-name
+```
+
+Anyone can pull binaries from your cache, so if you work in a team, you can ensure everyone can benefit from it.
+
+Take a look at [this package’s](https://github.com/b-rodrigues/rix/blob/master/.github/workflows/cachix-dev-env.yml)
+repository for an example of how this is done in practice.
