+++ R/make_nixpkgs_url.R
@@ -1,0 +1,21 @@
+# WARNING - Generated by {fusen} from dev/flat_make_nixpkgs_url.Rmd: do not edit by hand
+
+#' find_rev Find the right Nix revision
+#' @param r_version Character. R version to look for, for example, "4.2.0". If a nixpkgs revision is provided instead, this gets returned.
+#' @return A character. The url to use
+#'
+#' @examples
+#' find_rev("4.2.0")
+#' @noRd
+make_nixpkgs_url <- function(r_ver){
+
+  if(r_ver %in% c("bleeding_edge", "froze_edge")){
+    github_repo <- "rstats-on-nix/nixpkgs/"
+  } else {
+    github_repo <- "NixOS/nixpkgs/"
+  }
+
+  paste0("https://github.com/", github_repo, "archive/", get_latest(r_ver), ".tar.gz")
+
+}
+
