+++ dev/flat_make_nixpkgs_url.Rmd
@@ -1,0 +1,33 @@
+---
+title: "rix"
+output: html_document
+editor_options:
+  chunk_output_type: console
+---
+
+```{r development, include=FALSE}
+library(testthat)
+```
+
+This function generates the correct URL for the `nixpkgs` repo to use.
+
+```{r function-find_rev}
+#' find_rev Find the right Nix revision
+#' @param r_version Character. R version to look for, for example, "4.2.0". If a nixpkgs revision is provided instead, this gets returned.
+#' @return A character. The url to use
+#'
+#' @examples
+#' find_rev("4.2.0")
+make_nixpkgs_url <- function(r_ver){
+
+  if(r_ver %in% c("bleeding_edge", "froze_edge")){
+    github_repo <- "rstats-on-nix/nixpkgs/"
+  } else {
+    github_repo <- "NixOS/nixpkgs/"
+  }
+
+  paste0("https://github.com/", github_repo, "archive/", get_latest(r_ver), ".tar.gz")
+
+}
+
+```
