+++ dev/flat_rix.Rmd
@@ -161,49 +161,6 @@ for more details.")
   rix_version <- utils::packageVersion("rix")
 
 
-
-# If `r_ver = "bleeding_edge"`, then this function gets used instead
-generate_bleeding_edge_header <- function(rix_version,
-                                          rix_call){
-sprintf('# This file was generated by the {rix} R package v%s on %s
-# with following call:
-%s
-# It uses a bleeding edge fork of nixpkgs! Only use this if you 
-# really need bleeding edge packages. The environment generated
-# by this file is NOT reproducible! If you need reproducible
-# environments, replace "bleeding_edge" by "latest", or a
-# specific R version.
-# Report any issues to https://github.com/b-rodrigues/rix
-let
- pkgs = import (fetchTarball "https://github.com/rstats-on-nix/nixpkgs/archive/refs/heads/r-daily.tar.gz") {};
-',
-rix_version,
-Sys.Date(),
-generate_rix_call(rix_call, "bleeding_edge")
-)
-}
-
-# If `r_ver = "bleeding_edge"`, then this function gets used instead
-generate_bleeding_edge_header <- function(rix_version,
-                                          rix_call){
-sprintf('# This file was generated by the {rix} R package v%s on %s
-# with following call:
-%s
-# It uses a bleeding edge fork of nixpkgs! Only use this if you 
-# really need bleeding edge packages. The environment generated
-# by this file is NOT reproducible! If you need reproducible
-# environments, replace "bleeding_edge" by "latest", or a
-# specific R version.
-# Report any issues to https://github.com/b-rodrigues/rix
-let
- pkgs = import (fetchTarball "https://github.com/rstats-on-nix/nixpkgs/archive/refs/heads/r-daily.tar.gz") {};
-',
-rix_version,
-Sys.Date(),
-generate_rix_call(rix_call, "bleeding_edge")
-)
-}
-
   # Now we need to generate all the different sets of packages
   # to install. Let's start by the CRAN packages, current
   # and archived. The function below builds the strings.
