+++ R/make_nixpkgs_url.R
@@ -1,11 +1,11 @@
 # WARNING - Generated by {fusen} from dev/flat_make_nixpkgs_url.Rmd: do not edit by hand
 
-#' make_nixpkgs_url Find the right Nix revision
+#' find_rev Find the right Nix revision
 #' @param r_version Character. R version to look for, for example, "4.2.0". If a nixpkgs revision is provided instead, this gets returned.
 #' @return A character. The url to use
 #'
 #' @examples
-#' make_nixpkgs_url("4.2.0")
+#' find_rev("4.2.0")
 #' @noRd
 make_nixpkgs_url <- function(r_ver){
 
