+++ vignettes/get-started.Rmd
@@ -2,7 +2,7 @@
 title: "Get started"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Get started}
+  %\VignetteIndexEntry{get-started}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -14,12 +14,12 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r}
+```{r setup}
 library(rix)
 ```
 
 <!-- WARNING - This vignette is generated by {fusen} from dev/flat_make_nixpkgs_url.Rmd: do not edit by hand -->
 
-
-
 This function generates the correct URL for the `nixpkgs` repo to use.
+
+
