+++ dev/flat_get_latest.Rmd
@@ -30,14 +30,12 @@ get_latest <- function(r_ver) {
 
   if(!is_online){
     stop("ERROR! You don't seem to be connected to the internet.")
-  } else {
-  if(r_version == "bleeding_edge"){
+  } else if(r_version == "bleeding_edge"){
     latest_commit <- "refs/heads/r-daily"
   } else {
     latest_commit <- get_latest_or_frozen(r_ver)
   }
   latest_commit
-  }
 }
 ```
 
@@ -59,8 +57,6 @@ get_latest_or_frozen <- function(r_ver) {
     cat("Error:", e$message, "\n")
     return(NULL)
   })
-
-  }
 
 }
 ```
