+++ vignettes/b1-setting-up-and-using-rix-on-linux-and-windows.Rmd
@@ -1,0 +1,196 @@
+---
+title: "b1 - Setting up and using rix on Linux and Windows"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{b1-setting-up-and-using-rix-on-linux-and-windows}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(rix)
+```
+
+<!-- WARNING - This vignette is generated by {fusen} from dev/b1-linux_win.Rmd: do not edit by hand -->
+
+*This vignette will discuss Linux and Windows-specific topics. If you're not
+using either of these systems, you can ignore this vignette, and read the
+`vignette("b2-setting-up-and-using-rix-on-macos")`
+vignette instead.*
+
+
+## Introduction
+
+When it comes to Nix, there are really only two supported operating systems:
+macOS and Linux distributions. Windows is "supported" because it is actually
+running Linux thanks to WSL2. In practice this means that Linux distributions
+and Windows can be considered one system, and macOS another, separate, system.
+Because Windows is really running Linux under the hood thanks to WSL2, this
+means that WSL2 needs to be running on your Windows system before you attempt to
+install Nix. But it is important to know that you can run `{rix}` even if you
+don't have Nix installed, which means you can generate Nix expressions, you
+just can't build them. So if you can’t install Nix on your system, but have 
+R already installed, you can skip to the last section of this vignette to 
+simply install the `{rix}` package.
+
+
+## Installing Nix
+
+### Windows pre-requisites
+
+If you are on Windows, you need the Windows Subsystem for Linux 2 (WSL2) to run
+Nix. If you are on a recent version of Windows 10 or 11, you can simply run this
+as an administrator in PowerShell:
+
+```
+wsl --install
+```
+
+You can find further installation notes at [this official MS
+documentation](https://learn.microsoft.com/en-us/windows/wsl/install).
+
+We recommend to activate systemd in Ubuntu WSL2, mainly because this supports 
+other users than `root` running Nix. To set this up, please do as outlined
+[this official Ubuntu blog entry](https://ubuntu.com/blog/ubuntu-wsl-enable-systemd):
+
+```sh
+
+# in WSL2 Ubuntu shell
+
+sudo -i
+nano /etc/wsl.conf
+
+# add this entry
+
+[boot]
+systemd=true
+
+# then restart running instance from PowerShell
+
+wsl --shutdown
+
+# relaunch Ubuntu WSL2
+
+```
+
+Afterwards, you can install Nix like business as usual. You can proceed with the
+Determinate Systems installer.
+
+
+### The Determinate Systems installer
+
+You can use `{rix}` to generate Nix expressions even if you don't have Nix
+installed on your system, but obviously, you need to install Nix if you actually
+want to build the defined development environment and use them. Installing (and
+uninstalling) Nix is quite simple, thanks to the installer from [Determinate
+Systems](https://determinate.systems/posts/determinate-nix-installer), a company
+that provides services and tools built on Nix. Simply open a terminal and run
+the following line (on Windows, if you cannot or have decided not to activate
+systemd, then you have to append `--init none` to the command. You can find more
+details about this on [The Determinate Nix
+Installer page](https://github.com/DeterminateSystems/nix-installer)):
+
+
+```{sh parsermd-chunk-1, eval = FALSE}
+curl --proto '=https' --tlsv1.2 -sSf \
+    -L https://install.determinate.systems/nix | \
+     sh -s -- install
+```
+
+Once you have Nix installed, you can build the expressions you generate with `{rix}`!
+
[…]
