+++ vignettes/a-getting-started.Rmd
@@ -2,7 +2,7 @@
 title: "a - Getting started"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{a - Getting started}
+  %\VignetteIndexEntry{a-getting-started}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -14,12 +14,11 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r}
+```{r setup}
 library(rix)
 ```
 
 <!-- WARNING - This vignette is generated by {fusen} from dev/a-getting_started.Rmd: do not edit by hand -->
-
 
 ## Non-technical intro to Nix
 
@@ -124,6 +123,7 @@ will get exactly, byte by byte, the same environment as you when you initially
 started the project. And this also regardless of the operating system that is
 used.
 
+
 ## The Nix package manager
 
 Nix is a package manager that can be installed on your computer (regardless of
@@ -154,6 +154,7 @@ packages' repository (called `nixpkgs`) for our project. Pinning a revision
 ensures that every package that Nix installs will always be at exactly the same
 versions, regardless of when in the future the packages get installed.
 
+
 ## rix workflow
 
 The idea of `{rix}` is for you to declare the environment you need using the
@@ -181,14 +182,13 @@ time read through the vignettes and you should be fine.
 - the path to save the `default.nix` file (by default the current working directory)
 
 For example:
-```{r}
-#| eval: no
 
+
+```{r parsermd-chunk-1, eval = FALSE}
 rix(r_ver = "latest",
     r_pkgs = c("dplyr", "chronicler"),
     ide = "other")
 ```
-
 
 The call above writes a `default.nix` file in the current working directory.
 This `default.nix` can in turn be used by Nix to build an environment containing
@@ -197,6 +197,7 @@ the latest version of R, with the `{dplyr}` and `{chronicler}` packages.
 Take note of the `ide = "other"` argument: this argument, and the values it
 can take, are further discussed in the vignette `vignette("e-interactive-use")`
 but continue reading this vignette and then vignettes numbered by a "d".
+
 
 ### Using default.nix files
 
@@ -231,3 +232,4 @@ installed on your system.
 
 - If you’re running either Linux or Windows, read the Linux or Windows vignette: `vignette("b1-setting-up-and-using-rix-on-linux-and-windows")`
 - If you’re running macOS, read the macOS vignette: `vignette("b2-setting-up-and-using-rix-on-macos")`
+
