+++ DESCRIPTION
@@ -27,7 +27,7 @@ Suggests:
     testthat
 VignetteBuilder: 
     knitr
-Config/fusen/version: 0.6.0
+Config/fusen/version: 0.5.2.9000
 Config/testthat/edition: 3
 Encoding: UTF-8
 LazyData: true
