+++ tests/testthat/test-rix.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_rix.Rmd: do not edit by hand
-
 testthat::test_that("rix(), ide is 'rstudio', Linux", {
 
   skip_if(Sys.info()["sysname"] == "Darwin")
