+++ tests/testthat/test-rix_init.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_rix_init.Rmd: do not edit by hand
-
 testthat::test_that("Snapshot test of rix_init()", {
 
   path_env_nix <- tempdir()
