+++ tests/testthat/test-tar_nix_ga.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_cicd.Rmd: do not edit by hand
-
 testthat::test_that("Snapshot test of all features of rix", {
 
   testthat::announce_snapshot_file("cicd/run-pipeline.yaml")
