+++ R/get_latest.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_get_latest.Rmd: do not edit by hand
-
 #' get_latest Get the latest R version and packages
 #' @return A character. The commit hash of the latest nixpkgs-unstable revision
 #' @importFrom httr content GET stop_for_status
