+++ R/fetchgit.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_fetchers.Rmd: do not edit by hand
-
 #' fetchgit Downloads and installs a package hosted of Git
 #' @param git_pkg A list of four elements: "package_name", the name of the package, "repo_url", the repository's url, "branch_name", the name of the branch containing the code to download and "commit", the commit hash of interest. 
 #' @return A character. The Nix definition to download and build the R package from Github.
