+++ tests/testthat/test-detect_os.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_get_os.Rmd: do not edit by hand
-
 testthat::test_that("Test that generate_locale_archive fails", {
   testthat::expect_error(
     generate_locale_archive("Plan9"),
