+++ R/detect_versions.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_cran_archive.Rmd: do not edit by hand
-
 #' detect_versions Detects if CRAN packages need to be downloaded from the archive.
 #' @param r_pkgs A list of packages, to get from CRAN (either current packages or archived packages).
 #' @return A list of two elements, "cran_packages" and "archive_packages"
