+++ tests/testthat/test-with_nix.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_with_nix.Rmd: do not edit by hand
-
 testthat::test_that("Testing `with_nix()` if Nix is installed", {
 
   skip_on_covr()
