+++ R/get_sri_hash_deps.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_get_sri_hash_deps.Rmd: do not edit by hand
-
 #' get_sri_hash_deps Get the SRI hash of the NAR serialization of a Github repo
 #' @param repo_url A character. The URL to the package's Github repository or to the `.tar.gz` package hosted on CRAN.
 #' @param branch_name A character. The branch of interest, NULL for archived CRAN packages.
