+++ R/rix_init.R
@@ -1,8 +1,6 @@
-# WARNING - Generated by {fusen} from dev/flat_rix_init.Rmd: do not edit by hand
-
 #' Initiate and maintain an isolated, project-specific, and runtime-pure R
 #' setup via Nix.
-#' 
+#'
 #' Creates an isolated project folder for a Nix-R configuration. `rix::rix_init()`
 #' also adds, appends, or updates with or without backup a custom `.Rprofile` 
 #' file with code that initializes a startup R environment without system's user 
