+++ tests/testthat/test-get_sri_hash_deps.R
@@ -1,5 +1,3 @@
-# WARNING - Generated by {fusen} from dev/flat_get_sri_hash_deps.Rmd: do not edit by hand
-
 testthat::test_that("get_sri_hash_deps returns correct sri hash and dependencies of R packages", {
   testthat::expect_equal(
               get_sri_hash_deps("https://github.com/rap4all/housing/",
