+++ tests/testthat/test-nix_build.R
@@ -1,15 +1,12 @@
-# WARNING - Generated by {fusen} from dev/flat_nix_build.Rmd: do not edit by hand
-
-
 testthat::test_that("Testing that `nix_build()` builds derivation", {
-  
+
   if (isFALSE(is_nix_rsession())) {
     # needed for the GitHub test runners with system's R
     set_nix_path()
   }
-  
+
   skip_if_not(nix_shell_available())
-  
+
   skip_on_covr()
 
   path_subshell <- tempdir()
