+++ vignettes/z-advanced-topic-building-an-environment-for-literate-programming.Rmd
@@ -14,7 +14,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-```{r setup}
+```{r, include=FALSE}
 library(rix)
 ```
 
