+++ vignettes/get-started.Rmd
@@ -1,25 +1,0 @@
----
-title: "Get started"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{get-started}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-```{r setup}
-library(rix)
-```
-
-<!-- WARNING - This vignette is generated by {fusen} from dev/flat_make_nixpkgs_url.Rmd: do not edit by hand -->
-
-This function generates the correct URL for the `nixpkgs` repo to use.
-
-
