### 2024-05-19 121b9b7
fixed make_nixpkgs_url
