### 2024-05-19 0fca7af
basic functionality seems to be working, now on to tests
