+++ R/rix.R
@@ -130,10 +130,13 @@ for more details.")
     r_ver_text <- paste0("version ", r_ver)
   }
 
-  # Find the Nix revision to use
-  # TODO: don’t just return nix revision
-  # but write a function to return the whole url
-  nix_revision <- find_rev(r_ver)
+  # Find url to use
+  # In case of bleeding or frozen edge, the rstats-on-nix/nixpkgs
+  # fork is used. Otherwise, upstream NixOS/nixpkgs
+  nix_url <- make_nixpkgs_url(r_ver)
+
+  # Get the revision
+  nix_revision <- gsub("(.*archive\\/)|(\\.tar\\.gz)", "", nix_url)
 
   project_path <- file.path(project_path)
 
@@ -335,6 +338,7 @@ flag_rpkgs
   default.nix <- paste(
     generate_header(rix_version,
                     nix_revision,
+                    nix_url,
                     r_ver_text,
                     rix_call),
     generate_rpkgs(cran_pkgs$rPackages),
@@ -364,11 +368,11 @@ flag_rpkgs
 
 #' generate_rix_call Internal function used to generate the call to `rix()` as shown in `default.nix`
 #' @param rix_call Character, call to rix().
-#' @param nix_revision Character, nixpkgs revision to be used.
+#' @param nix_url Character, nixpkgs revision to be used.
 #' @noRd
-generate_rix_call <- function(rix_call, nix_revision){
+generate_rix_call <- function(rix_call, nix_url){
 
-  rix_call$r_ver <- nix_revision
+  rix_call$r_ver <- nix_url
 
   rix_call <- paste0("# >", deparse1(rix_call))
 
@@ -383,15 +387,16 @@ generate_rix_call <- function(rix_call, nix_revision){
 #' @noRd
 generate_header <- function(rix_version,
                             nix_revision,
+                            nix_url,
                             r_ver_text,
                             rix_call){
 
   if(identical(Sys.getenv("TESTTHAT"), "true")){
      sprintf('
 let
- pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/%s.tar.gz") {};
+ pkgs = import (fetchTarball "%s") {};
 ',
-nix_revision)
+nix_url)
  } else {
    sprintf('# This file was generated by the {rix} R package v%s on %s
 # with following call:
@@ -400,14 +405,14 @@ nix_revision)
 # which will install R %s
 # Report any issues to https://github.com/b-rodrigues/rix
 let
- pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/%s.tar.gz") {};
+ pkgs = import (fetchTarball "%s") {};
 ',
 rix_version,
 Sys.Date(),
-generate_rix_call(rix_call, nix_revision),
+generate_rix_call(rix_call, nix_url),
 nix_revision,
 r_ver_text,
-nix_revision
+nix_url
 )
   }
 }
