### 2024-05-19 2a4a1a1
tests for bleeding and frozen edge passing LFG
