+++ vignettes/z-binary_cache.Rmd
@@ -1,0 +1,131 @@
+---
+title: "z - Advanced topic: Rolling out your own binary cache"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{z-advanced-topic-rolling-out-your-own-binary-cache}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r, include=FALSE}
+library(rix)
+```
+
+*We recommend you first read `vignette("z-bleeding_edge")` before this one.*
+
+## Introduction
+
+Most of the software that the Nix package manager installs gets built from
+source, which can take a very long time. For example, installing R or RStudio
+from source could take up several hours, dependending on your hardware. Some R
+packages also must be compiled, so depending on the environment you need, this
+could take several hours to build. This vignette will explain how you can cut
+down on building times by building the environment once, and then caching it, so
+you don’t need to rebuild it on new machines.
+
+Most packages you will need, get pre-build using
+[Hydra](https://hydra.nixos.org/project/nixpkgs), and the resulting binaries are
+then made avaible on NixOS’s [public cache](https://cache.nixos.org/).
+Concretely, this means that you likely will never have to build packages from
+source: the Nix package manager first checks if what you need is in the public
+cache, and if yes, it downloads a binary from there instead of building it on
+your computer. This means that building an environment is just a matter of
+waiting for packages to download.
+
+However, if you use `rix::rix(r_ver = "bleeding", ...)`, you will generate an
+expression that when built will result in an environment with bleeding edge R
+packages. These packages will get installed from source, and so it could take a
+very long time to build. This is where we advise that you first: use our cache
+on top of the public cache (as explained in `vignette("z-bleeding_edge")`), and
+then, if our cache didn’t cover your needs, set up your own if you had to build
+many packages from source. This vignette explains how.
+
+## Using Github Actions to build the environment
+
+Before creating a cache to hold the binaries for your development environment,
+you first need to build the environment once. You can build the environment on
+your computer, or on Github Actions. The advantage of using Github Actions is
+that you can automate the process of building and pushing the binaries each time
+you change the definition of the environment.
+
+For example, imagine you have the following `generate_env.R` file for a project,
+which you version on Github:
+
+```
+library(rix)
+
+rix(r_ver = "bleeding_edge",
+    r_pkgs = c("dplyr", "ggplot2"),
+    system_pkgs = NULL,
+    git_pkgs = NULL,
+    tex_pkgs = NULL,
+    ide = "rstudio",
+    project_path = ".")
+```
+
+if you add a packages to it, or re-run it, you’ll end up with a new
+`default.nix` file, and so you will need to rebuild the environment. Again,
+depending on the packages you include, this could take quite some time to build.
+To generate a Github Actions workflow file that will build the environment on
+Github Actions, run `rix::ga_cachix(cache_name = "", path = "")` where
+`cache_name` is the name of the cache you made on Cachix, and `path` is the path
+to the `default.nix` file generated by `generate_env.R`. By default, the
+environment gets rebuilt every time you push changes to the `master/main` branch
+of your repository, but you can also re-build the environment periodically, by
+changing these lines:
+
+```
+on:
+  push:
+    branches: [ master, main ]
+```
+
+to this:
+
+```
+on:
+  push:
+    branches: [ master, main ]
+  schedule:
+    - cron: '30 0 * * *'
+```
+
+By using [cron syntax](https://en.wikipedia.org/wiki/Cron#Overview) you can
[…]
