### 2024-05-19 28f3f3f
removed test for frozen_edge as it'll fail as soon as there'll be another commit
