### 2024-05-20 f561235
finished binaries vignettes
