### 2024-05-21 f45ce03
added test for frozen edge case
