### 2024-05-21 a838451
- Added ga_cachix(), a function to create a GA workflow file that builds and pushes an environment to a Cachix cache
