+++ NAMESPACE
@@ -1,6 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
 export(available_r)
+export(ga_cachix)
 export(nix_build)
 export(rix)
 export(rix_init)
