### 2024-05-21 1289c19
- Added the possibility to create 'bleeding_edge' and 'frozen_edge' environments
