+++ R/rix.R
@@ -6,6 +6,8 @@
 #' @param r_ver Character, defaults to "latest". The required R version, for example "4.0.0".
 #'   You can check which R versions are available using `available_r`.
 #'   For reproducibility purposes, you can also provide a nixpkgs revision directly.
+#'   It is also possible to provide either "bleeding_edge" or "frozen_edge" if you
+#'   need an environment with bleeding edge packages. Read more in the "Details" below.
 #' @param r_pkgs Vector of characters. List the required R packages for your
 #'   analysis here.
 #' @param system_pkgs Vector of characters. List further software you wish to install that
@@ -69,6 +71,20 @@
 #'   preferred locale variables via
 #'   `options(rix.nix_locale_variables = list(LANG = "de_CH.UTF-8", <...>)`
 #'   and the aforementioned locale variable names.
+#'   It is possible to use "bleeding_edge" or "frozen_edge" as the value for
+#'   the `r_ver` argument. This will create an environment with the very
+#'   latest R packages. "bleeding_edge" means that every time you will build
+#'   the environment, the packages will get updated. This is especially useful
+#'   for environments that need to be constantly updated, for example when
+#'   developing a package. In contrast, "frozen_edge" will create an
+#'   environment that will remain stable at build time. So if you create
+#'   a `default.nix` file using "bleeding_edge", each time you build it
+#'   using `nix-build` that environment will be up-to-date. With "frozen_edge"
+#'   that environment will be up-to-date on the date that the `default.nix`
+#'   will be generated, and then each subsequent call to `nix-build` will
+#'   result in the same environment. We highly recommend you read the
+#'   vignette titled
+#'   "z - Advanced topic: Understanding the rPackages set release cycle and using bleeding edge packages".
 #' @export
 #' @examples
 #' \dontrun{
@@ -94,6 +110,15 @@ rix <- function(r_ver = "latest",
                 overwrite = FALSE,
                 print = FALSE,
                 shell_hook = NULL){
+
+  if(r_ver %in% c("bleeding_edge", "frozen_edge")){
+    warning(
+      "You chose 'bleeding_edge' or 'frozen_edge' as the value for
+`r_ver`. Please read the vignette
+https://b-rodrigues.github.io/rix/articles/z-advanced-topic-understanding-the-rpackages-set-release-cycle-and-using-bleeding-edge-packages.html
+before continuing."
+    )
+  }
 
   ide <- match.arg(ide, c("other", "code", "radian", "rstudio", "rserver"))
 
