### 2024-05-21 077dcd5
skip this test on windows as sed is not available unless running on wsl
