+++ tests/testthat/test-ga_cachix.R
@@ -1,5 +1,7 @@
 testthat::test_that("Snapshot test of ga_cachix()", {
 
+  skip_if(Sys.info()["sysname"] == "Windows")
+
   testthat::announce_snapshot_file("ga_cachix/cachix_dev_env.yaml")
 
   testthat::expect_snapshot_file(
