### 2024-05-22 db95bf0
added create_dev_env.R file to generate dev env to hack on rix
