+++ create_dev_env.R
@@ -1,0 +1,15 @@
+library(rix)
+
+latest_commit <- sys::as_text(sys::exec_internal("git", c("rev-parse", "master"))$stdout)
+
+
+rix(r_ver = "bleeding_edge",
+    r_pkgs = c("devtools", "diffviewer", "fledge", "codetools", "jsonlite",
+               "httr", "sys", "testthat", "knitr", "rmarkdown"),
+    system_pkgs = c("R", "glibcLocalesUtf8", "pandoc", "nix"),
+    tex_pkgs = "scheme-small",
+    ide = "radian",
+    project_path = ".",
+    overwrite = TRUE,
+    print = FALSE,
+    shell_hook = NULL)
