+++ create_dev_env.R
@@ -8,7 +8,7 @@ rix(r_ver = "bleeding_edge",
                "httr", "sys", "testthat", "knitr", "rmarkdown"),
     system_pkgs = c("R", "glibcLocalesUtf8", "pandoc", "nix"),
     tex_pkgs = "scheme-small",
-    ide = "radian",
+    ide = "other",
     project_path = ".",
     overwrite = TRUE,
     print = FALSE,
