### 2024-05-23 c88bfcb
Update CONTRIBUTING.md
fix typo
