### 2024-05-23 c165687
moved helper functions to separate file
