+++ R/rix.R
@@ -170,38 +170,6 @@ for more details.")
   # Get the rix version
   rix_version <- utils::packageVersion("rix")
 
-
-  # Now we need to generate all the different sets of packages
-  # to install. Let's start by the CRAN packages, current
-  # and archived. The function below builds the strings.
-  get_rPackages <- function(r_pkgs){
-
-    # in case users pass something like c("dplyr", "tidyr@1.0.0")
-    # r_pkgs will be "dplyr" only
-    # and "tidyr@1.0.0" needs to be handled by fetchzips
-    r_and_archive_pkgs <- detect_versions(r_pkgs)
-
-    # overwrite r_pkgs
-    r_pkgs <- r_and_archive_pkgs$cran_packages
-
-    # get archive_pkgs
-    archive_pkgs <- r_and_archive_pkgs$archive_packages
-
-    r_pkgs <- if(ide == "code"){
-                c(r_pkgs, "languageserver")
-              } else {
-                r_pkgs
-              }
-
-    rPackages <- paste(r_pkgs, collapse = ' ')
-
-    rPackages <- gsub('\\.', '_', rPackages)
-
-    list("rPackages" = rPackages,
-         "archive_pkgs" = archive_pkgs)
-
-  }
-
   # Get the two lists. One list is current CRAN packages
   # the other is archived CRAN packages.
   cran_pkgs <- get_rPackages(r_pkgs)
@@ -214,34 +182,6 @@ for more details.")
                   "rpkgs"
                 }
 
-  # generate_* function generate the actual Nix code
-  generate_rpkgs <- function(rPackages) {
-    if (flag_rpkgs == ""){
-      NULL
-    } else {
-      sprintf('rpkgs = builtins.attrValues {
-  inherit (pkgs.rPackages) %s;
-};
-',
-rPackages)
-    }
-  }
-
-  # Texlive packages
-  generate_tex_pkgs <- function(tex_pkgs) {
-    if (!is.null(tex_pkgs)) {
-
-      tex_pkgs <- unique(c("scheme-small", tex_pkgs))
-
-      tex_pkgs <- paste(tex_pkgs, collapse = ' ')
-
-    sprintf('tex = (pkgs.texlive.combine {
-  inherit (pkgs.texlive) %s;
-});
-',
-tex_pkgs)
-    }
-  }
 
   flag_tex_pkgs <- if(is.null(tex_pkgs)){
                      ""
@@ -249,92 +189,12 @@ tex_pkgs)
                      "tex"
                    }
 
-  # system packages
-  get_system_pkgs <- function(system_pkgs, r_pkgs){
-
-    system_pkgs <- c(system_pkgs, "R", "glibcLocales", "nix")
-
-    system_pkgs <- if(any(grepl("quarto", r_pkgs))){
-                unique(c(system_pkgs, "quarto"))
-              } else {
-                unique(system_pkgs)
-              }
-
-    paste(system_pkgs, collapse = ' ')
-  }
-
   flag_git_archive <- if(!is.null(cran_pkgs$archive_pkgs) | !is.null(git_pkgs)){
                         "git_archive_pkgs"
                       } else {
                         ""
                       }
 
-  generate_git_archived_packages <- function(git_pkgs, archive_pkgs){
-    if(flag_git_archive == ""){
-      NULL
-    } else {
-    sprintf('git_archive_pkgs = [%s];\n',
-            fetchpkgs(git_pkgs, archive_pkgs)
-            )
-    }
-  }
-
-
-  # `R` needs to be added. If we were using the rWrapper
-  # this wouldn't be needed, but we're not so we need
-  # to add it.
-  generate_system_pkgs <- function(system_pkgs, r_pkgs){
[…]
