+++ R/rix_helpers.R
@@ -1,0 +1,203 @@
+#' generate_header Internal function used to generate the header of the `default.nix` file.
+#' @param rix_version Character, current, version of the package
+#' @param nix_revision Character, nixpkgs revision to be used.
+#' @param r_ver_text Character, human-readable text giving the version.
+#' @param rix_call Character, call to rix().
+#' @noRd
+generate_header <- function(rix_version,
+                            nix_revision,
+                            nix_url,
+                            r_ver_text,
+                            rix_call){
+
+  if(identical(Sys.getenv("TESTTHAT"), "true")){
+     sprintf('
+let
+ pkgs = import (fetchTarball "%s") {};
+',
+nix_url)
+ } else {
+   sprintf('# This file was generated by the {rix} R package v%s on %s
+# with following call:
+%s
+# It uses nixpkgs\' revision %s for reproducibility purposes
+# which will install R %s
+# Report any issues to https://github.com/b-rodrigues/rix
+let
+ pkgs = import (fetchTarball "%s") {};
+',
+rix_version,
+Sys.Date(),
+generate_rix_call(rix_call, nix_url),
+nix_revision,
+r_ver_text,
+nix_url
+)
+  }
+}
+
+# Now we need to generate all the different sets of packages
+# to install. Let's start by the CRAN packages, current
+# and archived. The function below builds the strings.
+#' generate_rPackages Internal function used to generate the header of the `default.nix` file.
+#' @param r_pkgs Character, list of R packages to install.
+#' @noRd
+get_rPackages <- function(r_pkgs){
+
+  # in case users pass something like c("dplyr", "tidyr@1.0.0")
+  # r_pkgs will be "dplyr" only
+  # and "tidyr@1.0.0" needs to be handled by fetchzips
+  r_and_archive_pkgs <- detect_versions(r_pkgs)
+
+  # overwrite r_pkgs
+  r_pkgs <- r_and_archive_pkgs$cran_packages
+
+  # get archive_pkgs
+  archive_pkgs <- r_and_archive_pkgs$archive_packages
+
+  r_pkgs <- if(ide == "code"){
+              c(r_pkgs, "languageserver")
+            } else {
+              r_pkgs
+            }
+
+  rPackages <- paste(r_pkgs, collapse = ' ')
+
+  rPackages <- gsub('\\.', '_', rPackages)
+
+  list("rPackages" = rPackages,
+       "archive_pkgs" = archive_pkgs)
+
+}
+
+# generate_* function generate the actual Nix code
+generate_rpkgs <- function(rPackages) {
+  if (flag_rpkgs == ""){
+    NULL
+  } else {
+    sprintf('rpkgs = builtins.attrValues {
+  inherit (pkgs.rPackages) %s;
+};
+',
+rPackages)
+  }
+}
+
+# Texlive packages
+generate_tex_pkgs <- function(tex_pkgs) {
+  if (!is.null(tex_pkgs)) {
+
+    tex_pkgs <- unique(c("scheme-small", tex_pkgs))
+
+    tex_pkgs <- paste(tex_pkgs, collapse = ' ')
+
+    sprintf('tex = (pkgs.texlive.combine {
+  inherit (pkgs.texlive) %s;
+});
+',
+tex_pkgs)
+  }
+}
+
+# system packages
+get_system_pkgs <- function(system_pkgs, r_pkgs){
+
+  system_pkgs <- c(system_pkgs, "R", "glibcLocales", "nix")
+
+  system_pkgs <- if(any(grepl("quarto", r_pkgs))){
+                   unique(c(system_pkgs, "quarto"))
+                 } else {
[…]
