### 2024-05-23 5e301e4
Update z-contributing_to_nixpkgs.Rmd
