### 2024-05-23 d547f50
added generate_rix_call back
