+++ R/rix.R
@@ -210,8 +210,8 @@ for more details.")
                     nix_url,
                     r_ver_text,
                     rix_call),
-    generate_rpkgs(cran_pkgs$rPackages),
-    generate_git_archived_packages(git_pkgs, cran_pkgs$archive_pkgs),
+    generate_rpkgs(cran_pkgs$rPackages, flag_rpkgs),
+    generate_git_archived_packages(git_pkgs, cran_pkgs$archive_pkgs, flag_git_archive),
     generate_tex_pkgs(tex_pkgs),
     generate_system_pkgs(system_pkgs, r_pkgs),
     generate_wrapped_pkgs(ide, attrib, flag_git_archive, flag_rpkgs),
