### 2024-05-23 83e9a49
Update rix_helpers.R
