+++ R/internal_functions.R
@@ -5,19 +5,19 @@
 #' @param rix_call Character, call to rix().
 #' @noRd
 generate_header <- function(rix_version,
-                            nix_revision,
-                            nix_url,
-                            r_ver_text,
+                          nix_revision,
+                          nix_url,
+                          r_ver_text,
                             rix_call){
 
   if(identical(Sys.getenv("TESTTHAT"), "true")){
-     sprintf('
+      sprintf('
 let
  pkgs = import (fetchTarball "%s") {};
 ',
-nix_url)
- } else {
-   sprintf('# This file was generated by the {rix} R package v%s on %s
+  nix_url)
+  } else {
+    sprintf('# This file was generated by the {rix} R package v%s on %s
 # with following call:
 %s
 # It uses nixpkgs\' revision %s for reproducibility purposes
@@ -26,13 +26,13 @@ nix_url)
 let
  pkgs = import (fetchTarball "%s") {};
 ',
-rix_version,
-Sys.Date(),
-generate_rix_call(rix_call, nix_url),
-nix_revision,
-r_ver_text,
-nix_url
-)
+  rix_version,
+  Sys.Date(),
+  generate_rix_call(rix_call, nix_url),
+  nix_revision,
+  r_ver_text,
+  nix_url
+  )
   }
 }
 
@@ -50,13 +50,10 @@ generate_rix_call <- function(rix_call, nix_url){
 }
 
 
-# Now we need to generate all the different sets of packages
-# to install. Let's start by the CRAN packages, current
-# and archived. The function below builds the strings.
-#' generate_rPackages Internal function used to generate the header of the `default.nix` file.
+#' generate_rPackages Internal function that formats the R package names correctly for Nix.
 #' @param r_pkgs Character, list of R packages to install.
 #' @noRd
-get_rPackages <- function(r_pkgs, ide){
+get_rpkgs <- function(r_pkgs, ide){
 
   # in case users pass something like c("dplyr", "tidyr@1.0.0")
   # r_pkgs will be "dplyr" only
@@ -84,20 +81,27 @@ get_rPackages <- function(r_pkgs, ide){
 
 }
 
-# generate_* function generate the actual Nix code
+#' generate_rpkgs Internal function that generates the string containing the correct Nix expression to get R packages.
+#' @param r_pkgs Character, list of R packages to install.
+#' @param flag_rpkgs Character, are there any R packages at all?
+#' @noRd
 generate_rpkgs <- function(rPackages, flag_rpkgs) {
   if (flag_rpkgs == ""){
     NULL
   } else {
-    sprintf('rpkgs = builtins.attrValues {
+    sprintf('
+ rpkgs = builtins.attrValues {
   inherit (pkgs.rPackages) %s;
-};
+ };
 ',
 rPackages)
   }
+
 }
 
-# Texlive packages
+#' generate_tex_pkgs Internal function that generates the string containing the correct Nix expression to get LaTeX packages.
+#' @param tex_pkgs Character, list of LaTeX packages to install.
+#' @noRd
 generate_tex_pkgs <- function(tex_pkgs) {
   if (!is.null(tex_pkgs)) {
 
@@ -105,19 +109,25 @@ generate_tex_pkgs <- function(tex_pkgs) {
 
     tex_pkgs <- paste(tex_pkgs, collapse = ' ')
 
-    sprintf('tex = (pkgs.texlive.combine {
+    sprintf('
+ tex = (pkgs.texlive.combine {
   inherit (pkgs.texlive) %s;
-});
+ });
 ',
 tex_pkgs)
   }
 }
 
-# system packages
+#' generate_system_pkgs Internal function that formats the system package names correctly for Nix.
[…]
