+++ R/internal_functions.R
@@ -1,48 +1,88 @@
 #' generate_header Internal function used to generate the header of the `default.nix` file.
-#' @param rix_version Character, current, version of the package
-#' @param nix_revision Character, nixpkgs revision to be used.
-#' @param r_ver_text Character, human-readable text giving the version.
+#' @param nix_repo Character. nixpkgs reop to use (upstream or rstats-on-nix fork) with latest commit hash.
+#' @param r_version Character. R version to look for, for example, "4.2.0". If a nixpkgs revision is provided instead, this gets returned.
 #' @param rix_call Character, call to rix().
 #' @noRd
-generate_header <- function(rix_version,
-                          nix_revision,
-                          nix_url,
-                          r_ver_text,
+generate_header <- function(nix_repo,
+                            r_version,
                             rix_call){
 
   if(identical(Sys.getenv("TESTTHAT"), "true")){
-      sprintf('
+    sprintf('
 let
  pkgs = import (fetchTarball "%s") {};
 ',
-  nix_url)
-  } else {
-    sprintf('# This file was generated by the {rix} R package v%s on %s
+nix_repo$url)
+    } else {
+
+      # Generate the correct text for the header depending on wether
+      # an R version or a Nix revision is supplied to `r_ver`
+      if(nchar(r_version) > 20){
+        r_ver_text <- paste0("as it was as of nixpkgs revision: ", r_version)
+      } else {
+        r_ver_text <- paste0("version ", r_version)
+      }
+
+      # Get the rix version
+      rix_version <- utils::packageVersion("rix")
+
+      nix_url <- nix_repo$url
+
+      nix_revision <- nix_repo$latest_commit
+
+      if(grepl("NixOS", nix_url)){
+        sprintf('# This file was generated by the {rix} R package v%s on %s
 # with following call:
 %s
 # It uses nixpkgs\' revision %s for reproducibility purposes
-# which will install R %s
+# which will install R %s.
 # Report any issues to https://github.com/b-rodrigues/rix
 let
  pkgs = import (fetchTarball "%s") {};
 ',
   rix_version,
   Sys.Date(),
-  generate_rix_call(rix_call, nix_url),
+  generate_rix_call(rix_call, nix_repo),
   nix_revision,
   r_ver_text,
   nix_url
   )
-  }
+
+      } else { # if we're using rstats-on-nix
+        sprintf('# This file was generated by the {rix} R package v%s on %s
+# with following call:
+%s
+# It uses the `rstats-on-nix` fork of `nixpkgs` which provides bleeding
+# edge packages and R.
+# Only use bleeding edge packages if absolutely needed!
+# Read more on https://b-rodrigues.github.io/rix/articles/z-bleeding_edge.html
+# Report any issues to https://github.com/b-rodrigues/rix
+let
+ pkgs = import (fetchTarball "%s") {};
+',
+  rix_version,
+  Sys.Date(),
+  generate_rix_call(rix_call, nix_repo),
+  nix_url
+  )
+
+    }
+    }
 }
 
 #' generate_rix_call Internal function used to generate the call to `rix()` as shown in `default.nix`
 #' @param rix_call Character, call to rix().
-#' @param nix_url Character, nixpkgs revision to be used.
+#' @param nix_repo Character. nixpkgs reop to use (upstream or rstats-on-nix fork) with latest commit hash.
 #' @noRd
-generate_rix_call <- function(rix_call, nix_url){
+generate_rix_call <- function(rix_call, nix_repo){
 
[…]
