+++ NEWS.md
@@ -1,5 +1,9 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# rix 0.7.0 (2024-05-24)
+
+- Better messages of comments on top of the generated `default.nix` files.
+
 # rix 0.7.0 (2024-05-21)
 
 - Added the possibility to create "bleeding_edge" and "frozen_edge" environments
