### 2024-05-24 be00ba6
vignette contributing to nixpkgs
