+++ vignettes/z-contributing_to_nixpkgs.Rmd
@@ -17,3 +17,428 @@ knitr::opts_chunk$set(
 ```{r, include=FALSE}
 library(rix)
 ```
+
+## Introduction
+
+If you want to help us make the R ecosystem better on Nix by fixing packages,
+then read on!
+
+This Vignette will help you help us! It lists common ways that packages can
+break and teaches you how to fix these R packages for inclusion to `nixpkgs`.
+Every package available on CRAN and Bioconductor gets built on Hydra and made
+available through the `rPackages` package set. So there's no need to manually
+package them. However, some packages don't build successfully, and require
+manual fixing. Most of them are very quick, one-line fixes, but others require a
+bit more work. The goal of this Vignette is to make you quickly familiar with
+the main reasons a package may be broken and explain to you how to fix it, and
+why certain packages get fixed in certain ways.
+
+This vignette assumes that you are comfortable with Git and with the concepts
+of branches and PRs (pull requests) and with compiling software on Linux.
+
+## Setting up
+
+We first need to set up some thing. Fork the [`nixpkgs` repository](https://github.com/NixOS/nixpkgs)
+and clone it to your computer. Then, add the original repository as a remote:
+
+```
+git checkout master
+
+# Add upstream as a remote
+git remote add upstream git@github.com:NixOS/nixpkgs.git
+```
+
+This way, each time you want to fix a package, you can fetch the latest updates
+from upstream and merge them to your local copy:
+
+```
+# Fetch latest updates
+git fetch upstream
+
+# Merge latest upstream/master to your master
+git merge upstream/master
+```
+
+Make sure to merge the latest commits from upstream regularly, because `nixpkgs`
+gets updated very frequently each day. We can now look for a package to fix.
+
+If for some reason you cannot update your local copy, make sure you're on `master`
+and reset hard:
+
+```
+git reset --hard upstream/master
+```
+
+Always make sure to create a new branch to fix a package, don't work from `master`,
+you'll just be asking for trouble!
+
+You might also want to join our Matrix chat room to coordinate efforts. You can
+join it [here](https://matrix.to/#/#r:nixos.org).
+
+## Where to find packages to fix
+
+The first step to help fix a package is to find a package to fix: you should
+visit the latest `rPackages` evaluation over
+[here](https://hydra.nixos.org/jobset/nixpkgs/r-updates). Click on the "Still
+failing jobs" tab to see which packages' builds didn't succeed and click on a
+package. You should see something like this:
+
+![AIUQ build steps](images/AIUQ_failing.png)
+
+From there, you can see that `{AIUQ}`'s build failed because of another package,
+`{SuperGauss}`, so fixing `{SuperGauss}` will likely fix this one as well.
+
+You can look for `{SuperGauss}` in the "Still failing jobs" tab, and see why
+`{SuperGauss}` failed, or you could check out a little dashboard I built that
+you can find
+[here](https://raw.githack.com/b-rodrigues/nixpkgs-r-updates-fails/targets-runs/output/r-updates-fails.html).
+This dashboard shows essentially the same information you find on the "Still
+failing jobs" tab from before, but with several added niceties. First of all,
+there's a column called `fails_because_of` that shows the name of the package
[…]
