### 2024-05-24 9038738
updated snapshot test for cachix
