+++ tests/testthat/_snaps/ga_cachix/cachix_dev_env.yaml
@@ -15,6 +15,10 @@ jobs:
         os: [ubuntu-latest, macos-latest]
     steps:
     - uses: actions/checkout@v4
+
+    - name: Nix cache
+      uses: DeterminateSystems/magic-nix-cache-action@main
+
     - uses: cachix/install-nix-action@v25
       with:
         nix_path: nixpkgs=channel:nixos-unstable
