### 2024-05-24 57bb468
Update NEWS.md
