### 2024-05-30 d0408a3
Update z-contributing_to_nixpkgs.Rmd
