### 2024-05-30 3cb58f1
Update z-contributing_to_nixpkgs.Rmd
