### 2024-06-08 2f1dfda
implement `message_type = "quiet"`
