+++ R/with_nix.R
@@ -88,8 +88,8 @@
 #' the Nix software environment in which you want to run `expr`. See
 #'  `project_path` argument as an alternative way to specify the environment.
 #' @param message_type String how detailed output is. Currently, there is 
-#' either `"simple"` (default) or `"verbose"`, which shows the script that runs
-#' via `nix-shell`.
+#' either `"simple"` (default), `"quiet` or `"verbose"`, which shows the script
+#' that runs via `nix-shell`.
 #' @importFrom codetools findGlobals checkUsage
 #' @export
 #' @return 
@@ -148,7 +148,7 @@ with_nix <- function(expr,
                      exec_mode = c("blocking", "non-blocking"),
                      project_path = ".",
                      nix_file = NULL,
-                     message_type = c("simple", "verbose")) {
+                     message_type = c("simple", "quiet", "verbose")) {
   if (is.null(nix_file)) {
     nix_file <- file.path(project_path, "default.nix")
   }
@@ -226,6 +226,8 @@ with_nix <- function(expr,
   exec_mode <- match.arg(exec_mode)
   message_type <- match.arg(message_type)
   
+  is_quiet <- message_type == "quiet"
+  
   if (program == "R") {
   
     # get the function arguments as a pairlist;
@@ -236,8 +238,10 @@ with_nix <- function(expr,
     # and bound to `"."` (project root)
     args <- as.list(formals(expr))
     
-    cat("\n### Prepare to exchange arguments and globals for `expr`",
-      "between the host and Nix R sessions ###\n")
+    if (message_type == "verbose") {
+      cat("\n### Prepare to exchange arguments and globals for `expr`",
+          "between the host and Nix R sessions ###\n")
+    }
     
     # 1) save all function args onto a temporary folder each with
     # `<tag.Rds>` and `value` as serialized objects from RAM ---------------------
@@ -257,10 +261,15 @@ with_nix <- function(expr,
     # https://github.com/cran/codetools/blob/master/R/codetools.R
     # http://adv-r.had.co.nz/Expressions.html#ast-funs
     
-    # code inspection: generates messages with potential problems
-    check_expr(expr)
-    
-    globals_expr <- recurse_find_check_globals(expr, args_vec)
+    # do code inspection checks to report messages with potential code problems,
+    # and find global variables of `expr` recursively
+    # using {codetools} wrapper
+    if (isFALSE(is_quiet)) {
+      globals_expr <- recurse_find_check_globals(expr, args_vec)
+    } else {
+      globals_expr <- recurse_find_check_globals(expr, args_vec,
+        message_type = "quiet")
+    }
     
     # wrapper around `serialize_lobjs()`
     globals <- serialize_globals(globals_expr, temp_dir)
@@ -276,7 +285,8 @@ with_nix <- function(expr,
     rnix_file <- file.path(temp_dir, "with_nix_r.R")
     
     rnix_quoted <- quote_rnix(
-      expr, program, message_type, args_vec, globals, pkgs, temp_dir, rnix_file
+      expr, program, message_type = message_type,
+      args_vec, globals, pkgs, temp_dir, rnix_file
     )
     rnix_deparsed <- deparse_chr1(expr = rnix_quoted, collapse = "\n")
     
@@ -287,10 +297,11 @@ with_nix <- function(expr,
     writeLines(text = rnix_deparsed, file(rnix_file))
     
     # 3) run expression in nix session, based on temporary script
-    cat(paste0("==> Running deparsed expression via `nix-shell`", " in ",
[…]
