### 2024-06-08 9b816b8
make `with_nix` fully quiet if requested
