### 2024-06-08 925f815
`rix_init()` implement "quiet" and improve "simple" message_type
