+++ R/rix_init.R
@@ -75,16 +75,21 @@
 #' \dontrun{
 #' # create an isolated, runtime-pure R setup via Nix
 #' project_path <- "./sub_shell"
+#' if (!dir.exist(project_path)) dir.create(project_path)
 #' rix_init(
 #'   project_path = project_path,
-#'   rprofile_action = "create_missing"
+#'   rprofile_action = "create_missing",
+#'   message_type = c("simple")
 #' )
 #' }
 rix_init <- function(project_path = ".",
                  rprofile_action = c("create_missing", "create_backup",
                    "overwrite", "append"),
-                 message_type = c("simple", "verbose")) {
-  message_type <- match.arg(message_type, choices = c("simple", "verbose"))
+                 message_type = c("simple", "quiet", "verbose")) {
+  message_type <- match.arg(message_type,
+    choices = c("simple", "quiet", "verbose"))
+  is_quiet <- message_type == "quiet"
+  
   rprofile_action <- match.arg(rprofile_action,
     choices = c("create_missing", "create_backup", "overwrite", "append"))
   stopifnot(
@@ -92,16 +97,22 @@ rix_init <- function(project_path = ".",
       is.character(project_path) && length(project_path) == 1L
   )
   
-  cat("\n### Bootstrapping isolated, project-specific, and runtime-pure",
-    "R setup via Nix ###\n\n")
+  if (isFALSE(is_quiet)) {
+    cat("\n### Bootstrapping isolated, project-specific, and runtime-pure",
+        "R setup via Nix ###\n\n")
+  }
   if (isFALSE(dir.exists(project_path))) {
     dir.create(path = project_path, recursive = TRUE)
     project_path <- normalizePath(path = project_path)
-    cat("==> Created isolated nix-R project folder:\n", project_path, "\n")
+    if (isFALSE(is_quiet)) {
+      cat("==> Created isolated nix-R project folder:\n", project_path, "\n")
+    }
   } else {
     project_path <- normalizePath(path = project_path)
-    cat("==> Existing isolated nix-R project folder:\n", project_path,
-      "\n")
+    if (isFALSE(is_quiet)) {
+      cat("==> Existing isolated nix-R project folder:\n", project_path,
+        "\n")
+    }
   }
   
   # create project-local `.Rprofile` with pure settings
@@ -118,8 +129,8 @@ rix_init <- function(project_path = ".",
     )
   }
   
-  is_nixr <- is_nix_rsession()
-  is_rstudio <- is_rstudio_session()
+  is_nixr <- is_nix_rsession(message_type)
+  is_rstudio <- is_rstudio_session(message_type)
   
   rprofile_exists <- file.exists(rprofile_file)
   timestamp <- format(Sys.time(), "%Y-%m-%dT%H:%M:%S%z")
@@ -128,28 +139,35 @@ rix_init <- function(project_path = ".",
   switch(rprofile_action,
     create_missing = {
       if (isTRUE(rprofile_exists)) {
-        cat(
-          "\n* Keep existing `.Rprofile`. in `project_path`:\n",
-          paste0(project_path, "/"), "\n"
-        )
+        if (isFALSE(is_quiet)) {
+          cat(
+            "\n* Keep existing `.Rprofile`. in `project_path`:\n",
+            paste0(project_path, "/"), "\n"
+          )
+        }
       } else {
         write_rprofile(rprofile_text, rprofile_file)
-        message_rprofile(action_string = "Added", project_path = project_path)
+        if (isFALSE(is_quiet)) {
+          message_rprofile(action_string = "Added", project_path = project_path)
+        }
       }
       set_message_session_PATH(message_type = message_type)
     },
     create_backup = {
[…]
