### 2024-06-08 0bf1b8e
update docs with new `"quiet"` messaging option
