### 2024-06-09 a207a08
separate boolean answer and message for `is_nix_r_session()`; adjust tests for rix init .Rprofile
