+++ tests/testthat/test-nix_build.R
@@ -1,6 +1,6 @@
 testthat::test_that("Testing that `nix_build()` builds derivation", {
 
-  if (isFALSE(is_nix_rsession())) {
+  if (isFALSE(is_nix_r_session())) {
     # needed for the GitHub test runners with system's R
     set_nix_path()
   }
