### 2024-06-09 76cfaf2
refactor, simplify and separate internal helpers
