+++ R/with_nix_helpers.R
@@ -1,0 +1,539 @@
+#' @noRd
+stop_no_nix_shell <- function() {
+  stop(
+    paste0("`nix-shell` is needed but is not available in your current ",
+      "shell environment.\n",
+      "* If you are in an R session of your host operating system, you
+      either\n1a) need to install Nix first, or if you have already done so ",
+      "\n",
+      "To install, we suggest you follow https://zero-to-nix.com/start/install .",
+      "\n",
+      "1b) make sure that the location of the nix store is in the `PATH`
+      variable of this R session (mostly necessary in RStudio).\n",
+      "* If you ran `with_nix()` from R launched in a `nix-shell`, you need
+      to make sure that `pkgs.nix` is in the `buildInput` for ",
+      "`pkgs.mkShell`.\nIf you used `rix::rix()` to generate your main nix
+      configuration of this session, just regenerate it with the additonal
+      argument `system_pkgs = 'nix'."),
+      call. = FALSE
+  )
+}
+
+
+#' serialize language objects
+#' @noRd
+serialize_lobjs <- function(lobjs, temp_dir) {
+  invisible({
+    for (i in seq_along(lobjs)) {
+      if (!any(nzchar(deparse(lobjs[[i]])))) {
+        # for unnamed arguments like `expr = function(x) print(x)`
+        # x would be an empty symbol, see also ; i.e. arguments without 
+        # default expressions; i.e. tagged arguments with no value
+        # https://stackoverflow.com/questions/3892580/create-missing-objects-aka-empty-symbols-empty-objects-needed-for-f
+        lobjs[[i]] <- as.symbol(names(lobjs)[i])
+      }
+      saveRDS(
+        object = lobjs[[i]],
+        file = file.path(temp_dir, paste0(names(lobjs)[i], ".Rds"))
+      )
+    }
+  })
+}
+
+#' serialize arguments
+#' @noRd
+serialize_args <- function(args, temp_dir) {
+  invisible({
+    for (i in seq_along(args)) {
+      if (!nzchar(deparse(args[[i]]))) {
+        # for unnamed arguments like `expr = function(x) print(x)`
+        # x would be an empty symbol, see also ; i.e. arguments without 
+        # default expressions; i.e., tagged arguments with no value
+        # https://stackoverflow.com/questions/3892580/create-missing-objects-aka-empty-symbols-empty-objects-needed-for-f
+        args[[i]] <- as.symbol(names(args)[i])
+      }
+      args[[i]] <- get(as.character(args[[i]]))
+      saveRDS(
+        object = args[[i]],
+        file = file.path(temp_dir, paste0(names(args)[i], ".Rds"))
+      )
+    }
+  })
+}
+
+#' @noRd
+# to determine which extra packages to load in Nix R prior evaluating `expr`
+get_expr_extra_pkgs <- function(globals_expr) {
+  envs_check <- lapply(globals_expr, where)
+  names_envs_check <- vapply(envs_check, environmentName, character(1L))
+  
+  default_pkgnames <- paste0("package:", getOption("defaultPackages"))
+  pkgenvs_attached <- setdiff(
+    grep("^package:", names_envs_check, value = TRUE), 
+    c(default_pkgnames, "base")
+  )
+  if (!length(pkgenvs_attached) == 0L) {
+    pkgs_to_attach <- gsub("^package:", "", pkgenvs_attached)
+    return(pkgs_to_attach)
+  } else {
+    return(NULL)
+  }
+}
+
+
+#' @noRd
[…]
