+++ R/with_nix.R
@@ -161,6 +161,13 @@ with_nix <- function(expr,
     "`expr` needs to be a call or function for `program = R`, and character of length 1 for `program = shell`" =
       is.function(expr) || is.call(expr) || (is.character(expr) && length(expr) == 1L)
   )
+
+  program <- match.arg(program, choices = c("R", "shell"))
+  exec_mode <- match.arg(exec_mode, choices = c("blocking", "non-blocking"))
+  message_type <- match.arg(message_type,
+    choices = c("simple", "quiet", "verbose"))
+  
+  is_quiet <- message_type == "quiet"
   
   # ad-hoc solution for RStudio's limitation that R sessions cannot yet inherit
   # proper `PATH` from custom `.Rprofile` on macOS (2023-01-17)
@@ -175,7 +182,7 @@ with_nix <- function(expr,
     set_nix_path()
   }
   
-  if (isTRUE(nzchar(Sys.getenv("NIX_STORE")))) {
+  if (isTRUE(is_nix_r)) {
     # for Nix R sessions, guarantee that the system's user library 
     # (R_LIBS_USER) is not in the search path for packages => run-time purity
     current_libpaths <- .libPaths()
@@ -209,31 +216,8 @@ with_nix <- function(expr,
   has_nix_shell <- nix_shell_available() # TRUE if yes, FALSE if no
   
   if (isFALSE(has_nix_shell)) {
-    stop(
-      paste0("`nix-shell` is needed but is not available in your current ",
-        "shell environment.\n",
-        "* If you are in an R session of your host operating system, you
-        either\n1a) need to install Nix first, or if you have already done so ",
-        "\n",
-        "To install, we suggest you follow https://zero-to-nix.com/start/install .",
-        "\n",
-        "1b) make sure that the location of the nix store is in the `PATH`
-        variable of this R session (mostly necessary in RStudio).\n",
-        "* If you ran `with_nix()` from R launched in a `nix-shell`, you need
-        to make sure that `pkgs.nix` is in the `buildInput` for ",
-        "`pkgs.mkShell`.\nIf you used `rix::rix()` to generate your main nix
-        configuration of this session, just regenerate it with the additonal
-        argument `system_pkgs = 'nix'."),
-      call. = FALSE
-    )
+    stop_no_nix_shell()
   }
-  
-  program <- match.arg(program, choices = c("R", "shell"))
-  exec_mode <- match.arg(exec_mode, choices = c("blocking", "non-blocking"))
-  message_type <- match.arg(message_type,
-    choices = c("simple", "quiet", "verbose"))
-  
-  is_quiet <- message_type == "quiet"
   
   if (program == "R") {
   
@@ -313,7 +297,7 @@ with_nix <- function(expr,
     # command to run deparsed R expression via nix-shell
     cmd_rnix_deparsed <- c(
       file.path(project_path, "default.nix"),
-      "--pure", # required for to have nix glibc
+      "--pure", # required so that nix glibc is used
       "--run",
       sprintf(
         "Rscript --no-site-file --no-environ --no-restore '%s'",
@@ -372,7 +356,7 @@ with_nix <- function(expr,
     cat("\n### Finished code evaluation in `nix-shell` ###\n")
   }
   
-  if (isTRUE(nzchar(Sys.getenv("NIX_STORE")))) {
+  if (isTRUE(is_nix_r)) {
     # set back library paths to state before calling `with_nix()`
      .libPaths(current_libpaths)
   } else {
@@ -402,521 +386,4 @@ with_nix <- function(expr,
[…]
