### 2024-06-09 d64b507
fix: `is_nix` => `is_nix_r`; implement polling with message types
