### 2024-06-09 427a711
simplified code inspection message only for "verbose"
