### 2024-06-09 560de33
consistent message styles for `nix_init()`, `nix_build()`, and `with_nix()`
