+++ R/rix_init.R
@@ -275,10 +275,10 @@ message_r_session_nix_rstudio <- function(is_nix_r,
   
   if (isTRUE(is_nix_r)) {
     nix_r_msg <-
-      "\n==> Source R session running inside Nix environment (nixpkgs)"
+      "\n* current R session running inside Nix environment (nixpkgs)"
   } else {
     nix_r_msg <-
-      "\n==> Source R session running outside Nix environment (nixpkgs)"
+      "\n* current R session running outside Nix environment (nixpkgs)"
   }
   
   if (isTRUE(is_rstudio)) {
