### 2024-06-09 c56d2d3
"verbose" messages: better report temp dir for serializating
