### 2024-06-09 1c0c947
hide serializatioon path and objects for "simple" messaging.
