### 2024-06-09 965e9da
also print file of nix expression used unless quiet
