### 2024-06-09 36a91ea
message correct `nix-build` on dir like happening
