### 2024-06-09 f4e881d
remove unused code line
