### 2024-06-09 0d38e50
improve docs for `rix_init()`
