### 2024-06-09 668be96
document internal helpers for `rix_init()`
