+++ R/rix_init.R
@@ -209,6 +209,12 @@ rix_init <- function(project_path = ".",
 }
 
 #' @noRd
+#' Get character vector of length two with comment and code write `.Rprofile`
+#' to evaluate R expressions in a pure R library runtime and also RStudio IDE
+#' on macOS
+#' 
+#' @param rprofile_deparsed deparsed string with containing `.Rprofile` code.
+#' @return Character vector of length 2.
 get_rprofile_text <- function(rprofile_deparsed) {
   c(
 "### File generated by `rix::rix_init()` ###
@@ -225,6 +231,9 @@ get_rprofile_text <- function(rprofile_deparsed) {
 }
 
 #' @noRd
+#' Print message for `.Rprofile` addition
+#' @param action_string string
+#' @param project_path string with project path
 message_rprofile <- function(action_string = "Added",
                              project_path = ".") {
   msg <- paste0(
@@ -239,6 +248,14 @@ message_rprofile <- function(action_string = "Added",
 }
 
 #' @noRd
+#' Get current `PATH` entries, report and modify to include default Nix profile
+#' path
+#' 
+#' Print `PATH` environent variable, and modfiy it as as side effect so that `` 
+#' `"/nix/var/nix/profiles/default/bin"` is included. Confirm message with
+#' by printing modified PATH.
+#' @return Character vector that lists `PATH` entries after modification, which
+#' are separated by `":"`.
 set_message_session_PATH <- function(message_type =
                                        c("simple", "quiet", "verbose")) {
   message_type <- match.arg(message_type,
@@ -259,6 +276,13 @@ set_message_session_PATH <- function(message_type =
 }
 
 #' @noRd
+#' Report whether the current R session is running in Nix and RStudio, or not.
+#' @param is_nix_r logical scalar. `TRUE` means in a Nix R environment
+#' @param is_rstudio `TRUE` means source R session is inside RStudio
+#' @param message_type character vector of lenght one. Either `"simple"`
+#' (default), `"quiet"`, or `"verbose"`. Currently, `"simple"` and `"verbose"``
+#' create identical messages, while `"quiet"` omits diagnostics messages
+#' @return NULL
 message_r_session_nix_rstudio <- function(is_nix_r, 
                                           is_rstudio,
                                           message_type =
@@ -298,18 +322,36 @@ message_r_session_nix_rstudio <- function(is_nix_r,
 
 
 #' @noRd
+#' Is the current R session running in a Nix software environment or not?
+#' 
+#' Query `NIX_STORE` environmental variable in current R session. Only nonzero
+#' if inside a Nix R.
+#' session.
+#' @return Logical vector of length one.
 is_nix_r_session <- function() {
   is_nix_r <- nzchar(Sys.getenv("NIX_STORE"))
   return(is_nix_r)
 }
 
 #' @noRd
+#' Has the current R session been launched from RStudio or not?
+#' 
+#' Query `RSTUDIO` environmental variable in current R session. Value is `"1"`
+#' if inside RStudio R session.
 is_rstudio_session <- function(message_type = c("simple", "quiet", "verbose")) {
   is_rstudio <- Sys.getenv("RSTUDIO") == "1"
   return(is_rstudio)
 }
 
 #' @noRd
+#' If not yet present, add the Nix default path of the system-wide profile to 
+#' `PATH` environment variable inside R session.
+#' 
+#' The default profile for the system is typically located at 
+#' `/nix/var/nix/profiles/default`.
+#' @details creates the side effect of adding the .
+#' `"/nix/var/nix/profiles/default/bin"`, when it is not yet part of `PATH`
[…]
