### 2024-06-09 41596fd
internal code organization: `nix_build()` helpers in separate files
