### 2024-06-09 8cb92fc
document `nix_build()` helper functions
