+++ R/nix_build_helpers.R
@@ -1,3 +1,20 @@
+#' Avoid impure R library paths in the Nix runtime caused by default
+#' `.libPaths()` mechanism in Nix.
+#' 
+#' Remove the global library tree of the default user library location
+#' (`R_LIBS_USER`) in Nix, which is the same as the used by the host operating
+#' system. This entry is at the first positio of `.libPaths()`. It addresses an
+#' issue that arises when combining R environments managed by the host operating
+#  system with those managed by Nix on the same computer. Due to the current
+#' packaging approach of R in Nix, it is challenging to maintain runtime-pure R
+#' library paths. `nix-shell --pure` also does not resolve the issue. By
+#' default, R packages are loaded from user-specific library directories of the
+#' host operating system, which can lead wrong behavior on Unix systems and
+#' segmentation faults on MacOS (Darwin).  In this context, "runtime-pure"
+#' refers to ensuring  that R only uses packages from `.libPaths()` in the Nix
+#' store, avoiding any unintended loading of R packages installed outside Nix.
+#'  @return Invisibly returns previous (if no `R_LIBS_USER` not in 
+#' `.libPaths()`) or updated `.libPaths()` library paths as character vector.
 #' @noRd
 remove_r_libs_user <- function() {
   current_paths <- .libPaths() 
@@ -18,6 +35,24 @@ remove_r_libs_user <- function() {
   })
 }
 
+#' Set `LD_LIBRARY_PATH` to `""`.
+
+#' We currently use this helper when not in a Nix R session in both `nix_build()`
+#' and `with_nix()`.
+#' On some systems, like Ubuntu 22.04, we found that a preset 
+#' `LD_LIBRARY_PATH` environment variable in the system's R session
+#' (R installed via apt) is responsible for causing  a segmentation fault
+#' for both `nix-build` and `nix-shell` when invoked via 
+#' `sys::exec_internal()`, `base::system()` or `base::system2()` from R.
+#' This seems due to incompatible linked libraries or permission issue that
+#' conflict when mixing Nix packages and libraries from the system.
+#' Therefore, we set it to `""` and set  back the default (old)
+#' `LD_LIBRARY_PATH` when `with_nix()` exits. For newer RStudio versions,
+#'  LD_LIBRARY_PATH is not `""` anymore
+#' https://github.com/rstudio/rstudio/issues/12585 => seems to work since
+#' RStudio 2023.12.0-daily+346 ; however, since we want to go back in time, too,
+#' we'll leave that tweak to ensure properly working behavior.
+#' @return 
 #' @noRd
 fix_ld_library_path <- function() {
   old_ld_library_path <- Sys.getenv("LD_LIBRARY_PATH")
@@ -25,14 +60,16 @@ fix_ld_library_path <- function() {
   invisible(old_ld_library_path)
 }
 
+#' 
 #' @noRd
 poll_sys_proc_blocking <- function(cmd, proc,
                                    what = c("nix-build", "expr"),
                                    message_type = 
                                      c("simple", "quiet", "verbose")
-) {
-  what <- match.arg(what)
-  message_type <- match.arg(message_type)
+                                   ) {
+  what <- match.arg(what, choices = c("nix-build", "expr"))
+  message_type <- match.arg(message_type,
+    choices = c("simple", "quiet", "verbose"))
   is_quiet <- message_type == "quiet"
   
   status <- proc$status
@@ -46,9 +83,19 @@ poll_sys_proc_blocking <- function(cmd, proc,
     }
   }
   
-  return(invisible(status))
+  # return(invisible(status))
 }
 
+#' Poll running non-blocking process started previously via.
+#' `sys::exec_background()`
+#' 
+#' Typically, it is used for a `nix-build` process launched via `nix_build()`
[…]
