### 2024-06-09 160f369
simplify message in `rix_init()`
