### 2024-06-09 71446d9
document more `with_nix()` helpers
