+++ R/with_nix_helpers.R
@@ -22,7 +22,17 @@ stop_no_nix_shell <- function() {
 }
 
 
-#' serialize language objects
+#' Serialize a list of R expressions as `.Rds` to disk
+#' 
+#' This helper is currently only called from the wrapper `serialize_globals()`,
+#' where all recursively found global R objects of `expr` need to be saved
+#' on disk in a temporary directory, so that they can later be deserialized
+#' inside the Nix R environment
+#' @param lobjs list of **R** expressions
+#' @param file path of temporary directory where list of expressions are 
+#' saved as individual `.Rds` files
+#' @details It is called for its side effects to save expressions as `.Rds`
+#' files.
 #' @noRd
 serialize_lobjs <- function(lobjs, temp_dir) {
   invisible({
@@ -42,7 +52,13 @@ serialize_lobjs <- function(lobjs, temp_dir) {
   })
 }
 
-#' serialize arguments
+#' Get all function args of `expr` as R objects and save them into `.Rds` files.
+#' 
+#' Save function arguments into a folder each with `<tag.Rds>` and `value`.
+#' This is used for the first serialization step in the source environment
+#' inside `with_nix()`.
+#' @param args list of symbols where names of elements are character
+#' representations of symbols, or list containing empty symbol(s)
 #' @noRd
 serialize_args <- function(args, temp_dir) {
   invisible({
@@ -63,8 +79,8 @@ serialize_args <- function(args, temp_dir) {
   })
 }
 
+#' to determine which extra packages to load in Nix R prior evaluating `expr`
 #' @noRd
-# to determine which extra packages to load in Nix R prior evaluating `expr`
 get_expr_extra_pkgs <- function(globals_expr) {
   envs_check <- lapply(globals_expr, where)
   names_envs_check <- vapply(envs_check, environmentName, character(1L))
@@ -83,10 +99,21 @@ get_expr_extra_pkgs <- function(globals_expr) {
 }
 
 
+#' Check if the current environment is the empty environment
+#' @return logical vector of length one
 #' @noRd
 is_empty <- function(x) identical(x, emptyenv())
 
 
+#' Find the environment where R object is defined
+#' 
+#' Is used by helper `classify_globals()`, to return the environment where
+#' the object called `name`. The environment stack is queried until the empty
+#' environment is reached.
+#' @param name  string with the name of the R object (a global)
+#' @param env environment (class) where to search is started in direction to
+#' the empty environment
+#' @return environment (class), where object of called `<name>` is found
 #' @noRd
 where <- function(name, env = parent.frame()) {
   while(!is_empty(env)) {
@@ -99,7 +126,8 @@ where <- function(name, env = parent.frame()) {
 }
 
 #' Finds and checks global functions and variables recursively for closure
-#' `expr`
+#' @param expr an **R** expression
+#' @param args_vec character vector with arguments
 #' @noRd
 recurse_find_check_globals <- function(expr,
                                        args_vec,
@@ -382,8 +410,25 @@ serialize_pkgs <- function(globals_expr, temp_dir) {
   return(pkgs)
 }
 
-# build deparsed script via language objects;
-# reads like R code, and avoids code injection
+#' Quote language objects via partial substitution of expressions
+#' reads like R code, and avoids code injection.
+#' 
+#' This is used to boilerplate a custom R script that is evaluated by calling
+#' `Rscript` in Nix target environment
+#' @param expr Typically a function
+#' @param program string, currently `"R"`
+#' @param message_type character vector of length one with message type;
[…]
