### 2024-06-10 #625 @ldecicco-USGS
@b-rodrigues Just a heads up that I'll be on leave from June 17-July 8 and may or may not be able to get to this issue during that time.  @assignUser and @wdwatkins , when the updates and response are available, feel free to get started on responding to the changes.

If all looks good, you can use this template for the final reviewer approval:
https://devdevguide.netlify.app/approval2template
