### 2024-06-13 feb246c
Update b1-setting-up-and-using-rix-on-linux-and-windows.Rmd
