### 2024-06-14 69276fe
general vignettes cleanup
