+++ vignettes/z-binary_cache.Rmd
@@ -102,18 +102,18 @@ instance, for package development).
 
 ## Using your cache
 
-Whether you decide to build the packages on Github or locally, to then use your cache, you need to open an
-account on [Cachix](https://www.cachix.org/). The free tier includes 5GB of
-space, which is more than enough for several development environments. Once your
-account is done, create a personal auth token so that the Github Actions workflow (or
-your computer, if building locally) can
-authenticate to your Cachix account. Then, copy this token, and go to your
-Github repository’s settings, then *Secrets and variables > Actions* and add a
-new repository secret. Copy the token in the `Secret` field and name the secret
-`CACHIX_AUTH`. Now the action can authenticate to Cachix and push the binaries
-it builds! If building locally, simply run `cachix authtoken <TOKEN>` in your 
-terminal before building (follow the instructions on Cachix website to learn
-how to push the binaries afterwards).
+Whether you decide to build the packages on Github or locally, to then use your
+cache, you need to open an account on [Cachix](https://www.cachix.org/). The
+free tier includes 5GB of space, which is more than enough for several
+development environments. Once your account is done, create a personal auth
+token so that the Github Actions workflow (or your computer, if building
+locally) can authenticate to your Cachix account. Then, copy this token, and go
+to your Github repository’s settings, then *Secrets and variables > Actions* and
+add a new repository secret. Copy the token in the `Secret` field and name the
+secret `CACHIX_AUTH`. Now the action can authenticate to Cachix and push the
+binaries it builds! If building locally, simply run `cachix authtoken <TOKEN>`
+in your terminal before building (follow the instructions on Cachix website to
+learn how to push the binaries afterwards).
 
 To use your personalized cache on your computers, run the following commands on
 your computer. First, install the Cachix client:
