### 2024-06-14 3854c32
started work on install local packages
