### 2024-06-15 44a334e
factor out try_get_request(); no more httr
