### 2024-06-15 4de65f0
remove unneeded functions
