### 2024-06-15 2970692
document `classify_globals()`
