### 2024-06-15 3e07b6f
Document `serialize_globals()`
