### 2024-06-15 37f7d19
enable quiet `with_nix()` when external R packages are detected
