### 2024-06-15 1e1894f
`fix_ld_library_path`: complete return
