### 2024-06-15 b4a4ee2
`serialize_pkgs`: add internal docs and allow quiet messaging
