### 2024-06-15 3a55289
httr -> curl; remove httr dep; implement in curl
