+++ R/get_latest.R
@@ -1,24 +1,24 @@
-#' get_latest Get the latest R version and packages
+#' Get the latest R version and packages
 #' @param r_version Character. R version to look for, for example, "4.2.0". If a nixpkgs revision is provided instead, this gets returned.
 #' @return A character. The commit hash of the latest nixpkgs-unstable revision
-#' @importFrom httr content GET stop_for_status
+#' @importFrom curl new_handle curl_fetch_memory handle_reset
 #' @importFrom jsonlite fromJSON
 #' @importFrom curl has_internet
 #'
 #' @noRd
 get_latest <- function(r_version) {
-
+  
   is_online <- has_internet()
-
+  
   stopifnot("r_version has to be a character." = is.character(r_version))
-
+  
   # If the use provides a commit, then the commit gets used.
   # User needs to know which repo it belongs to
   if(nchar(r_version) == 40){
     return(r_version)
   } else if(
-           !(r_version %in% c("bleeding_edge", "frozen_edge", available_r()))
-         ){
+    !(r_version %in% c("bleeding_edge", "frozen_edge", available_r()))
+  ){
     stop("The provided R version is likely wrong.\nPlease check that you provided a correct R version.\nYou can list available versions using `available_r()`.\nYou can also directly provide a commit, but you need to make sure it points to the right repo used by `rix()`.\nYou can also use 'bleeding_edge' and 'frozen_edge'.")
   } else if(!is_online){
     stop("ERROR! You don't seem to be connected to the internet.")
@@ -33,33 +33,50 @@ get_latest <- function(r_version) {
 
 #' @noRd
 get_right_commit <- function(r_version) {
-
+  
   if(r_version == "frozen_edge"){
     api_url <- "https://api.github.com/repos/rstats-on-nix/nixpkgs/commits?sha=r-daily"
-
-  } else if(r_version %in% Filter(\(x)`!=`(x, "latest"), available_r())){ #all but latest
+    
+  } else if (r_version %in% Filter(\(x)`!=`(x, "latest"), available_r())){ #all but latest
     temp <- new.env(parent = emptyenv())
-
+    
     data(list = "r_nix_revs",
          package = "rix",
          envir = temp)
-
+    
     get("r_nix_revs", envir = temp)
-
+    
     return(r_nix_revs$revision[r_nix_revs$version == r_version])
-
+    
   } else {
     api_url <- "https://api.github.com/repos/NixOS/nixpkgs/commits?sha=nixpkgs-unstable"
   }
-  tryCatch({
-    response <- httr::GET(url = api_url)
-    httr::stop_for_status(response)
-    commit_data <- jsonlite::fromJSON(httr::content(response, "text"))
-    latest_commit <- commit_data$sha[1]
-    return(latest_commit)
-  }, error = function(e) {
-    cat("Error:", e$message, "\n")
-    return(NULL)
-  })
+  
+  # handle to get error for status code 404
+  h <- curl::new_handle(failonerror = TRUE)
+  
+  req <- try_get_request(url = api_url, handle = h)
+  
+  curl::handle_reset(h)
+  
+  commit_data <- jsonlite::fromJSON(rawToChar(req$content))
+  latest_commit <- commit_data$sha[1]
+  
+  return(latest_commit)
+}
 
+
+#' Try fetch contents of an URL with handle and stop with propagating the curl
+#' error and also show URL for context
+#' @noRd
+try_get_request <- function(url, handle) {
+  tryCatch({
+    req <- curl::curl_fetch_memory(url, handle)
+  }, error = function(e) {
+    stop("`curl::curl_fetch_memory(", 
+         paste0("url = ", "'", api_url, "'", ")` "), "failed:\n ",
+         e$message[1], call. = FALSE)
+  })
+  
+  return(req)
 }
