### 2024-06-17 e243ea5
fix vignette build; `api_url` -> `url`
