### 2024-06-17 c3a582f
more check args to fix
