+++ .github/workflows/tests-r-via-nix.yaml
@@ -34,7 +34,7 @@ jobs:
       - name: devtools::test() via nix-shell
         run: nix-shell --pure -p cacert nix R rPackages.covr rPackages.codetools rPackages.xml2 rPackages.sys rPackages.devtools --run "Rscript -e 'devtools::test()'"
       #- name: Test coverage
-      #  run: nix-shell --pure -p nix git cacert R rPackages.covr rPackages.codetools rPackages.sys rPackages.xml2 rPackages.httr rPackages.testthat --run "Rscript -e 'covr::codecov()' "
+      #  run: nix-shell --pure -p nix git cacert R rPackages.covr rPackages.codetools rPackages.sys rPackages.xml2 rPackages.testthat --run "Rscript -e 'covr::codecov()' "
 
       - name: Show testthat output
         if: always()
