### 2024-06-17 cf1fa84
added possibility to install packages from source
