+++ NAMESPACE
@@ -10,6 +10,7 @@ export(with_nix)
 importFrom(codetools,checkUsage)
 importFrom(codetools,findGlobals)
 importFrom(curl,has_internet)
+importFrom(desc,description)
 importFrom(httr,GET)
 importFrom(httr,content)
 importFrom(httr,http_error)
