### 2024-06-19 067986a
removed test for fetchlocals, it's going to be on GA instead
