+++ R/zzz.R
@@ -1,4 +1,4 @@
 #' zzz Global imports
 #' @noRd
 #' @importFrom utils data
-utils::globalVariables(c("r_nix_revs", "is_internet_down"))
+utils::globalVariables(c("r_nix_revs", "is_internet_down", "type", "package"))
