+++ NAMESPACE
@@ -16,4 +16,6 @@ importFrom(httr,content)
 importFrom(httr,http_error)
 importFrom(httr,stop_for_status)
 importFrom(jsonlite,fromJSON)
+importFrom(stats,na.omit)
 importFrom(utils,data)
+importFrom(utils,tail)
