### 2024-06-19 84b0327
Update test-fetchlocal.yml
