### 2024-06-19 6997f6a
Update test-fetchlocal.yml
