+++ .github/workflows/test-fetchlocal.yml
@@ -33,9 +33,16 @@ jobs:
       - name: Nix cache
         uses: DeterminateSystems/magic-nix-cache-action@main
 
-      - name: Build default.nix with local packages
+      - name: Build default.nix with 2 local packages
         run: |
           nix-shell ./inst/extdata/default.nix --run "Rscript -e \"library(rix);rix('4.3.1', local_pkgs = c('chronicler_0.2.1.tar.gz', 'knitr_1.43.tar.gz'), overwrite = TRUE)\""
+
+      - name: Build default.nix
+        run: nix-build
+
+      - name: Build default.nix with 1 local package
+        run: |
+          nix-shell ./inst/extdata/default.nix --run "Rscript -e \"library(rix);rix('4.3.1', local_pkgs = 'chronicler_0.2.1.tar.gz', overwrite = TRUE)\""
 
       - name: Build default.nix
         run: nix-build
