+++ NAMESPACE
@@ -9,12 +9,11 @@ export(tar_nix_ga)
 export(with_nix)
 importFrom(codetools,checkUsage)
 importFrom(codetools,findGlobals)
+importFrom(curl,curl_fetch_memory)
+importFrom(curl,handle_reset)
 importFrom(curl,has_internet)
+importFrom(curl,new_handle)
 importFrom(desc,description)
-importFrom(httr,GET)
-importFrom(httr,content)
-importFrom(httr,http_error)
-importFrom(httr,stop_for_status)
 importFrom(jsonlite,fromJSON)
 importFrom(stats,na.omit)
 importFrom(utils,data)
