+++ DESCRIPTION
@@ -18,7 +18,6 @@ Imports:
     codetools,
     curl,
     desc,
-    httr,
     jsonlite,
     sys,
     utils
