### 2024-06-22 174484e
remove `exec_mode` API; add `message_type`
