### 2024-06-22 770e2ef
kill process on exit; depreciate `exec_mode`
