+++ R/with_nix.R
@@ -70,13 +70,6 @@
 #' @param program String stating where to evaluate the expression. Either `"R"`,
 #' the default, or `"shell"`. `where = "R"` will evaluate the expression via
 #' `RScript` and `where = "shell"` will run the system command in `nix-shell`.
-#' @param exec_mode Either `"blocking"` (default) or `"non-blocking`. This
-#' will either block the R session while `expr` is running in a `nix-shell`
-#' environment, oor running it in the background ("non-blocking"). While
-#' `program = R` will yield identical results for foreground and background
-#' evaluation (R object), `program = "shell"` will return list of exit status,
-#' standard output and standard error of the system command and as text in
-#' blocking mode.
 #' @param project_path Path to the folder where the `default.nix` file resides. 
 #' The default is `"."`, which is the working directory in the current R
 #' session. This approach also useful when you have different subfolders 
@@ -112,7 +105,7 @@
 #' # stream messages, and bring output back to current R session
 #' out <- with_nix(
 #'   expr = function(mtcars) nrow(mtcars),
-#'   program = "R", exec_mode = "non-blocking", project_path = project_path,
+#'   program = "R", project_path = project_path,
 #'   message_type = "simple"
 #' )
 #' 
@@ -128,7 +121,7 @@
 #' 
 #' out <- with_nix(
 #'   expr = get_sample(seed = 1234, n = 5),
-#'   program = "R", exec_mode = "non-blocking",
+#'   program = "R",
 #'   project_path = ".",
 #'   message_type = "simple"
 #' )
@@ -139,7 +132,6 @@
 #' }
 with_nix <- function(expr,
                      program = c("R", "shell"),
-                     exec_mode = c("blocking", "non-blocking"),
                      project_path = ".",
                      message_type = c("simple", "quiet", "verbose")) {
   nix_file <- file.path(project_path, "default.nix")
@@ -154,7 +146,6 @@ with_nix <- function(expr,
   )
 
   program <- match.arg(program, choices = c("R", "shell"))
-  exec_mode <- match.arg(exec_mode, choices = c("blocking", "non-blocking"))
   message_type <- match.arg(message_type,
     choices = c("simple", "quiet", "verbose"))
   
@@ -278,8 +269,7 @@ with_nix <- function(expr,
     
     # 3) run expression in nix session, based on temporary script
     if (isFALSE(is_quiet)) {
-      cat(paste0("\n==> running deparsed expression via `nix-shell`", " in ",
-        exec_mode, " mode ...\n\n"
+      cat(paste0("\n==> running deparsed expression via `nix-shell`...\n\n"
       ))
     }
 
@@ -294,20 +284,11 @@ with_nix <- function(expr,
       )
     )
     
-    proc <- switch(exec_mode,
-      "blocking" = sys::exec_internal(cmd = "nix-shell", cmd_rnix_deparsed),
-      "non-blocking" = sys::exec_background(
-        cmd = "nix-shell", cmd_rnix_deparsed),
-      stop('invalid `exec_mode`. Either use "blocking" or "non-blocking"')
-    )
-    if (exec_mode == "non-blocking") {
-      poll_sys_proc_nonblocking(
-        cmd = cmd_rnix_deparsed, proc, what = "expr",
-        message_type = message_type)
-    } else if (exec_mode == "blocking") {
-      poll_sys_proc_blocking(cmd = cmd_rnix_deparsed, proc, what = "expr",
-        message_type = message_type)
-    }
[…]
