### 2024-06-22 798b880
remove `exec_mode` interface; adjust tests
