+++ NAMESPACE
@@ -16,5 +16,6 @@ importFrom(curl,new_handle)
 importFrom(desc,description)
 importFrom(jsonlite,fromJSON)
 importFrom(stats,na.omit)
+importFrom(tools,pskill)
 importFrom(utils,data)
 importFrom(utils,tail)
