+++ R/with_nix.R
@@ -120,7 +120,7 @@
 #' }
 #' 
 #' out <- with_nix(
-#'   expr = get_sample(seed = 1234, n = 5),
+#'   expr = function() get_sample(seed = 1234, n = 5),
 #'   program = "R",
 #'   project_path = ".",
 #'   message_type = "simple"
