### 2024-06-22 f916f22
also indent fetchzip(s) lines
