### 2024-06-22 47558e5
update tests for fetchzip(s) and fetchgit(s)
