+++ tests/testthat/test-fetchgit.R
@@ -7,8 +7,8 @@ testthat::test_that("Test fetchgit works", {
            commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"
            )
     ),
-    "\n  (pkgs.rPackages.buildRPackage {\n    name = \"housing\";\n    src = pkgs.fetchgit {\n     url = \"https://github.com/rap4all/housing/\";\n     branchName = \"fusen\";\n     rev = \"1c860959310b80e67c41f7bbdc3e84cef00df18e\";\n     sha256 = \"sha256-s4KGtfKQ7hL0sfDhGb4BpBpspfefBN6hf+XlslqyEn4=\";\n    };\n    propagatedBuildInputs = builtins.attrValues {\n     inherit (pkgs.rPackages) dplyr ggplot2 janitor purrr readxl rlang rvest stringr tidyr;\n    };\n  })\n"
-    )
+    "\n    (pkgs.rPackages.buildRPackage {\n      name = \"housing\";\n      src = pkgs.fetchgit {\n        url = \"https://github.com/rap4all/housing/\";\n        branchName = \"fusen\";\n        rev = \"1c860959310b80e67c41f7bbdc3e84cef00df18e\";\n        sha256 = \"sha256-s4KGtfKQ7hL0sfDhGb4BpBpspfefBN6hf+XlslqyEn4=\";\n      };\n      propagatedBuildInputs = builtins.attrValues {\n        inherit (pkgs.rPackages) dplyr ggplot2 janitor purrr readxl rlang rvest stringr tidyr;\n      };\n    })\n"
+  )
 })
 
 testthat::test_that("Test fetchgit fails gracefully", {
@@ -21,14 +21,14 @@ testthat::test_that("Test fetchgit fails gracefully", {
            )
     ), "Are these correct?"
   )
-  })
+})
 
 
 
 testthat::test_that("Test fetchzip works", {
   testthat::expect_equal(
     fetchzip("AER@1.2-8"),
-    "\n  (pkgs.rPackages.buildRPackage {\n    name = \"AER\";\n    src = pkgs.fetchzip {\n     url = \"https://cran.r-project.org/src/contrib/Archive/AER/AER_1.2-8.tar.gz\";\n     sha256 = \"sha256-OqxXcnUX/2C6wfD5fuNayc8OU+mstI3tt4eBVGQZ2S0=\";\n    };\n    propagatedBuildInputs = builtins.attrValues {\n     inherit (pkgs.rPackages) car lmtest sandwich survival zoo Formula;\n    };\n  })\n"
+    "\n    (pkgs.rPackages.buildRPackage {\n      name = \"AER\";\n      src = pkgs.fetchzip {\n       url = \"https://cran.r-project.org/src/contrib/Archive/AER/AER_1.2-8.tar.gz\";\n       sha256 = \"sha256-OqxXcnUX/2C6wfD5fuNayc8OU+mstI3tt4eBVGQZ2S0=\";\n      };\n      propagatedBuildInputs = builtins.attrValues {\n        inherit (pkgs.rPackages) car lmtest sandwich survival zoo Formula;\n      };\n    })\n"
     )
 })
 
@@ -53,14 +53,14 @@ testthat::test_that("Test fetchgits", {
          commit = "d617172447d2947efb20ad6a4463742b8a5d79dc")
       )
     ),
[…]
