### 2024-06-22 4dbcc53
adjust pkgs.mkShell indentation
