+++ vignettes/z-advanced-topic-using-nix-inside-docker.Rmd
@@ -1,0 +1,197 @@
+---
+title: "z - Advanced topic: Using Nix inside Docker"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{z-advanced-topic-using-nix-inside-doceker}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r, include=FALSE}
+library(rix)
+```
+
+## Introduction
+
+It might look like Nix is an alternative to Docker, but that's not really
+the case. Docker is a containerisation tool, while Nix is a package manager.
+You can use Nix in such a way that you don't need Docker anymore, but if you're
+already invested in Docker, you don't have to abandon it and still benefit
+from Nix.
+
+## A generic Dockerfile
+
+This `Dockerfile` uses `ubuntu:latest` as a base image, and then uses the Nix
+package manager to set up a complete development environment:
+
+```
+FROM ubuntu:latest
+
+RUN apt update -y
+
+RUN apt install curl -y
+
+# We don't have R nor {rix} in this image, so we can bootstrap it by downloading
+# the default.nix file that comes with {rix}. You can also download it beforehand
+# and then copy it to the Docker image 
+RUN curl -O https://raw.githubusercontent.com/b-rodrigues/rix/master/inst/extdata/default.nix
+
+# Copy a script to generate the environment of interest using {rix}
+COPY generate_env.R .
+
+# The next 4 lines install Nix inside Docker. See the Determinate Systems installer's documentation
+RUN curl --proto '=https' --tlsv1.2 -sSf -L https://install.determinate.systems/nix | sh -s -- install linux \
+  --extra-conf "sandbox = false" \
+  --init none \
+  --no-confirm
+
+# Adds Nix to the path, as described by the Determinate Systems installer's documentation
+ENV PATH="${PATH}:/nix/var/nix/profiles/default/bin"
+
+# This will overwrite the default.nix we downloaded previously with a new
+# expression generated from running `generate_env.R`
+RUN nix-shell --run "Rscript generate_env.R"
+
+# We now build the environment
+RUN nix-build
+
+# Finally, we run `nix-shell`. This will get executed when running
+# containers from this image. You can of course put anything in here
+CMD nix-shell
+```
+
+It doesn't matter that we are using `ubuntu:latest` as a base image, which is
+usually not recommended for reproducibility purposes, since it is Nix that takes
+care of ensuring that our environment is reproducible.
+
+Here is an example of the `generate_env.R` file:
+
+```{r, eval = FALSE}
+library(rix)
+
+rix(r_ver = "4.3.1",
+    r_pkgs = c("dplyr", "ggplot2"),
+    ide = "other",
+    project_path = ".",
+  	shell_hook = "R",
+    overwrite = TRUE)
+```
+
+Using Nix to handle the setup of the environment, even inside Docker, creates a
+nice separation of concerns. On one hand, you can continue using Docker to serve
+applications, and on the other hand, you can use Nix to ensure you don't have to
+store images, as you can always rebuild the correct environment in a completely
+reproducible manner.
+
+## Dockerizing a Shiny application
+
+*Dockerizing* a Shiny application using Nix is very easy as well. You can keep
[…]
