+++ R/get_imports.R
@@ -1,0 +1,7 @@
+get_imports <- function(path){
+  tempdir <- tempdir()
+  untar(path, exdir = tempdir)
+  paths <- list.files(tempdir, full.names = TRUE, recursive = TRUE)
+  desc_path <- 
+  as.data.frame(read.dcf(desc_path))$Imports
+}
