+++ chronicler/inst/doc/maybe-monad.R
@@ -1,0 +1,34 @@
+## ---- include = FALSE---------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
+## ----setup--------------------------------------------------------------------
+library(chronicler)
+library(maybe)
+
+## -----------------------------------------------------------------------------
+m_sqrt <- maybe(sqrt)
+
+m_sqrt(16)
+
+## -----------------------------------------------------------------------------
+m_sqrt("10")
+
+## -----------------------------------------------------------------------------
+r_sqrt <- record(sqrt)
+
+r_sqrt("16")
+
+## -----------------------------------------------------------------------------
+r_sqrt <- record(sqrt)
+
+r_sqrt(16)
+
+## -----------------------------------------------------------------------------
+pick(r_sqrt(16), "value")
+
+## -----------------------------------------------------------------------------
+(r_sqrt(16))$value
+
