+++ chronicler/tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(chronicler)
+
+test_check("chronicler")
