+++ chronicler/R/chronicle.R
@@ -1,0 +1,599 @@
+#' Creates the log_df element of a chronicle object.
+#' @param ops_number Tracks the number of the operation in a chain of operations.
+#' @param success Did the operation succeed?
+#' @param fstring The function call.
+#' @param args The arguments of the call.
+#' @param res_pure The result of the purely call.
+#' @param start Starting time.
+#' @param end Ending time.
+#' @param .g Optional. A function to apply to the intermediary results for monitoring purposes. Defaults to returning NA.
+#' @param diff_obj Optional. Output of the `diff` parameter in `record()`.
+#' @importFrom tibble tibble
+#' @importFrom maybe from_maybe nothing
+#' @return A tibble containing the log.
+make_log_df <- function(ops_number = 1,
+                        success,
+                        fstring,
+                        args,
+                        res_pure,
+                        start = Sys.time(),
+                        end = Sys.time(),
+                        .g = (\(x) NA),
+                        diff_obj = NULL){
+
+  outcome <- ifelse(success == 1,
+                    "OK! Success",
+                    "NOK! Caution - ERROR")
+
+  tibble::tibble(
+            "ops_number" = ops_number,
+            "outcome" = outcome,
+            "function" = fstring,
+            "arguments" = args,
+            "message" = paste0(res_pure$log_df, collapse = " "),
+            "start_time" = start,
+            "end_time" = end,
+            "run_time" = end - start,
+            "g" = list(.g(maybe::from_maybe(res_pure$value,
+                                            default = maybe::nothing()))),
+            "diff_obj" = list(diff_obj),
+            "lag_outcome" = NA
+          )
+
+}
+
+
+#' Reads the log of a chronicle.
+#' @param .c A chronicle object.
+#' @return The log of the object.
+#' @examples
+#' \dontrun{
+#' read_log(chronicle_object)
+#' }
+#' @export
+read_log <- function(.c){
+
+  log_df <- .c$log_df
+
+  make_func_call <- function(log_df, i){
+
+    paste0(paste0(log_df[i, c("function", "arguments")],
+                  collapse = "("),
+           ")")
+
+  }
+
+  is_success <- function(log_df, i){
+
+    ifelse(grepl("Success", log_df$outcome[i]),
+           "successfully",
+           paste0("unsuccessfully with following exception: ", log_df$message[i]))
+
+  }
+
+  success_symbol <- function(log_df, i){
+
+    ifelse(grepl("Success", log_df$outcome[i]),
+           "OK!",
+           "NOK!")
+
+  }
+
+  make_sentence <- function(log_df, i){
+
+    paste(success_symbol(log_df, i),
+          make_func_call(log_df, i),
+          "ran",
+          is_success(log_df, i),
+          "at",
+          log_df$start_time[i])
+
+  }
+
+  total_runtime <- function(log_df){
+
+    total_time <- log_df$run_time
+
+    unit <- attr(total_time, "units")
+
+    paste(as.numeric(sum(log_df$run_time)), unit)
+
+  }
+
+
+  sentences <- vector(length = nrow(log_df))
+
+  for(i in 1:nrow(log_df)){
+
+  sentences[i] <-  make_sentence(log_df, i)
+
+  }
+
+  c("Complete log:", sentences, paste("Total running time:", total_runtime(log_df)))
+
+}
+
+
+#' Print method for chronicle objects.
+#' @param x A chronicle object.
+#' @param ... Unused.
+#' @return No return value, called for side effects (printing the object on screen).
+#' @details
+#' `chronicle` object are, at their core, lists with the following elements:
[…]
