+++ chronicler/R/zzz.R
@@ -1,0 +1,1 @@
+utils::globalVariables(c("outcome"))
