+++ chronicler/man/avia.Rd
@@ -1,0 +1,16 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{avia}
+\alias{avia}
+\title{Air passenger transport between the main airports of Luxembourg and their main partner airports}
+\format{
+A data frame with 510 rows and 238 columns.
+}
+\usage{
+avia
+}
+\description{
+A non-tidy dataset from EUROSTAT which can be found \href{https://ec.europa.eu/eurostat/databrowser/bookmark/f9e65cb1-39d2-4b80-9a8b-3b3076b95575?lang=en}{here}.
+}
+\keyword{datasets}

[…]
