+++ chronicler/inst/doc/advanced-topics.R
@@ -1,219 +1,0 @@
-## ---- include = FALSE---------------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-
-## ----setup, include = FALSE---------------------------------------------------
-library(chronicler)
-library(testthat)
-
-## -----------------------------------------------------------------------------
-my_sqrt <- function(x){
-
-  sqrt(x)
-
-}
-
-
-## -----------------------------------------------------------------------------
-my_sqrt <- function(x, log = ""){
-
-  list(sqrt(x),
-       c(log,
-         paste0("Running sqrt with input ", x)))
-
-}
-
-
-## -----------------------------------------------------------------------------
-my_log <- function(x, log = ""){
-
-  list(log(x),
-       c(log,
-         paste0("Running log with input ", x)))
-
-}
-
-
-## -----------------------------------------------------------------------------
-10 |>
-  sqrt() |>
-  log()
-
-
-## ---- eval = FALSE------------------------------------------------------------
-#  10 |>
-#    my_sqrt() |>
-#    my_log()
-#  
-
-## -----------------------------------------------------------------------------
-log_it <- function(.f, ..., log = NULL){
-
-  fstring <- deparse(substitute(.f))
-
-  function(..., .log = log){
-
-    list(result = .f(...),
-         log = c(.log,
-                 paste0("Running ", fstring, " with argument ", ...)))
-  }
-}
-
-
-## -----------------------------------------------------------------------------
-l_sqrt <- log_it(sqrt)
-
-l_sqrt(10)
-
-l_log <- log_it(log)
-
-l_log(10)
-
-
-## -----------------------------------------------------------------------------
-bind <- function(.l, .f, ...){
-
-  .f(.l$result, ..., .log = .l$log)
-
-}
-
-## -----------------------------------------------------------------------------
-10 |>
-  l_sqrt() |>
-  bind(l_log)
-
-
-## -----------------------------------------------------------------------------
-log(sqrt(10))
-
-## -----------------------------------------------------------------------------
-unit <- log_it(identity)
-
-## -----------------------------------------------------------------------------
-fmap <- function(m, f, ...){
-
-  fstring <- deparse(substitute(f))
-
-  list(result = f(m$result, ...),
-       log = c(m$log,
-               paste0("fmapping ", fstring, " with arguments ", paste0(m$result, ..., collapse = ","))))
-}
-
-
-## -----------------------------------------------------------------------------
-# Let’s use unit(), which we defined above, for this.
-
-(m <- unit(10))
-
-## -----------------------------------------------------------------------------
-fmap(m, log)
-
-## -----------------------------------------------------------------------------
-fmap(m, l_log)
-
-## -----------------------------------------------------------------------------
-flatten <- function(m){
-
-  list(result = m$result$result,
-       log = c(m$log))
-
-}
-
-
-## -----------------------------------------------------------------------------
-flatten(fmap(m, l_log))
-
-## -----------------------------------------------------------------------------
-# I first define a composition operator for functions
-`%.%` <- \(f,g)(function(...)(f(g(...))))
-
-# I now compose flatten() and fmap()
-# flatten %.% fmap is read as "flatten after fmap"
-flatmap <- flatten %.% fmap
-
-
-## -----------------------------------------------------------------------------
-10 |>
-  l_sqrt() |>
-  bind(l_log)
-
-
[…]
