+++ chronicler/R/data.R
@@ -1,6 +1,0 @@
-#' Air passenger transport between the main airports of Luxembourg and their main partner airports
-#'
-#' A non-tidy dataset from EUROSTAT which can be found [here](https://ec.europa.eu/eurostat/databrowser/bookmark/f9e65cb1-39d2-4b80-9a8b-3b3076b95575?lang=en).
-#'
-#' @format A data frame with 510 rows and 238 columns.
-"avia"
