+++ chronicler/tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(chronicler)
-
-test_check("chronicler")
