### 2024-06-23 961f686
removed unneeded script
