+++ R/rix.R
@@ -1,24 +1,29 @@
-#' rix Generates a Nix expression that builds a reproducible development environment
+#' Generate a Nix expression that builds a reproducible development
+#' environment
 #' @return Nothing, this function only has the side-effect of writing a file
 #'   called "default.nix" in the working directory. This file contains the
 #'   expression to build a reproducible environment using the Nix package
 #'   manager.
-#' @param r_ver Character, defaults to "latest". The required R version, for example "4.0.0".
-#'   You can check which R versions are available using `available_r`.
-#'   For reproducibility purposes, you can also provide a nixpkgs revision directly.
-#'   It is also possible to provide either "bleeding_edge" or "frozen_edge" if you
-#'   need an environment with bleeding edge packages. Read more in the "Details" below.
+#' @param r_ver Character, defaults to "latest". The required R version, for
+#'   example "4.0.0". You can check which R versions are available using
+#'   `available_r`. For reproducibility purposes, you can also provide a nixpkgs
+#'   revision directly. It is also possible to provide either "bleeding_edge" or
+#'   "frozen_edge" if you need an environment with bleeding edge packages. Read
+#'   more in the "Details" below.
 #' @param r_pkgs Vector of characters. List the required R packages for your
 #'   analysis here.
-#' @param system_pkgs Vector of characters. List further software you wish to install that
-#'   are not R packages such as command line applications for example. You can look for
-#'   available software on the NixOS website \url{https://search.nixos.org/packages?channel=unstable&from=0&size=50&sort=relevance&type=packages&query=}
-#' @param git_pkgs List. A list of packages to install from Git. See details for more information.
-#' @param local_pkgs List. A list of paths to local packages to install. These packages need to be in the `.tar.gz` or `.zip` formats.
-#' @param tex_pkgs Vector of characters. A set of tex packages to install.
-#'   Use this if you need to compile `.tex` documents,
-#'   or build PDF documents using Quarto. If you don't know which package to add,
-#'   start by adding "amsmath". See the Vignette "Authoring LaTeX documents" for more details.
+#' @param system_pkgs Vector of characters. List further software you wish to
+#'   install that are not R packages such as command line applications for
+#'   example. You can look for available software on the NixOS website
+#'   \url{https://search.nixos.org/packages?channel=unstable&from=0&size=50&sort=relevance&type=packages&query=}
+#' @param git_pkgs List. A list of packages to install from Git. See details for
+#' more information.
+#' @param local_pkgs List. A list of paths to local packages to install. These
+#' packages need to be in the `.tar.gz` or `.zip` formats.
+#' @param tex_pkgs Vector of characters. A set of tex packages to install. Use
+#'   this if you need to compile `.tex` documents, or build PDF documents using
+#'   Quarto. If you don't know which package to add, start by adding "amsmath".
+#'   See the Vignette "Authoring LaTeX documents" for more details.
 #' @param ide Character, defaults to "other". If you wish to use RStudio to work
 #'   interactively use "rstudio" or "rserver" for the server version. Use "code"
 #'   for Visual Studio Code. You can also use "radian", an interactive REPL. For
@@ -28,9 +33,10 @@
 #'   Where to write `default.nix`, for example "/home/path/to/project".
 #'   The file will thus be written to the file
 #'   "/home/path/to/project/default.nix".
-#' @param overwrite Logical, defaults to FALSE. If TRUE, overwrite the `default.nix`
-#'   file in the specified path.
-#' @param print Logical, defaults to FALSE. If TRUE, print `default.nix` to console.
+#' @param overwrite Logical, defaults to FALSE. If TRUE, overwrite the
+#'   `default.nix` file in the specified path.
[…]
