+++ create_dev_env.R
@@ -4,9 +4,10 @@ latest_commit <- sys::as_text(sys::exec_internal("git", c("rev-parse", "master")
 
 
 rix(r_ver = "bleeding_edge",
-    r_pkgs = c("devtools", "diffviewer", "fledge", "codetools", "jsonlite",
-               "httr", "sys", "testthat", "knitr", "rmarkdown"),
-    system_pkgs = c("R", "glibcLocalesUtf8", "pandoc", "nix"),
+    r_pkgs = c("devtools", "diffviewer", "fledge", "styler", "precommit",
+               "codetools", "jsonlite",  "httr", "sys", "testthat", "knitr",
+               "rmarkdown", "renv", "digest"),
+    system_pkgs = c("R", "glibcLocalesUtf8", "pandoc", "nix", "pre-commit"),
     tex_pkgs = "scheme-small",
     ide = "other",
     project_path = ".",
