### 2024-06-23 9bcec9d
uninstall precommit setup
