+++ R/fetchers.R
@@ -1,9 +1,8 @@
 #' fetchgit Downloads and installs a package hosted of Git
-#' @param git_pkg A list of four elements: "package_name", the name of the package, "repo_url", the repository's url, "branch_name", the name of the branch containing the code to download and "commit", the commit hash of interest. 
+#' @param git_pkg A list of four elements: "package_name", the name of the package, "repo_url", the repository's url, "branch_name", the name of the branch containing the code to download and "commit", the commit hash of interest.
 #' @return A character. The Nix definition to download and build the R package from Github.
 #' @noRd
-fetchgit <- function(git_pkg){
-
+fetchgit <- function(git_pkg) {
   package_name <- git_pkg$package_name
   repo_url <- git_pkg$repo_url
   branch_name <- git_pkg$branch_name
@@ -13,9 +12,10 @@ fetchgit <- function(git_pkg){
   sri_hash <- output$sri_hash
   imports <- output$deps
   imports <- unlist(strsplit(imports, split = " "))
-  imports <- paste(c("", imports), collapse = '\n          ')
+  imports <- paste(c("", imports), collapse = "\n          ")
 
-  sprintf('
+  sprintf(
+    '
     (pkgs.rPackages.buildRPackage {
       name = \"%s\";
       src = pkgs.fetchgit {
@@ -29,14 +29,13 @@ fetchgit <- function(git_pkg){
       };
     })
 ',
-  package_name,
-  repo_url,
-  branch_name,
-  commit,
-  sri_hash,
-  imports
-)
-
+    package_name,
+    repo_url,
+    branch_name,
+    commit,
+    sri_hash,
+    imports
+  )
 }
 
 
@@ -44,31 +43,32 @@ fetchgit <- function(git_pkg){
 #' @param archive_pkg A character of the form "dplyr@0.80"
 #' @return A character. The Nix definition to download and build the R package from CRAN.
 #' @noRd
-fetchzip <- function(archive_pkg, sri_hash = NULL){
-
+fetchzip <- function(archive_pkg, sri_hash = NULL) {
   pkgs <- unlist(strsplit(archive_pkg, split = "@"))
 
   cran_archive_link <- paste0(
     "https://cran.r-project.org/src/contrib/Archive/",
     pkgs[1], "/",
     paste0(pkgs[1], "_", pkgs[2]),
-    ".tar.gz")
+    ".tar.gz"
+  )
 
   package_name <- pkgs[1]
   repo_url <- cran_archive_link
 
-  if(is.null(sri_hash)){
+  if (is.null(sri_hash)) {
     output <- get_sri_hash_deps(repo_url, branch_name = NULL, commit = NULL)
     sri_hash <- output$sri_hash
     imports <- output$deps
     imports <- unlist(strsplit(imports, split = " "))
-    imports <- paste(c("", imports), collapse = '\n          ')
+    imports <- paste(c("", imports), collapse = "\n          ")
   } else {
     sri_hash <- sri_hash
     imports <- NULL
   }
 
-  sprintf('
+  sprintf(
+    '
     (pkgs.rPackages.buildRPackage {
       name = \"%s\";
       src = pkgs.fetchzip {
@@ -80,11 +80,11 @@ fetchzip <- function(archive_pkg, sri_hash = NULL){
       };
     })
 ',
-  package_name,
-  repo_url,
-  sri_hash,
-  imports
-)
+    package_name,
+    repo_url,
+    sri_hash,
+    imports
+  )
 }
 
 
@@ -93,14 +93,13 @@ fetchzip <- function(archive_pkg, sri_hash = NULL){
 #' @importFrom stats na.omit
 #' @return Atomic vector of packages without base packages
 #' @noRd
-remove_base <- function(list_imports){
-
[…]
