+++ R/with_nix.R
@@ -10,7 +10,7 @@
 #'
 #'
 #'
-#' `with_nix()` gives you the power of evaluating a main function `expr` 
+#' `with_nix()` gives you the power of evaluating a main function `expr`
 #' and its function call stack that are defined in the current R session
 #' in an encapsulated nix-R session defined by Nix expression (`default.nix`),
 #' which is located in at a distinct project path (`project_path`).
@@ -49,20 +49,20 @@
 #' well as the \{codetools\} package by Luke Tierney.
 #'
 #' Some of the key steps that are done behind the scene:
-#' 1. recursively find, classify, and export global objects (globals) in the 
+#' 1. recursively find, classify, and export global objects (globals) in the
 #' call stack of `expr` as well as propagate R package environments found.
 #' 2. Serialize (save to disk) and deserialize (read from disk) dependent
 #'  data structures as `.Rds` with necessary function arguments provided,
-#'  any relevant globals in the call stack, packages, and `expr` outputs 
+#'  any relevant globals in the call stack, packages, and `expr` outputs
 #'  returned in a temporary directory.
-#' 3. Use pure `nix-shell` environments to execute a R code script 
+#' 3. Use pure `nix-shell` environments to execute a R code script
 #'   reconstructed catching expressions with quoting; it is launched by commands
 #'  like this via `{sys}` by Jeroen Ooms:
 #'  `nix-shell --pure --run "Rscript --vanilla"`.
 #'
 #' @param expr Single R function or call, or character vector of length one with
 #' shell command and possibly options (flags) of the command to be invoked.
-#' For `program = R`, you can both use a named or an anonymous function. 
+#' For `program = R`, you can both use a named or an anonymous function.
 #' The function provided in `expr` should not evaluate when you pass arguments,
 #' hence you need to wrap your function call like
 #' `function() your_fun(arg_a = "a", arg_b = "b")`, to avoid evaluation and make
@@ -70,18 +70,18 @@
 #' @param program String stating where to evaluate the expression. Either `"R"`,
 #' the default, or `"shell"`. `where = "R"` will evaluate the expression via
 #' `RScript` and `where = "shell"` will run the system command in `nix-shell`.
-#' @param project_path Path to the folder where the `default.nix` file resides. 
+#' @param project_path Path to the folder where the `default.nix` file resides.
 #' The default is `"."`, which is the working directory in the current R
-#' session. This approach also useful when you have different subfolders 
+#' session. This approach also useful when you have different subfolders
 #' with separate software environments defined in different `default.nix` files.
-#' @param message_type String how detailed output is. Currently, there is 
+#' @param message_type String how detailed output is. Currently, there is
 #' either `"simple"` (default), `"quiet` or `"verbose"`, which shows the script
 #' that runs via `nix-shell`.
 #' @importFrom codetools findGlobals checkUsage
 #' @export
-#' @return 
+#' @return
 #' - if `program = "R"`, R object returned by function given in `expr`
-#' when evaluated via the R environment in `nix-shell` defined by Nix 
+#' when evaluated via the R environment in `nix-shell` defined by Nix
 #' expression.
 #' - if `program = "shell"`, list with the following elements:
 #'     - `status`: exit code
@@ -108,9 +108,9 @@
 #'   program = "R", project_path = project_path,
[…]
