### 2024-06-23 f45b911
update function docs
