+++ create_dev_env.R
@@ -4,7 +4,7 @@ latest_commit <- sys::as_text(sys::exec_internal("git", c("rev-parse", "master")
 
 
 rix(r_ver = "bleeding_edge",
-    r_pkgs = c("devtools", "diffviewer", "fledge", "lintr", "styler"
+    r_pkgs = c("devtools", "diffviewer", "fledge", "lintr", "styler",
                "codetools", "jsonlite",  "httr", "sys", "testthat", "knitr",
                "rmarkdown", "precommit"),
     system_pkgs = c("R", "glibcLocalesUtf8", "pandoc", "nix"),
