### 2024-06-24 b-rodrigues/rix#200 @eliocamp
`Error in available_r() : object 'r_nix_revs' not found` if the package is not loaded
`rix::rix:()` seems to fail if the package is not loaded:

``` r
path_default_nix <- tempdir()

rix::rix(r_ver = "latest",
    r_pkgs = c("dplyr", "ggplot2"),
    system_pkgs = NULL,
    git_pkgs = NULL,
    ide = "code",
    project_path = path_default_nix,
    overwrite = TRUE,
    print = TRUE)
#> Error in available_r(): object 'r_nix_revs' not found

library(rix)
rix::rix(r_ver = "latest",
         r_pkgs = c("dplyr", "ggplot2"),
         system_pkgs = NULL,
         git_pkgs = NULL,
         ide = "code",
         project_path = path_default_nix,
         overwrite = TRUE,
         print = TRUE)
#> # This file was generated by the {rix} R package v0.7.1 on 2024-06-24
#> # with following call:
#> # >rix::rix(r_ver = "9693852a2070b398ee123a329e68f0dab5526681",
#> #  > r_pkgs = c("dplyr",
#> #  > "ggplot2"),
#> #  > system_pkgs = NULL,
#> #  > git_pkgs = NULL,
#> #  > ide = "code",
#> #  > project_path = path_default_nix,
#> #  > overwrite = TRUE,
#> #  > print = TRUE)
#> # It uses nixpkgs' revision 9693852a2070b398ee123a329e68f0dab5526681 for reproducibility purposes
#> # which will install R version latest.
#> # Report any issues to https://github.com/b-rodrigues/rix
#> let
#>  pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/9693852a2070b398ee123a329e68f0dab5526681.tar.gz") {};
#>
#>   rpkgs = builtins.attrValues {
#>     inherit (pkgs.rPackages)
#>       dplyr
#>       ggplot2
#>       languageserver;
#>   };
#>
#>   system_packages = builtins.attrValues {
#>     inherit (pkgs)
#>       R
#>       glibcLocales
#>       nix;
#>   };
#>
#> in
#>
#> pkgs.mkShell {
#>   LOCALE_ARCHIVE = if pkgs.system == "x86_64-linux" then  "${pkgs.glibcLocales}/lib/locale/locale-archive" else "";
#>   LANG = "en_US.UTF-8";
#>    LC_ALL = "en_US.UTF-8";
#>    LC_TIME = "en_US.UTF-8";
#>    LC_MONETARY = "en_US.UTF-8";
#>    LC_PAPER = "en_US.UTF-8";
#>    LC_MEASUREMENT = "en_US.UTF-8";
#>
#>   buildInputs = [  rpkgs  system_packages   ];
#>
#> }
```

This seems to be an issue with `rix:::available_r()`.
https://github.com/b-rodrigues/rix/blob/86fac584c4cf16f1d6cc89745caf81088b8ae7f8/R/available_r.R#L14
