### 2024-06-24 b-rodrigues/rix#200 @philipp-baumann
Thanks for reporting @eliocamp . Can confirm the bug.

@b-rodrigues I think we best put it in `R/sysdata.rda`.

From `?data`

"...Use of data within a function without an envir argument has the almost always undesirable side-effect of putting an object in the user's workspace (and indeed, of replacing any object of that name already there). It would almost always be better to put the object in the current evaluation environment by data(..., envir = environment()). However, two alternatives are usually preferable, both described in the ‘Writing R Extensions’ manual.

For sets of data, set up a package to use lazy-loading of data.

For objects which are system data, for example lookup tables used in calculations within the function, use a file ‘R/sysdata.rda’ in the package sources or create the objects by R code at package installation time.

A sometimes important distinction is that the second approach places objects in the namespace but the first does not. So if it is important that the function sees mytable as an object from the package, it is system data and the second approach should be used. In the unusual case that a package uses a lazy-loaded dataset as a default argument to a function, that needs to be specified by [::](http://127.0.0.1:37983/help/library/utils/help/%3A%3A), e.g., survival::survexp.us.
"
