### 2024-06-24 7e7fbd0
updated man page for rix
