+++ tests/testthat/test-rix.R
@@ -184,6 +184,31 @@ testthat::test_that("If on darwin and ide = rstudio, raise warning", {
 
 })
 
+testthat::test_that("If R version is 4.4.0, raise warning", {
+
+  path_default_nix <- tempdir()
+
+  save_default_nix_test <- function(path_default_nix) {
+
+    rix(r_ver = "4.4.0",
+        ide = "other",
+        r_pkgs = NULL,
+        project_path = path_default_nix,
+        overwrite = TRUE,
+        shell_hook = NULL
+      )
+
+    paste0(path_default_nix, "/default.nix")
+
+  }
+
+  testthat::expect_warning(
+    save_default_nix_test(path_default_nix),
+    regexp = "version is not available"
+  )
+
+})
+
 
 testthat::test_that("rix(), bleeding_edge", {
 
