+++ R/included-datasets.R
@@ -1,12 +1,12 @@
-#' r_nix_revs
+#' sysdata
 #'
 #' Nix revisions with old R releases
 #'
-#' @format A data frame with 34 rows and 3 variables:
+#' @format A data frame with 38 rows and 3 variables:
 #' \describe{
 #'   \item{ version }{ R version }
 #'   \item{ revision }{ Nix revision containing this version of R }
 #'   \item{ date }{ Date when this revision was available }
 #' }
 #' @source Nix Package Version, by Marcelo Lazaroni
-"r_nix_revs"
+"sysdata"
