+++ R/included-datasets.R
@@ -2,7 +2,7 @@
 #'
 #' Nix revisions with old R releases
 #'
-#' @format A data frame with 38 rows and 3 variables:
+#' @format A data frame with 39 rows and 3 variables:
 #' \describe{
 #'   \item{ version }{ R version }
 #'   \item{ revision }{ Nix revision containing this version of R }
