+++ man/sysdata.Rd
@@ -1,11 +1,11 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/included-datasets.R
 \docType{data}
-\name{r_nix_revs}
-\alias{r_nix_revs}
-\title{r_nix_revs}
+\name{sysdata}
+\alias{sysdata}
+\title{sysdata}
 \format{
-A data frame with 34 rows and 3 variables:
+A data frame with 39 rows and 3 variables:
 \describe{
 \item{ version }{ R version }
 \item{ revision }{ Nix revision containing this version of R }
@@ -16,7 +16,7 @@ A data frame with 34 rows and 3 variables:
 Nix Package Version, by Marcelo Lazaroni
 }
 \usage{
-r_nix_revs
+sysdata
 }
 \description{
 Nix revisions with old R releases
