### 2024-06-24 2b3af57
removed unneeded vignette
