### 2024-06-24 4bb9afe
added trimws for good measure
