+++ R/fetchers.R
@@ -130,7 +130,7 @@ get_imports <- function(path) {
   # Remove version of R that may be listed in 'Depends'
   output <- Filter(function(x) !grepl("R \\(.*\\)", x), output)
 
-  output <- gsub("\\(.*?\\)", "", output)
+  output <- trimws(gsub("\\(.*?\\)", "", output))
 
   output <- remove_base(unique(output))
 
