+++ R/fetchers.R
@@ -130,6 +130,7 @@ get_imports <- function(path) {
   # Remove version of R that may be listed in 'Depends'
   output <- Filter(function(x) !grepl("R \\(.*\\)", x), output)
 
+  # Remove minimum package version for example 'packagename ( > 1.0.0)'
   output <- trimws(gsub("\\(.*?\\)", "", output))
 
   output <- remove_base(unique(output))
