### 2024-06-24 bec04f9
globals
