### 2024-06-24 e963d71
updated subshell vignette
