+++ NAMESPACE
@@ -18,3 +18,4 @@ importFrom(stats,na.omit)
 importFrom(tools,pskill)
 importFrom(utils,data)
 importFrom(utils,tail)
+importFrom(utils,untar)
