### 2024-06-24 02ece7d
updates to docker documentation
