### 2024-06-24 b-rodrigues/rix#200 @eliocamp
With the latest master, I get a different object that is not found 😅

``` r
rix::rix(r_ver = "latest",
    r_pkgs = NULL,
    system_pkgs = NULL,
    git_pkgs = NULL,
    ide = "rstudio",
    project_path = "../nixtest2/",
    overwrite = TRUE,
    print = TRUE)
#> Error in available_r(): object 'sysdata' not found
```
