### 2024-06-24 b-rodrigues/rix#200 @philipp-baumann
Bruno has just pushed a fix:

```r
devtools::install_github("https://github.com/b-rodrigues, ref = "simplify_available_r")
```

This should work.
