+++ R/get_latest.R
@@ -35,15 +35,8 @@ get_right_commit <- function(r_version) {
   if (r_version == "frozen_edge") {
     api_url <- "https://api.github.com/repos/rstats-on-nix/nixpkgs/commits?sha=r-daily"
   } else if (r_version %in% Filter(function(x) `!=`(x, "latest"), available_r())) { # all but latest
-    temp <- new.env(parent = emptyenv())
 
-    data(list = "sysdata",
-         package = "rix",
-         envir = temp)
-
-    get("sysdata", envir = temp)
-
-    return(sysdata$revision[sysdata$version == r_version])
+    return(rix::sysdata$revision[sysdata$version == r_version])
 
   } else {
     api_url <- "https://api.github.com/repos/NixOS/nixpkgs/commits?sha=nixpkgs-unstable"
