### 2024-06-24 b-rodrigues/rix#200 @philipp-baumann
https://github.com/b-rodrigues/rix/pull/220
