+++ R/get_latest.R
@@ -36,7 +36,7 @@ get_right_commit <- function(r_version) {
     api_url <- "https://api.github.com/repos/rstats-on-nix/nixpkgs/commits?sha=r-daily"
   } else if (r_version %in% Filter(function(x) `!=`(x, "latest"), available_r())) { # all but latest
 
-    return(rix::sysdata$revision[rix::sysdata$version == r_version])
+    return(sysdata$revision[sysdata$version == r_version])
 
   } else {
     api_url <- "https://api.github.com/repos/NixOS/nixpkgs/commits?sha=nixpkgs-unstable"
