### 2024-06-25 5dc72d3
hash_git working, but not hash_archive
