+++ R/nix_hash.R
@@ -1,11 +1,11 @@
-#' Return the sri hash of a path using `nix hash path --sri path`
+#' Return the sri hash of a path using `nix hash path --sri path` if Nix is available locally
 #' @param repo_url URL to Github repository
-#' @param branchName Branch to checkout
+#' @param branch_name Branch to checkout
 #' @param commit Commit hash
-nix_hash <- function(repo_url, branchName, commit) {
+nix_hash <- function(repo_url, branch_name, commit) {
 
   if(grepl("github", repo_url)){
-    hash_git(repo_url, branchName, commit)
+    hash_git(repo_url, branch_name, commit)
   } else if(grepl("cran.*Archive.*", repo_url)){
     hash_cran(repo_url)
   } else {
@@ -39,7 +39,10 @@ hash_cran <- function(repo_url){
 
   sri_hash <- system(command, intern = TRUE)
 
-  deps <- get_imports(paste0(path_to_src, "/DESCRIPTION"))
+  paths <- list.files(path_to_src, full.names = TRUE, recursive = TRUE)
+  desc_path <- grep("DESCRIPTION", paths, value = TRUE)
+
+  deps <- get_imports(desc_path)
 
   unlink(path_to_folder, recursive = TRUE, force = TRUE)
 
@@ -53,10 +56,10 @@ hash_cran <- function(repo_url){
 
 #' Return the sri hash of a Github repository
 #' @param repo_url URL to Github repository
-#' @param branchName Branch to checkout
+#' @param branch_name Branch to checkout
 #' @param commit Commit hash
 #' @importFrom git2r clone checkout
-hash_git <- function(repo_url, branchName, commit){
+hash_git <- function(repo_url, branch_name, commit){
 
   path_to_repo <- paste0(tempdir(), "repo",
                          paste0(sample(letters, 5), collapse = ""))
@@ -64,7 +67,8 @@ hash_git <- function(repo_url, branchName, commit){
   git2r::clone(
            url = repo_url,
            local_path = path_to_repo,
-           branch = branchName
+           branch = branch_name,
+           progress = FALSE
          )
 
   git2r::checkout(path_to_repo, branch = commit)
@@ -87,4 +91,61 @@ hash_git <- function(repo_url, branchName, commit){
       "sri_hash" = sri_hash,
       "deps" = deps)
   )
+}
+
+
+#' Get the SRI hash of the NAR serialization of a Github repo, if nix is not available locally
+#' @param repo_url A character. The URL to the package's Github repository or to the `.tar.gz` package hosted on CRAN.
+#' @param branch_name A character. The branch of interest, NULL for archived CRAN packages.
+#' @param commit A character. The commit hash of interest, for reproducibility's sake, NULL for archived CRAN packages.
+#' @return list with following elements:
+#' - `sri_hash`: string with SRI hash of the NAR serialization of a Github repo
+#' - `deps`: string with R package dependencies separarated by space.
+#' @noRd
+nix_hash_online <- function(repo_url, branch_name, commit) {
+  # handle to get error for status code 404
+  h <- curl::new_handle(failonerror = TRUE)
+
+  url <- paste0(
+    "http://git2nixsha.dev:1506/hash?repo_url=",
+    repo_url, "&branchName=", branch_name, "&commit=", commit
+  )
+
+  # extra diagnostics
+  extra_diagnostics <-
+    c(
+      "\nIf it's a Github repo, check the url, branch name and commit.\n",
+      "Are these correct? If it's an archived CRAN package, check the name\n",
+      "of the package and the version number."
+    )
+
+  req <- try_get_request(
+    url = url, handle = h,
+    extra_diagnostics = extra_diagnostics
+  )
+
+  # plumber endpoint delivers list with
[…]
