### 2024-06-26 4b90c80
relevant extra diagnostics
