### 2024-06-27 7fc0e31
Update README.Rmd
fixes #225
