### 2024-06-27 c81fb5d
Update README.Rmd
fixes #224
