+++ README.Rmd
@@ -50,20 +50,39 @@ packages and manage configurations in a declarative way. This makes sure that
 software gets installed in fully reproducible manner, on any operating system or
 hardware.
 
-`{rix}` provides functions to help you write and deploy Nix expressions (written
+`{rix}` provides functions to help you write Nix expressions (written
 in the Nix language). These expressions will be the inputs for the Nix package
 manager, to build sets of software packages and provide them in a reproducible
 development environment. These environments can be used for interactive data
-analysis, or reproduced when running pipelines in CI/CD systems. On the 
-[Nixpkgs collection](https://github.com/nixos/nixpkgs), there are currently more
+analysis, or reproduced when running pipelines in CI/CD systems. The 
+[Nixpkgs collection](https://github.com/nixos/nixpkgs) includes currently more
 than 100.000 pieces of software available through the Nix package manager. 
-Through {rix}, you can define and build isolated R environments through Nix
-package manager with ease. Like this, environments contain R and all the
-required packages that you need for your project. You can also add any other
-software tool available. The Nix R ecosystem currently includes the entirety of
-CRAN and Bioconductor packages. Like with any other programming language and
+
+With `{rix}`, you can define development environments, or shells, that contain
+the required tools needed to analyze data using R. These environments are
+isolated from each other and project-specific: this means that a project can use one version 
+of R and R packages, and another environment another version of R and R packages. 
+However, extra care is required if you already have R installed through the usual 
+method for your operating system, as
+these development environments are not totally isolated from the rest of your system.
+
+Unlike Docker, where a running container cannot acces anything from the host system, unless
+explicitely configured, Nix development shells are nothing but environments that 
+add more software to the list of already available software (the so-called `PATH`). As
+such, it is possible to access anything (files but also software) already present on the 
+system from a running Nix shell. Thus, `{rix}` also provides a function called `rix_init()`
+that helps isolate R sessions running inside Nix environments from the rest of your system. This avoids
+clashes between the Nix-specific library of R packages and the user library of R packages
+should you already have R installed and managed by the usual method for your operating system.
+
+It is also possible to add any other
+software package available on Nixpkgs, for example IDEs such as RStudio or VS Code. 
+The Nix R ecosystem currently includes almost the entirety of
+CRAN and Bioconductor packages (there is around a hundred CRAN or Biocondcuctor packages 
+that are unavailable through Nix). Like with any other programming language and
 software, it is also possible to install older releases of R packages, or
-install packages from GitHub at defined states.
+install packages from GitHub at defined states, as well as local packages 
+in the `.tar.gz` format.
 
 The Nix package manager is extremely powerful. Not only does it handle all the
 dependencies of any package extremely well in a deterministic manner, it is also
