### 2024-06-27 6751f2b
Update d1-installing-r-packages-in-a-nix-environment.Rmd
fixes #223
