+++ R/fetchers.R
@@ -1,6 +1,10 @@
 #' fetchgit Downloads and installs a package hosted of Git
-#' @param git_pkg A list of four elements: "package_name", the name of the package, "repo_url", the repository's url, "branch_name", the name of the branch containing the code to download and "commit", the commit hash of interest.
-#' @return A character. The Nix definition to download and build the R package from Github.
+#' @param git_pkg A list of four elements: "package_name", the name of the
+#' package, "repo_url", the repository's url, "branch_name", the name of the
+#' branch containing the code to download and "commit", the commit hash of
+#' interest.
+#' @return A character. The Nix definition to download and build the R package
+#' from Github.
 #' @noRd
 fetchgit <- function(git_pkg) {
   package_name <- git_pkg$package_name
@@ -40,7 +44,7 @@ fetchgit <- function(git_pkg) {
 
 
 #' fetchzip Downloads and installs an archived CRAN package
-#' @param archive_pkg A character of the form "dplyr@0.80"
+#' @param archive_pkg A character of the form `"dplyr@0.80"`
 #' @return A character. The Nix definition to download and build the R package from CRAN.
 #' @noRd
 fetchzip <- function(archive_pkg, sri_hash = NULL) {
@@ -112,25 +116,18 @@ remove_base <- function(list_imports) {
 #' @return Atomic vector of packages
 #' @noRd
 get_imports <- function(path) {
-
   tmp_dir <- tempdir()
 
   # Is the path pointing to a tar.gz archive
   # or directly to a DESCRIPTION file?
-  if(grepl("\\.tar\\.gz", path)){
-
+  if (grepl("\\.tar\\.gz", path)) {
     untar(path, exdir = tmp_dir)
     paths <- list.files(tmp_dir, full.names = TRUE, recursive = TRUE)
     desc_path <- grep("DESCRIPTION", paths, value = TRUE)
-
-  } else if(grepl("DESCRIPTION", path)){
-
+  } else if (grepl("DESCRIPTION", path)) {
     desc_path <- path
-
   } else {
-
     stop("Path is neither a .tar.gz archive, nor pointing to a DESCRIPTION file directly.")
-
   }
 
   columns_of_interest <- c("Depends", "Imports", "LinkingTo")
@@ -145,10 +142,10 @@ get_imports <- function(path) {
 
   output <- unname(trimws(unlist(strsplit(unlist(imports), split = ","))))
 
-                                        # Remove version of R that may be listed in 'Depends'
+  # Remove version of R that may be listed in 'Depends'
   output <- Filter(function(x) !grepl("R \\(.*\\)", x), output)
 
-                                        # Remove minimum package version for example 'packagename ( > 1.0.0)'
+  # Remove minimum package version for example 'packagename ( > 1.0.0)'
   output <- trimws(gsub("\\(.*?\\)", "", output))
 
   output <- remove_base(unique(output))
@@ -158,7 +155,9 @@ get_imports <- function(path) {
 
 
 #' fetchlocal Installs a local R package
-#' @param local_pkg A list of local package names ('.tar.gz' archives) to install. These packages need to be in the same folder as the generated `default.nix` file.
+#' @param local_pkg A list of local package names ('.tar.gz' archives) to
+#' install. These packages need to be in the same folder as the generated
+#' `default.nix` file.
 #' @importFrom utils tail
 #' @return A character. The Nix definition to build the R package from local sources.
 #' @noRd
@@ -190,8 +189,10 @@ fetchlocal <- function(local_pkg) {
 }
 
 #' fetchlocals Installs a local R package
-#' @param local_pkgs Either a list of paths to local packages, or a path to a single package
-#' @return A character. The Nix definition to build the local R packages from local sources.
+#' @param local_pkgs Either a list of paths to local packages, or a path to a
+#' single package
[…]
