+++ README.Rmd
@@ -34,55 +34,71 @@ knitr::opts_chunk$set(
 
 ## Introduction
 
-`{rix}` is an R package that leverages [Nix](https://nixos.org/),
-a powerful package manager focusing on reproducible builds. With Nix, it is
-possible to create project-specific environments that contain a project-specific
-version of R and R packages (as well as other tools or languages, if needed).
-You can use `{rix}` and Nix to replace `{renv}` and Docker with one single tool.
-Nix is an incredibly useful piece of software for ensuring reproducibility of
-projects, in research or otherwise. For example, it allows you run web 
-applications like Shiny apps or `{plumber}` APIs in a controlled environment.
+`{rix}` is an R package that leverages [Nix](https://nixos.org/), a powerful
+package manager focusing on reproducible builds. With Nix, it is possible to
+create project-specific environments that contain a project-specific version of
+R and R packages (as well as other tools or languages, if needed). This
+project-specific environment will also include all the required system-level
+dependencies that can be difficult to install, such as `GDAL` for packages for
+geospatial analysis for example. This is how Nix installs software: it installs
+software as a complete "bundle" that include all of its dependencies, and all of
+the dependencies' dependencies and so on. Nix is an incredibly useful piece of
+software for ensuring reproducibility of projects, in research or otherwise. For
+example, it allows you run web applications like Shiny apps or `{plumber}` APIs
+in a controlled environment, or run `{targets}` pipelines with the right version
+of R and dependencies, and it is also possible to use environments managed by Nix
+to work interactively using an IDE.
+
+In essence, this means that you can use `{rix}` and Nix to replace `{renv}` and
+Docker with one single tool, but the approach is quite different: `{renv}`
+records specific versions of individual packages, while `{rix}` provides a
+complete snapshot of the R ecosystem at a specific point in time, but also
+snapshots all the required dependencies to make your project-specific R
+environment work. To ensure complete reproducibility with `{renv}`, it must be
+combined with Docker, in order to include system-level dependencies (like `GDAL`,
+as per the example above).
 
 Nix has a fairly high entry cost though. Nix is a complex piece of software that
 comes with its own programming language, which is also called Nix. Its purpose
 is to solve a complex problem: defining instructions on how to build software
-packages and manage configurations in a declarative way. This makes sure that 
-software gets installed in fully reproducible manner, on any operating system or
-hardware.
+packages and manage configurations in a declarative and way. This makes sure
+that software gets installed in fully reproducible manner, on any operating
+system or hardware.
 
-`{rix}` provides functions to help you write Nix expressions (written
-in the Nix language). These expressions will be the inputs for the Nix package
-manager, to build sets of software packages and provide them in a reproducible
-development environment. These environments can be used for interactive data
-analysis, or reproduced when running pipelines in CI/CD systems. The 
-[Nixpkgs collection](https://github.com/nixos/nixpkgs) includes currently more
-than 100.000 pieces of software available through the Nix package manager. 
+`{rix}` provides functions to help you write Nix expressions (written in the Nix
+language). These expressions will be the inputs for the Nix package manager, to
+build sets of software packages and provide them in a reproducible development
+environment. These environments can be used for interactive data analysis, or
+reproduced when running pipelines in CI/CD systems. The [Nixpkgs
+collection](https://github.com/nixos/nixpkgs) includes currently more than
+100.000 pieces of software available through the Nix package manager.
 
 With `{rix}`, you can define development environments, or shells, that contain
 the required tools needed to analyze data using R. These environments are
-isolated from each other and project-specific: this means that a project can use one version 
-of R and R packages, and another environment another version of R and R packages. 
-However, extra care is required if you already have R installed through the usual 
-method for your operating system, as
[…]
