### 2024-06-28 b2d4abe
warn users that generate an expression for an old version of R
