### 2024-06-28 262a242
rix_init_after_rix
