### 2024-06-28 2f8c143
better warning if default.nix exists already
