+++ R/with_nix.R
@@ -38,11 +38,10 @@
 #' can easily do so by providing Nix expressions in custom `.nix` or
 #' `default.nix` files in different subfolders of the project.
 #'
-#' It is recommended that you use `rix_init()` to generate a custom `.Rprofile`
+#' `rix_init()` is run automatically to generate a custom `.Rprofile`
 #' file for the subshell in `project_dir`. The defaults in that file ensure
 #' that only R packages from the Nix store, that are defined in the subshell
 #' `.nix` file are loaded and system's libraries are excluded.
-#'
 #'
 #' To do its job, `with_nix()` heavily relies on patterns that manipulate
 #' language expressions (aka computing on the language) offered in base R as
