+++ vignettes/c-using-rix-to-build-project-specific-environments.Rmd
@@ -163,7 +163,6 @@ version 4.0.0 for example in a subshell (or execute a function that requires an
 old version of a package in that subshell), and get the result of the
 computation back into the main R session.
 
-
 ## Nix environments are not completely isolated from your system
 
 It is important to know that an environment built by Nix is not totally isolated
@@ -172,8 +171,20 @@ your system, and suppose you build a Nix environment that also comes with `sl`.
 If you activate that environment, the version of `sl` that will run when called
 is the one included in the Nix environment. If, however, you start `sl` in a Nix
 environment that does not come with it, then your system's `sl` will get used
-instead. It is also possible to completely isolate an environment built with Nix
-using the provided `rix_init()` function and activate your environment using
-`nix-shell --pure` instead of only `nix-shell`. This is especially useful for
-subshells.
+instead.
+
+In the context of the R programming language, this means that if you have a user
+or system library of packages (meaning, a library of packages generated by a
+regular installation of R), these packages may be loaded by an R version running
+from a Nix shell. To avoid issues with this, calling `rix()` automatically runs
+`rix_init()` as well, which generates a custom `.Rprofile` file in the project's
+directory. This way, your regular user library of packages will not interfere
+with R environments managed by Nix. Moreover, this custom `.Rprofile` also
+redefines `install.packages()` and makes it throw an error: indeed, if you wish
+to add packages to your an R environment managed by Nix, you should add these
+packages to the `default.nix` file instead, and rebuild the environment.
+
+If you want to even make other programs inaccessible to a running Nix
+environment, you can drop into it using `nix-shell --pure` instead of only
+`nix-shell`.
 

[…]
