### 2024-06-28 6a3b138
implement github repo and cran hashing
