### 2024-06-28 53e180d
get rid of git2r
