+++ .github/workflows/tests-r-via-nix.yaml
@@ -32,7 +32,7 @@ jobs:
           name: rstats-on-nix
       
       - name: devtools::test() via nix-shell
-        run: nix-shell --pure -p cacert nix R rPackages.git2r rPackages.covr rPackages.codetools rPackages.xml2 rPackages.sys rPackages.devtools --run "Rscript -e 'devtools::test()'"
+        run: nix-shell --pure -p cacert nix R rPackages.covr rPackages.codetools rPackages.xml2 rPackages.sys rPackages.devtools --run "Rscript -e 'devtools::test()'"
       #- name: Test coverage
       #  run: nix-shell --pure -p nix git cacert R rPackages.covr rPackages.codetools rPackages.sys rPackages.xml2 rPackages.testthat --run "Rscript -e 'covr::codecov()' "
 
