+++ tests/testthat/test-rix.R
@@ -1,31 +1,34 @@
 testthat::test_that("rix(), ide is 'rstudio', Linux", {
-
   skip_if(Sys.info()["sysname"] == "Darwin")
 
   path_default_nix <- tempdir()
 
   save_default_nix_test <- function(ide, path_default_nix) {
-
-    rix(r_ver = "4.3.1",
-        r_pkgs = c("dplyr", "janitor", "AER@1.2-8", "quarto"),
-        tex_pkgs = c("amsmath"),
-        git_pkgs = list(
-          list(package_name = "housing",
-               repo_url = "https://github.com/rap4all/housing/",
-               branch_name = "fusen",
-               commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"),
-          list(package_name = "fusen",
-               repo_url = "https://github.com/ThinkR-open/fusen",
-               branch_name = "main",
-               commit = "d617172447d2947efb20ad6a4463742b8a5d79dc")
+    rix(
+      r_ver = "4.3.1",
+      r_pkgs = c("dplyr", "janitor", "AER@1.2-8", "quarto"),
+      tex_pkgs = c("amsmath"),
+      git_pkgs = list(
+        list(
+          package_name = "housing",
+          repo_url = "https://github.com/rap4all/housing/",
+          branch_name = "fusen",
+          commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"
         ),
-        ide = ide,
-        project_path = path_default_nix,
-        overwrite = TRUE,
-        shell_hook = NULL)
+        list(
+          package_name = "fusen",
+          repo_url = "https://github.com/ThinkR-open/fusen",
+          branch_name = "main",
+          commit = "d617172447d2947efb20ad6a4463742b8a5d79dc"
+        )
+      ),
+      ide = ide,
+      project_path = path_default_nix,
+      overwrite = TRUE,
+      shell_hook = NULL
+    )
 
     paste0(path_default_nix, "/default.nix")
-
   }
 
   testthat::announce_snapshot_file("rix/rstudio_default.nix")
@@ -34,260 +37,268 @@ testthat::test_that("rix(), ide is 'rstudio', Linux", {
     path = save_default_nix_test(ide = "rstudio", path_default_nix),
     name = "rstudio_default.nix",
   )
-
-
 })
 
 
 testthat::test_that("rix(), ide is 'other' or 'code'", {
-
-  path_default_nix <- normalizePath(tempdir())
+  path_default_nix <- paste0(
+    tempdir(), paste0(sample(letters, 5), collapse = "")
+  )
+  dir.create(path_default_nix)
+  path_default_nix <- normalizePath(path_default_nix)
 
   save_default_nix_test <- function(ide, path_default_nix) {
-
-    rix(r_ver = "4.3.1",
-        r_pkgs = c("data.table", "janitor", "AER@1.2-8", "quarto"),
-        tex_pkgs = c("amsmath"),
-        git_pkgs = list(
-          list(package_name = "housing",
-               repo_url = "https://github.com/rap4all/housing/",
-               branch_name = "fusen",
-               commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"),
-          list(package_name = "fusen",
-               repo_url = "https://github.com/ThinkR-open/fusen",
-               branch_name = "main",
-               commit = "d617172447d2947efb20ad6a4463742b8a5d79dc")
+    rix(
+      r_ver = "4.3.1",
+      r_pkgs = c("data.table", "janitor", "AER@1.2-8", "quarto"),
+      tex_pkgs = c("amsmath"),
+      git_pkgs = list(
+        list(
+          package_name = "housing",
+          repo_url = "https://github.com/rap4all/housing/",
+          branch_name = "fusen",
[…]
