### 2024-06-30 65e1b87
expand docs; shorter nix-hash command for sri base64 hash
