+++ man/hash_cran.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nix_hash.R
+\name{hash_cran}
+\alias{hash_cran}
+\title{Return the SRI hash of a CRAN package source using \verb{nix hash path --sri path}}
+\usage{
+hash_cran(repo_url)
+}
+\arguments{
+\item{url}{URL to CRAN package source}
+}
+\description{
+Return the SRI hash of a CRAN package source using \verb{nix hash path --sri path}
+}
