+++ man/nix_hash.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nix_hash.R
+\name{nix_hash}
+\alias{nix_hash}
+\title{Return the sri hash of a path using \verb{nix hash path --sri path} if Nix is
+available locally}
+\usage{
+nix_hash(repo_url, branch_name, commit)
+}
+\arguments{
+\item{repo_url}{URL to Github repository}
+
+\item{branch_name}{Branch to checkout}
+
+\item{commit}{Commit hash}
+}
+\description{
+Return the sri hash of a path using \verb{nix hash path --sri path} if Nix is
+available locally
+}
