+++ man/hash_git.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nix_hash.R
+\name{hash_git}
+\alias{hash_git}
+\title{Return the SRI hash of a GitHub repository at a given unique commmit ID}
+\usage{
+hash_git(repo_url, branch_name, commit)
+}
+\arguments{
+\item{repo_url}{URL to GitHub repository}
+
+\item{commit}{Commit hash}
+}
+\description{
+Return the SRI hash of a GitHub repository at a given unique commmit ID
+}
+\details{
+\code{hash_git} will retrieve an archive of the repository URL
+<https://github.com/\if{html}{\out{<user>}}/\if{html}{\out{<repo>}} at a given commit ID. It will fetch
+a .tar.gz file from
+<https://github.com/\if{html}{\out{<user>}}/\if{html}{\out{<repo>}}/archive/\if{html}{\out{<commit-id>}}.tar.gz. Then, it will
+ungzip and unarchive the downloaded \code{tar.gz} file. Then, on the extracted
+directory, it will run \code{nix-hash}
+(NAR) hash
+NAR
+}
