### 2024-06-30 201aa35
do more checking for `options(rix.sri_hash=`)
