+++ man/get_sri_hash_deps.Rd
@@ -1,0 +1,33 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/nix_hash.R
+\name{get_sri_hash_deps}
+\alias{get_sri_hash_deps}
+\title{Return the sri hash of a path using \verb{nix-hash --type sha256 --sri <path>}
+with local Nix, or using an online API service (equivalent
+\verb{nix hash path --sri <path>}) if Nix is not available}
+\usage{
+get_sri_hash_deps(repo_url, branch_name, commit)
+}
+\arguments{
+\item{repo_url}{A character. The URL to the package's Github repository or to
+the \code{.tar.gz} package hosted on CRAN.}
+
+\item{branch_name}{A character. The branch of interest, NULL for archived
+CRAN packages.}
+
+\item{commit}{A character. The commit hash of interest, for reproducibility's
+sake, NULL for archived CRAN packages.}
+}
+\value{
+list with following elements:
+\itemize{
+\item \code{sri_hash}: string with SRI hash of the NAR serialization of a Github repo
+at a given deterministic git commit ID (SHA-1)
+\item \code{deps}: string with R package dependencies separarated by space.
+}
+}
+\description{
+Return the sri hash of a path using \verb{nix-hash --type sha256 --sri <path>}
+with local Nix, or using an online API service (equivalent
+\verb{nix hash path --sri <path>}) if Nix is not available
+}
