### 2024-06-30 afcf1c4
fix tests; make sure defaults
