### 2024-07-01 b876211
change test to make it pass, probably missed before because of skip_on_darwin
