### 2024-07-01 23be343
handle situations where two description files are present in the package (one for testing)
