+++ vignettes/d1-installing-r-packages-in-a-nix-environment.Rmd
@@ -78,12 +78,8 @@ and your system library of packages don't interfere with each other.
 It is possible to install an arbitrary version of a package that has been
 archived on CRAN:
 
-
-```{r parsermd-chunk-1}
-path_default_nix <- paste0(
-  tempdir(), "repo",
-  paste0(sample(letters, 5), collapse = "")
-)
+```{r, eval = F}
+path_default_nix <- tempdir()
 
 rix(r_ver = "4.2.1",
     r_pkgs = c("dplyr@0.8.0", "janitor@1.0.0"),
@@ -94,8 +90,90 @@ rix(r_ver = "4.2.1",
     overwrite = TRUE)
 ```
 
-```{r parsermd-chunk-2, echo = F}
-cat(readLines(paste0(path_default_nix, "/default.nix")), sep = "\n")
+```
+#> # This file was generated by the {rix} R package v0.7.1 on 2024-07-01
+#> # with following call:
+#> # >rix(r_ver = "79b3d4bcae8c7007c9fd51c279a8a67acfa73a2a",
+#> #  > r_pkgs = c("dplyr@0.8.0",
+#> #  > "janitor@1.0.0"),
+#> #  > system_pkgs = NULL,
+#> #  > git_pkgs = NULL,
+#> #  > ide = "other",
+#> #  > project_path = path_default_nix,
+#> #  > overwrite = TRUE)
+#> # It uses nixpkgs' revision 79b3d4bcae8c7007c9fd51c279a8a67acfa73a2a for reproducibility purposes
+#> # which will install R version 4.2.1.
+#> # Report any issues to https://github.com/b-rodrigues/rix
+#> let
+#>  pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/79b3d4bcae8c7007c9fd51c279a8a67acfa73a2a.tar.gz") {};
+#>   
+#>   git_archive_pkgs = [ 
+#>     (pkgs.rPackages.buildRPackage {
+#>       name = "dplyr";
+#>       src = pkgs.fetchzip {
+#>        url = "https://cran.r-project.org/src/contrib/Archive/dplyr/dplyr_0.8.0.tar.gz";
+#>        sha256 = "sha256-f30raalLd9KoZKZSxeTN71PG6BczXRIiP6g7EZeH09U=";
+#>       };
+#>       propagatedBuildInputs = builtins.attrValues {
+#>         inherit (pkgs.rPackages) 
+#>           assertthat
+#>           glue
+#>           magrittr
+#>           pkgconfig
+#>           R6
+#>           Rcpp
+#>           rlang
+#>           tibble
+#>           tidyselect
+#>           BH
+#>           plogr;
+#>       };
+#>     })
+#> 
+#> 
+#>     (pkgs.rPackages.buildRPackage {
+#>       name = "janitor";
+#>       src = pkgs.fetchzip {
+#>        url = "https://cran.r-project.org/src/contrib/Archive/janitor/janitor_1.0.0.tar.gz";
+#>        sha256 = "sha256-3NJomE/CNbOZ+ohuVZJWe2n1RVGUm8x8a0A0qzLmdN4=";
+#>       };
+#>       propagatedBuildInputs = builtins.attrValues {
+#>         inherit (pkgs.rPackages) 
+#>           dplyr
+#>           tidyr
+#>           snakecase
+#>           magrittr
+#>           purrr
+#>           rlang;
+#>       };
+#>     })
+#>   ];
+#>    
+#>   system_packages = builtins.attrValues {
+#>     inherit (pkgs) 
+#>       R
+#>       glibcLocales
+#>       nix;
+#>   };
+#>   
+#> in
+#> 
+#> pkgs.mkShell {
[…]
