+++ vignettes/c-using-rix-to-build-project-specific-environments.Rmd
@@ -50,7 +50,7 @@ file. You need to provide the following inputs to `rix()`:
 Run the following command to generate the right `default.nix` file:
 
 
-```{r parsermd-chunk-1}
+```{r, eval = F}
 path_default_nix <- tempdir()
 
 rix(r_ver = "latest",
@@ -61,6 +61,51 @@ rix(r_ver = "latest",
     project_path = path_default_nix,
     overwrite = TRUE,
     print = TRUE)
+#> # This file was generated by the {rix} R package v0.7.1 on 2024-07-01
+#> # with following call:
+#> # >rix(r_ver = "12a9c0004bc987afb1ff511ebb97b67497a68e22",
+#> #  > r_pkgs = c("dplyr",
+#> #  > "ggplot2"),
+#> #  > system_pkgs = NULL,
+#> #  > git_pkgs = NULL,
+#> #  > ide = "code",
+#> #  > project_path = path_default_nix,
+#> #  > overwrite = TRUE,
+#> #  > print = TRUE)
+#> # It uses nixpkgs' revision 12a9c0004bc987afb1ff511ebb97b67497a68e22 for reproducibility purposes
+#> # which will install R version latest.
+#> # Report any issues to https://github.com/b-rodrigues/rix
+#> let
+#>  pkgs = import (fetchTarball "https://github.com/NixOS/nixpkgs/archive/12a9c0004bc987afb1ff511ebb97b67497a68e22.tar.gz") {};
+#>  
+#>   rpkgs = builtins.attrValues {
+#>     inherit (pkgs.rPackages) 
+#>       dplyr
+#>       ggplot2
+#>       languageserver;
+#>   };
+#>     
+#>   system_packages = builtins.attrValues {
+#>     inherit (pkgs) 
+#>       R
+#>       glibcLocales
+#>       nix;
+#>   };
+#>   
+#> in
+#> 
+#> pkgs.mkShell {
+#>   LOCALE_ARCHIVE = if pkgs.system == "x86_64-linux" then  "${pkgs.glibcLocales}/lib/locale/locale-archive" else "";
+#>   LANG = "en_US.UTF-8";
+#>    LC_ALL = "en_US.UTF-8";
+#>    LC_TIME = "en_US.UTF-8";
+#>    LC_MONETARY = "en_US.UTF-8";
+#>    LC_PAPER = "en_US.UTF-8";
+#>    LC_MEASUREMENT = "en_US.UTF-8";
+#> 
+#>   buildInputs = [  rpkgs  system_packages   ];
+#>   
+#> }
 ```
 
 To start using this environment, open a terminal in the folder containing
@@ -114,9 +159,51 @@ rix(r_ver = "4.2.1",
 
 The file looks like this:
 
-
-```{r parsermd-chunk-3, echo = F}
-cat(readLines(paste0(path_default_nix, "/default.nix")), sep = "\n")
+```
+#> # This file was generated by the {rix} R package v0.7.1 on 2024-07-01
+#> # with following call:
+#> # >rix(r_ver = "79b3d4bcae8c7007c9fd51c279a8a67acfa73a2a",
+#> #  > r_pkgs = c("dplyr",
+#> #  > "janitor"),
+#> #  > system_pkgs = c("quarto"),
+#> #  > git_pkgs = NULL,
+#> #  > ide = "other",
+#> #  > project_path = path_default_nix,
+#> #  > overwrite = TRUE)
+#> # It uses nixpkgs' revision 79b3d4bcae8c7007c9fd51c279a8a67acfa73a2a for reproducibility purposes
+#> # which will install R version 4.2.1.
[…]
