### 2024-07-01 2b62be4
check for existence of .Rprofile
