+++ tests/testthat/test-rix.R
@@ -56,28 +56,33 @@ testthat::test_that("rix(), ide is 'other' or 'code'", {
   path_default_nix <- normalizePath(path_default_nix)
 
   save_default_nix_test <- function(ide, path_default_nix) {
-    rix(
-      r_ver = "4.3.1",
-      r_pkgs = c("data.table", "janitor", "AER@1.2-8", "quarto"),
-      tex_pkgs = c("amsmath"),
-      git_pkgs = list(
-        list(
-          package_name = "housing",
-          repo_url = "https://github.com/rap4all/housing/",
-          branch_name = "fusen",
-          commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"
+    # We need to add this because this function gets called
+    # twice, so the generated .Rprofile is there already and
+    # calling the function again raises the warning.
+    suppressWarnings(
+      rix(
+        r_ver = "4.3.1",
+        r_pkgs = c("data.table", "janitor", "AER@1.2-8", "quarto"),
+        tex_pkgs = c("amsmath"),
+        git_pkgs = list(
+          list(
+            package_name = "housing",
+            repo_url = "https://github.com/rap4all/housing/",
+            branch_name = "fusen",
+            commit = "1c860959310b80e67c41f7bbdc3e84cef00df18e"
+          ),
+          list(
+            package_name = "fusen",
+            repo_url = "https://github.com/ThinkR-open/fusen",
+            branch_name = "main",
+            commit = "d617172447d2947efb20ad6a4463742b8a5d79dc"
+          )
         ),
-        list(
-          package_name = "fusen",
-          repo_url = "https://github.com/ThinkR-open/fusen",
-          branch_name = "main",
-          commit = "d617172447d2947efb20ad6a4463742b8a5d79dc"
-        )
-      ),
-      ide = ide,
-      project_path = path_default_nix,
-      overwrite = TRUE,
-      shell_hook = NULL
+        ide = ide,
+        project_path = path_default_nix,
+        overwrite = TRUE,
+        shell_hook = NULL
+      )
     )
 
     file.path(path_default_nix, "default.nix")
@@ -114,13 +119,16 @@ testthat::test_that("Quarto gets added to sys packages", {
   path_default_nix <- normalizePath(tempdir())
 
   save_default_nix_test <- function(pkgs, interface, path_default_nix) {
-    rix(
-      r_ver = "4.3.1",
-      r_pkgs = pkgs,
-      ide = interface,
-      project_path = path_default_nix,
-      overwrite = TRUE,
-      shell_hook = NULL
+    # Because of rix_init, see above
+    suppressWarnings(
+      rix(
+        r_ver = "4.3.1",
+        r_pkgs = pkgs,
+        ide = interface,
+        project_path = path_default_nix,
+        overwrite = TRUE,
+        shell_hook = NULL
+      )
     )
 
     file.path(path_default_nix, "default.nix")
@@ -147,6 +155,10 @@ testthat::test_that("Quarto gets added to sys packages", {
     ),
     name = "yes_quarto_default.nix"
   )
+
+  on.exit({
+    unlink(path_default_nix, recursive = TRUE, force = TRUE)
+  })
 })
 
 
@@ -171,13 +183,17 @@ testthat::test_that("r_pkgs = NULL and ide = 'rstudio' work together", {
   testthat::announce_snapshot_file("rix/null_pkgs_rstudio.nix")
 
   testthat::expect_snapshot_file(
-    path = save_default_nix_test(
-      pkgs = NULL,
-      interface = "rstudio",
-      path_default_nix
-    ),
-    name = "null_pkgs_rstudio.nix"
-  )
[…]
