+++ R/rix.R
@@ -207,12 +207,19 @@ for more details."
     paste0(project_path, "/default.nix")
   }
 
+  .Rprofile_path <- if (project_path == ".") {
+                      ".Rprofile"
+                    } else {
+                      paste0(project_path, "/.Rprofile")
+                    }
+
   # Find url to use
   # In case of bleeding or frozen edge, the rstats-on-nix/nixpkgs
   # fork is used. Otherwise, upstream NixOS/nixpkgs
   nix_repo <- make_nixpkgs_url(r_ver)
 
   default.nix_path <- file.path(default.nix_path)
+  .Rprofile_path <- file.path(.Rprofile_path)
 
   rix_call <- match.call()
 
@@ -288,16 +295,22 @@ for more details."
   if (!file.exists(default.nix_path) || overwrite) {
     writeLines(default.nix, default.nix_path)
 
-    rix_init(project_path = project_path,
-             rprofile_action = "create_missing",
-             # 'verbose' is too chatty for rix()
-             # hence why it's transformed to "simple"
-             message_type = ifelse(message_type == "verbose",
-                                   "simple", message_type))
+    if(!file.exists(.Rprofile_path)){
+      rix_init(project_path = project_path,
+               rprofile_action = "create_missing",
+               # 'verbose' is too chatty for rix()
+               # hence why it's transformed to "simple"
+               message_type = ifelse(message_type == "verbose",
+                                     "simple", message_type))
 
-    if(message_type != "quiet"){
-      message("\n\n### Successfully generated `default.nix` and `.Rprofile` ###\n\n")
-   }
+      if(message_type != "quiet"){
+        message("\n\n### Successfully generated `default.nix` and `.Rprofile` ###\n\n")
+      }
+
+    } else if(all(!grepl("File generated by `rix::rix_init()",
+                        readLines(.Rprofile_path)))){
+      warning("\n\n### .Rprofile file already exists. You may want to call rix_init(profile_action = 'append') manually to ensure correct functioning of your Nix environment. ###\n\n")
+    }
 
   } else {
     project_path <- if(project_path == ".") {
