### 2024-07-01 2f1cfb0
handling cases where .Rprofile was already generated by rix_init()
