+++ tests/testthat/test-rix.R
@@ -428,7 +428,7 @@ testthat::test_that("rix(), only one Github package", {
 
   save_default_nix_test <- function(path_default_nix) {
 
-    rix(r_ver = "latest",
+    rix(r_ver = "4.3.1",
       r_pkgs = NULL,
       system_pkgs = NULL,
       git_pkgs = list(package_name = "lookup",
@@ -466,7 +466,7 @@ testthat::test_that("rix(), conclusion message", {
 
   save_default_nix_test <- function(path_default_nix) {
 
-    rix(r_ver = "latest",
+    rix(r_ver = "4.3.1",
       ide = "other",
       project_path = path_default_nix,
       message_type = "simple",
@@ -485,3 +485,51 @@ testthat::test_that("rix(), conclusion message", {
     unlink(path_default_nix, recursive = TRUE, force = TRUE)
   })
 })
+
+
+testthat::test_that("rix(), warning message if rix_init() already called", {
+
+  testthat::skip_on_os
+
+  path_default_nix <- paste0(
+    tempdir(), paste0(sample(letters, 5), collapse = "")
+  )
+  dir.create(path_default_nix)
+  path_default_nix <- normalizePath(path_default_nix)
+
+  rix_init(
+      project_path = path_default_nix,
+      rprofile_action = "overwrite",
+      message_type = "simple"
+  )
+
+  # Remove lines starting with # so rix() thinks the
+  # .Rprofile file was not generated by rix_init()
+  system(
+    paste0("sed -i '/^#/d' ",
+           file.path(path_default_nix, ".Rprofile"))
+  )
+
+
+  save_default_nix_test <- function(path_default_nix) {
+
+    rix(r_ver = "4.3.1",
+      ide = "other",
+      project_path = path_default_nix,
+      message_type = "simple",
+      overwrite = TRUE
+    )
+
+    file.path(path_default_nix, "default.nix")
+  }
+
+  testthat::expect_warning(
+    save_default_nix_test(path_default_nix),
+    regexp = "You may"
+  )
+
+  on.exit({
+    unlink(path_default_nix, recursive = TRUE, force = TRUE)
+  })
+})
+
