### 2024-07-01 c7e124c
use `message_type = "simple"` as default (consistency in rix)
