+++ R/rix.R
@@ -294,8 +294,22 @@ for more details."
 
   if (!file.exists(default.nix_path) || overwrite) {
     writeLines(default.nix, default.nix_path)
-
-    if(!file.exists(.Rprofile_path)){
+    
+    if (file.exists(.Rprofile_path)) {
+      if (all(!grepl("File generated by `rix::rix_init()",
+        readLines(.Rprofile_path)))) {
+        if (message_type != "quiet"){
+          message("\n\n### Successfully generated `default.nix` ###\n\n")
+        }
+        warning("\n\n### .Rprofile file already exists. You may want to call rix_init(profile_action = 'append') manually to ensure correct functioning of your Nix environment. ###\n\n")
+      } else {
+        if (message_type != "quiet"){
+          message("\n\n### Successfully generated `default.nix`. ",
+            "Keeping `.Rprofile` generated by `rix::rix_init()`###\n\n")
+        }
+      }
+    }
+    else if (!file.exists(.Rprofile_path)){
       rix_init(project_path = project_path,
                rprofile_action = "create_missing",
                # 'verbose' is too chatty for rix()
@@ -303,21 +317,18 @@ for more details."
                message_type = ifelse(message_type == "verbose",
                                      "simple", message_type))
 
-      if(message_type != "quiet"){
+      if (message_type != "quiet"){
         message("\n\n### Successfully generated `default.nix` and `.Rprofile` ###\n\n")
       }
 
-    } else if(all(!grepl("File generated by `rix::rix_init()",
-                        readLines(.Rprofile_path)))){
-      warning("\n\n### .Rprofile file already exists. You may want to call rix_init(profile_action = 'append') manually to ensure correct functioning of your Nix environment. ###\n\n")
     }
 
   } else {
-    project_path <- if(project_path == ".") {
+    project_path <- if (project_path == ".") {
                       "current folder"
                     }  else {
                       project_path
                     }
-    stop(paste0("default.nix exists in ", project_path, ". Set `overwrite == TRUE` to overwrite."))
+    stop(paste0("`default.nix` exists in ", project_path, ". Set `overwrite == TRUE` to overwrite."))
   }
 }
