+++ R/rix.R
@@ -204,13 +204,13 @@ for more details."
   default.nix_path <- if (project_path == ".") {
     "default.nix"
   } else {
-    paste0(project_path, "/default.nix")
+    file.path(project_path, "default.nix")
   }
 
   .Rprofile_path <- if (project_path == ".") {
                       ".Rprofile"
                     } else {
-                      paste0(project_path, "/.Rprofile")
+                      file.path(project_path, ".Rprofile")
                     }
 
   # Find url to use
@@ -304,7 +304,7 @@ for more details."
         warning("\n\n### .Rprofile file already exists. You may want to call rix_init(profile_action = 'append') manually to ensure correct functioning of your Nix environment. ###\n\n")
       } else {
         if (message_type != "quiet"){
-          message("\n\n### Successfully generated `default.nix`. ",
+          message("\n\n### Successfully generated `default.nix` in . ",
             "Keeping `.Rprofile` generated by `rix::rix_init()`###\n\n")
         }
       }
