### 2024-07-01 8121ddf
message if keeping valid `rix_init()` `.Rprofile`
