+++ R/rix.R
@@ -306,7 +306,11 @@ for more details."
           "to ensure correct functioning of your Nix environment. ###\n\n")
       } else {
         if (message_type != "quiet"){
-          message("\n\n### Successfully generated `default.nix` in . ",
+          message(
+            sprintf(
+              "\n\n### Successfully generated `default.nix` in %s. ",
+              normalizePath(project_path)
+            ),
             "Keeping `.Rprofile` generated by `rix::rix_init()`###\n\n")
         }
       }
