+++ vignettes/z-advanced-topic-handling-packages-with-remote-dependencies.Rmd
@@ -44,8 +44,7 @@ Fortran. To create a reproducible development environment that makes
 [{lookup}](https://github.com/jimhester/lookup) available to you, you could use
 the following `rix::rix()` call:
 
-
-```{r, include = F }
+```{r, eval = F}
 path_default_nix <- tempdir()
 
 rix(r_ver = "latest",
