### 2024-07-02 #625 @philipp-baumann
@assignUser Thanks for your suggestions, they helped me polish `with_nix()` along with `rix_init()` and `nix_build()`.  These are aimed at using R and shell code in pinned nix software environments. Many of the improvements also affect the underlying helper functions. We are pleased to announce today that we will soon be releasing {rix} version 0.8.0, which will highlight these changes, among other improvements and bug fixes. I'll try to give some more explanations here. [@wdwatkins](https://github.com/wdwatkins) and https://github.com/assignUser ,f eel free to ping me with questions and clarifications. Here is the first batch of changes and improvements.

- https://github.com/b-rodrigues/rix/releases/tag/v0.8.0
- https://github.com/b-rodrigues/rix/blob/master/NEWS.md

> > with_nix is very verbose with ~ a terminal length of 'preamble' output of checks etc., I think a 'quite' option would be nice when running it repeatedly.
>
> What would be the best way to do so? Is there some way that we could add this without needing to touch the overall structure of the function?

Thanks for pointing this out. The `message_type = "quiet"` is an important option that was missing. In `with_nix()`, as well as `rix_init()` and `nix_build()`, we now have consistent message types with the following enumeration options available via `message_type

- `"simple"`: the default. This gives minimal but sufficient feedback to understand the main steps taken.
- `"quiet"`: No message is printed.
- `"verbose"`: a more detailed report for better diagnostics.

`message_type` is now propagated internally in the helpers, ensuring that the user can only use these three enums.

 Here is a quick illustration for "quiet" and "verbose, for `rix_init()`, much of this is made possible by https://github.com/b-rodrigues/rix/pull/177

## "simple" messaging

```r
> rix_init(project_path = "_my_env", rprofile_action = "create_missing", message_type = "simple")

### Bootstrapping isolated, project-specific, and runtime-pure R setup via Nix ###

==> Created isolated nix-R project folder:
 /home/philipp/git/philipp-baumann-shared/rix/_my_env

* current R session running outside Nix environment and from RStudio

==> Added `.Rprofile` file and code lines for new R sessions launched from:
/home/philipp/git/philipp-baumann-shared/rix/_my_env

* Added the location of the Nix store to `PATH` environmental variable for new R sessions on host/docker RStudio:
/nix/var/nix/profiles/default/bin

==> Also adjusting `PATH` via `Sys.setenv()`, so that system commands can invoke key Nix commands like `nix-build` in this RStudio session outside Nix
```

## "quiet" aka no messaging

```r
> rix_init(project_path = "_my_env", rprofile_action = "create_missing", message_type = "quiet")
```

## Reducing dependencies and reverse dependencies
We have removed {httr} for GET requests, and now use low-level fail-safe alternatives for downloading `tar.gz' archives from GitHub and CRAN based on {curl}.

Concerns issue https://github.com/b-rodrigues/rix/pull/177
addressed by

- https://github.com/b-rodrigues/rix/pull/189
- https://github.com/b-rodrigues/rix/pull/193
- for downloading archives before the new local `nix-hash' hashing: https://github.com/b-rodrigues/rix/pull/221

We now have a few important but high quality dependencies:

- codetools
- curl
- jsonlite
- sys
- utils

Low-dependency principles are an important way to keep the package stable and reproducible over time.

## Improve formatting of `default.nix` expressions
- The fine-tuning tries to respect indentation and list packages on separate lines for better version control handling and visual appearance.
- Addressed in https://github.com/b-rodrigues/rix/pull/195

## Allow local hashing of GitHub and CRAN archive packages
- The version of {rix} that we submitted to rOpenSci used an API server http://git2nixsha.dev that Bruno set up to allow Nix-specific hashing, specifically SHA-256 hashes in SRI format and base64 applied to the Nix Archive (NAR) serialisation format (for more details see https://github.com/NixOS/nix/pull/7690 https://nix.dev/manual/nix/2.23/store/file-system-object/content-address#nix-archive).
- Online hashing is a nice add-on, as it does not require Nix to be installed, and users can generate NIx expressions containing GitHub packages. However, if Nix is present, the same can be done locally. This motivated us to implement a local version.
- This implementation led to an internal redesign of the high-level `get_sri_hash_deps()` and the underlying infrastructure, here `nix_hash()`, which does the equivalent of `nix_hash_online()`.
- We have made both local and online hashing more modular and failsafe.
- There is also a new option `options(rix.nix_hash=)`; a user can set it to either `"local"` or `"api_server"`. If no option is set, `"check_nix"` is sent, which checks if Nix and CLIs are available aka in the PATH. If local hashing is requested by the global options, but nix is not installed, this will signal an error.
- The new local feature was added via https://github.com/b-rodrigues/rix/pull/221
