### 2024-07-04 #625 @b-rodrigues
We are now done with version 0.8.0 and will only do quick bug fixes if we find them, or if they’re raised by users. This is the version that we would like to "resubmit" for review, and if everything goes well, to CRAN as well.
