### 2024-07-04 #625 @philipp-baumann
Here the second batch of changes we have made :-)

- The data frame with the r `version` the nixpkgs `revision` (corresponding hexadecimal commit ID; SHA-1), and the corresponding `date` is now treated as system data as suggested in the R manual and help. We moved it to `R/sysdata`. The helper `available_r()`, which lazy-loads by default, which is the way to treat the data, since we use it for mapping the `r_pkgs` input if no GitHub commit digest is given.
  - We resolved it here: https://github.com/b-rodrigues/rix/pull/220 ; this was done in response to a user that found an issue when calling `rix::rix()` when not attaching the package via `library("rix")` before. https://github.com/b-rodrigues/rix/issues/200

- We now also support the installation of local R packages from `.tar.gz` sources. `rix::rix()` has gained an additional `local_r_pkgs` argument where packages in `.tar.gz` or `.zip` formats can be provided for reproducible local builds via Nix.
