### 2024-07-04 #625 @philipp-baumann
> > This is required because the underlying Nix function fetchgit requires both the repo and the commit. It might be more cumbersome from a UX perspective,
>
> Ok in that case leave branch out as it's redundant or allow adding repo + branch only and fetch the HEAD of the branch for the user.

Based on your input, and also because the [`fetchGit` section in the Nix manual](https://nix.dev/manual/nix/2.18/language/builtins.html#builtins-fetchGit) may have an incorrect statement about the need to enter branch names when revisions aka valid full GitHub commit hashes are provided; @b-rodrigues and I had a brief discussion about this via matrix and thought it would make sense to eventually remove these seemingly redundant entries for the user at the `git_pkgs` list argument.
_"Note

    It is nice to always specify the branch which a revision belongs to. Without the branch being specified, the fetcher might fail if the default branch changes. Additionally, it can be confusing to try a commit from a non-default branch and see the fetch fail. If the branch is specified the fault is much more obvious."_

Its super weird, branches are moving pointers to git commits. So this statement does not make any sense to me. Even after reading the discussion on https://discourse.nixos.org/t/rev-and-ref-attributes-in-builtins-fetchgit-and-maybe-flakes-too/20588/4, the recommendation does not seem to make sense.  Builtins.fetchGit expects the `rev` argument to be the full commit hash, a 40-byte SHA-1 hexadecimal string. I have seen no explanation as to why this is not sufficient to give the commit digest, i.e. why it should fail if the branch changes.

To be more specific, since we implemented local hashing for when Nix is installed on the local system using rix functionality, we only kept the `repo_url` and `commit` (equivalent to `rev` in the NIx fetchGit fetcher (which we don't use here since we download the GitHub and CRAN archives and derive hashes with the `nix-hash` CLI) -> https://github.com/b-rodrigues/rix/blob/51c19a9835ac172a08a864885fe40ced7375a769/R/nix_hash.R#L5.

However, `get_sri_hash_deps()` (https://github.com/b-rodrigues/rix/blob/51c19a9835ac172a08a864885fe40ced7375a769/R/nix_hash.R#L208), because when redirecting to the http://git2nixsha.dev:1506 API server to compute the hashes in case of `nix_hash_online()`, the server uses `git2r` to clone and checkout before hashing with `nix hash path` (this feature is newer but experimental), but gives the same SRI SHA-256 digest. So for this to work, the API server needs to be changed to use the functions used for local hashing (i.e. the same functions). This is a transition, and also for users, if we deprecate right now it might be a bit annoying and hinder reproducibility.

With that in mind, @b-rodrigues and I discussed removing the extra entries and just keeping `repo_url` and `` in a future release. wdwatkins thanks for the hint, do you also see the option to soft-depreciate or do you have any additional considerations?
