### 2024-07-11 #625 @b-rodrigues
> With that in mind, @b-rodrigues and I have discussed removing the requirement of a `branch_name` element in list inputted to the `git_pkgs` argument of `rix::rix()` in a future release. @wdwatkins we appreciate your hint! do you also see the option to soft-depreciate or do you have any additional considerations?

After more thought and discussion the `branch_name` argument is now removed.
