### 2024-09-03 #625 @ropensci-review-bot
Approved! Thanks @b-rodrigues for submitting and @wdwatkins, @assignUser for your reviews! :grin:
