### 2024-09-05 #625 @philipp-baumann
> Whoo hoo! Thanks everyone for a great and thorough job! @b-rodrigues and @philipp-baumann - really well done work! @wdwatkins and @assignUser thanks for the fantastic feedback!

Thanks very much for your excellent reviews! It was a really constructive process 🙏 {rix} has become mature and I'm really happy it is being sheltered under the ROpenSci umbrella.
