### 2024-09-06 #625 @b-rodrigues
@wdwatkins and @assignUser would you be OK if we add you as reviewers in the package's DESCRIPTION ?
