### 2024-09-12 #625 @b-rodrigues
Hi @ldecicco-USGS I'd need some guidance regarding the pkgdown website. We currently have the website up at https://ropensci.github.io/rix/

So from my understanding, all we need to do is remove the github action `pkgdown` that builds and deploys the website, and change every instance of `ropensci.github.io` with the string `docs.ropensci.org`, right? That point about redirecting is not needed, since the repo has been transferred to the ropensci organization, or?
