### 2024-09-13 #625 @philipp-baumann
Hi @ldecicco-USGS . While testing a pre-commit-lite ci hook for styling ( https://github.com/ropensci/rix/pull/295 ) we realized that I could not add the integration ( https://results.pre-commit.ci/ ). Could @ropensci-review-bot give me admin rights on the repo too? Thanks
