### 2024-09-13 #625 @ldecicco-USGS
@philipp-baumann - you should have an invitation to the rOpenSci slack channel to discuss adding pre-commit. One quote cracked me up: "I'm always happy to opine on precommit!"
