### 2024-09-13 #625 @b-rodrigues
> @philipp-baumann I have sent you an invitation to the rOpenSci organization and a team with admin access to the repository (together with @b-rodrigues). I have removed the personal write access both of you had, but you get _more_ access via the team.

sweet!
