### 2024-09-16 #625 @b-rodrigues
> That sounds right to me, but I'll ask around to see if we're missing anything.

I see that the webpage is up already: https://docs.ropensci.org/rix/

so I guess it happens automatically 😃
